LTFHPlus: Implementation of LT-FH++

Implementation of LT-FH++, an extension of the liability threshold family history (LT-FH) model. LT-FH++ uses a Gibbs sampler for sampling from the truncated multivariate normal distribution and allows for flexible family structures. LT-FH++ was first described in Pedersen, Emil M., et al. (2022) doi:10.1016/j.ajhg.2022.01.009 as an extension to LT-FH with more flexible family structures, and again as the age-dependent liability threshold (ADuLT) model Pedersen, Emil M., et al. (2023) https://www.nature.com/articles/s41467-023-41210-z as an alternative to traditional time-to-event genome-wide association studies, where family history was not considered.

See also

Author

Maintainer: Emil Michael Pedersen emp@ph.au.dk

Authors:

Florian Privé florian.prive.21@gmail.com [thesis advisor]

Jette Steinbach jst@econ.au.dk

Other contributors:

Bjarni Jóhann Vilhjálmsson bjv@econ.au.dk [thesis advisor]

Esben Agerbo ea@econ.au.dk [thesis advisor]

Lucas Rasmussen lar.ncrr@au.dk [contributor]