Verifies that the covariance matrix is positive definite (a properti of gaussian covariance matrices) by checking if all eigen values are positive.

Verifies that the covariance matrix is positive definite (a properti of gaussian covariance matrices) by checking if all eigen values are positive.

check_positive_definite(full_cov, corr_mat, correction_val = 0.99, n_sib = 0)

Arguments

full_cov
The full covariance matrix for the LT-FH++ method.

corr_mat
Matrix containing the heritabilities and correlations for each of the phenotypes provided. Heritabilities should be on the diagonal with the corresponing off-diagonal entry holding the correlation between two phenotypes.

correction_val
value to multiply to off-diagonal entries in an attempt to force positive definite.

n_sib
Number of Siblings to include in the matrix.

Value

Returns the full covariance matrix based off of the heritabilities and corralations provided in corr_mat.

Examples

check_positive_definite(get_cov(.5), matrix(.5, 1, 1))

#> Covariance matrix is not positive definite (not all eigen values are positive) - Slightly reducing correlation to mitigate. 

#>      [,1] [,2] [,3] [,4]
#> [1,] 0.50 0.50 0.25 0.25
#> [2,] 0.50 1.00 0.25 0.25
#> [3,] 0.25 0.25 1.00 0.00
#> [4,] 0.25 0.25 0.00 1.00