Estimate genetic liability based off of family history and prevalence information.

This function has been replaced by estimate_liability. This function is kept for legacy reasons.

estimate_gen_liability(h2, phen, thr, ind = c(1), id_col = "ids", tol = 0.01)

Arguments

h2

Liability scale heritability of the trait being analysed.

phen

tibble or data.frame with the IDs and status of the genotyped individual and the parents and siblings.

thr

tibble or data.frame with a row of each individual in the provided phen. Each row should contain the ID and threshold value needed for the model.

ind

Indices to return from the gibbs sampler. 1 corresponds to the genetic liability of the first phenotype. c(1,5) corresponds to the genetic liability of the first two phenotypes provided, but with no siblings in the model.

id_col

Column name for IDs of family members.

tol

Convergence criteria of the gibbs sampler. Default is 0.01, meaning a standard error of the mean below 0.01

Value

Returns the estimated genetic liabilities.

Examples

#See vignettes for example of use and input.