Estimating the genetic or full liability

estimate_liability estimates the genetic component of the full liability and/or the full liability for a number of individuals based on their family history.

estimate_liability(
  family,
  threshs,
  h2 = 0.5,
  pid = "PID",
  fam_id = "fam_ID",
  out = c(1),
  tol = 0.01,
  always_add = c("g", "o")
)

Arguments

family

A matrix, list or data frame that can be converted into a tibble. Must have at least two columns that hold the family identifier and the corresponding personal identifiers, respectively. That is, for each family in fam_id there should be a list holding all individuals belonging to that family in pid. Note that the personal identifiers for all individuals must have a special format. It must be end with _?, where ? is one of the following abbreviations

g (Genetic component of the full liability)

o (full liability)

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). See also construct_covmat.

threshs

A matrix, list or data frame that can be converted into a tibble. Must have at least three columns, one holding the personal identifier for all individuals, and the remaining two holding the lower and upper thresholds, respectively.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

pid

A string holding the name of the column in family and threshs that hold the personal identifier(s). Defaults to "PID".

fam_id

A string holding the name of the column in family that holds the family identifier. Defaults to "fam_ID".

out

A character or numeric vector indicating whether the genetic component of the full liability, the full liability or both should be returned. If out = c(1) or out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or out = c("full"), the full liability is estimated and returned. If out = c(1,2) or out = c("genetic", "full"), both components are estimated and returned. Defaults to c(1).

tol

A number that is used as the convergence criterion for the Gibbs sampler. Equals the standard error of the mean. That is, a tolerance of 0.2 means that the standard error of the mean is below 0.2. Defaults to 0.01.

always_add

A character vector or NULL. If always_add = c("g","o"), both the genetic component of the full liability as well as the full liability will be added to the list of family members. If always_add equals "g" or "o", the genetic component of the full liability or the full liability will be added, respectively. If always_add = NULL, no component will be added. Defaults to c("g","o").

Value

If family and threshs are two matrices, lists or data frames that can be converted into tibbles, if family has two columns named like the strings represented in pid and fam_id, if threshs has a column named like the string given in pid as well as a column named "lower" and a column named "upper" and if the heritability, out, tol and always_add are of the required form, then the function returns a tibble with either four or six columns (depending on the length of out). The first two columns correspond to the columns fam_id and pid from family. If out is equal to c(1) or c("genetic"), the third and fourth column hold the estimated genetic liability as well as the corresponding standard error, respectively. If out equals c(2) or c("full"), the third and fourth column hold the estimated full liability as well as the corresponding standard error, respectively. If out is equal to c(1,2) or c("genetic","full"), the third and fourth column hold the estimated genetic liability as well as the corresponding standard error, respectively, while the fifth and sixth column hold the estimated full liability as well as the corresponding standard error, respectively.

Details

This function can be used to estimate either the genetic component of the full liability, the full liability or both.

See also

Examples

sims <- simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, 
add_ind = TRUE, h2 = 0.5, n_sim=100, pop_prev = .05)

estimate_liability(family = sims$fam_ID, threshs = sims$thresholds, 
h2 = 0.5, pid = "indiv_ID", fam_id = "fam_ID", out = c(1), tol = 0.01,
always_add = c("g","o"))
#> The number of workers is 1
#> # A tibble: 100 × 4
#>    fam_ID    indiv_ID Posterior_genetic_liab Posterior_genetic_std_err
#>    <list>    <chr>                     <dbl>                     <dbl>
#>  1 <chr [4]> ID1                     -0.0485                   0.00415
#>  2 <chr [4]> ID2                     -0.0269                   0.00413
#>  3 <chr [4]> ID3                     -0.0389                   0.00403
#>  4 <chr [4]> ID4                     -0.0174                   0.00411
#>  5 <chr [4]> ID5                     -0.0387                   0.00388
#>  6 <chr [4]> ID6                      0.528                    0.00334
#>  7 <chr [4]> ID7                     -0.0339                   0.00396
#>  8 <chr [4]> ID8                     -0.0212                   0.00410
#>  9 <chr [4]> ID9                     -0.0347                   0.00400
#> 10 <chr [4]> ID10                    -0.0460                   0.00415
#> # … with 90 more rows
# 
sims <- simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, 
h2 = 0.5, n_sim=100, pop_prev = .05)
estimate_liability(family = sims$fam_ID, threshs = sims$thresholds, 
h2 = 0.5, pid = "indiv_ID", fam_id = "fam_ID", out = c("genetic"), 
tol = 0.01, always_add = c("g","o"))
#> The number of workers is 1
#> # A tibble: 100 × 4
#>    fam_ID    indiv_ID Posterior_genetic_liab Posterior_genetic_std_err
#>    <list>    <chr>                     <dbl>                     <dbl>
#>  1 <chr [1]> ID1                     1.22                      0.00154
#>  2 <chr [1]> ID2                    -0.00940                   0.00382
#>  3 <chr [1]> ID3                    -0.00663                   0.00371
#>  4 <chr [1]> ID4                    -0.00252                   0.00407
#>  5 <chr [1]> ID5                     0.00453                   0.00385
#>  6 <chr [1]> ID6                     0.00162                   0.00347
#>  7 <chr [1]> ID7                    -0.00743                   0.00357
#>  8 <chr [1]> ID8                     0.00348                   0.00380
#>  9 <chr [1]> ID9                    -0.00172                   0.00373
#> 10 <chr [1]> ID10                   -0.00905                   0.00370
#> # … with 90 more rows