Estimating the genetic or full liability for multiple phenotypes FUNCTIONS FOR MULTIPLE TRAITS IS STILL NOT COMPLETE. PLEASE DO NOT USE. estimate_liability_multi estimates the genetic component of the full liability and/or the full liability for a number of individuals based on their family history for a variable number of phenotypes.

This function can be used to estimate either the genetic component of the full liability, the full liability or both for a variable number of traits.

estimate_liability_multi(
  family,
  threshs,
  h2s,
  corrmat,
  pid = "PID",
  fam_id = "fam_ID",
  out = c(1),
  tol = 0.01,
  parallel = FALSE
)

Arguments

family

A matrix, list or data frame that can be converted into a tibble. Must have at least two columns that hold the family identifier and the corresponding personal identifiers, respectively. That is, for each family in fam_id there should be a list holding all individuals belonging to that family in pid. Note that the personal identifiers for all individuals must have a special format. It must be end with _?, where ? is one of the following abbreviations

g (Genetic component of the full liability)

o (full liability)

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). See also construct_covmat.

threshs

A matrix, list or data frame that can be converted into a tibble. Must have at least five columns; one holding the personal identifier for all individuals, and the remaining four holding the lower and upper thresholds for the first and second phenotype, respectively. It must be possible to tie each pair of lower and upper thresholds to a specific phenotype uniquely. This is done easily by adding _name_of_phenotype to the column names lower and upper, e.g. lower_p1 and upper_p1 for the lower and upper thresholds corresponding to the first phenotype.

h2s

A numeric vector representing the heritability on liability scale for all phenotypes. All entries in h2s must be non-negative and at most 1.

corrmat

A numeric matrix holding the genetic correlation between the desired number of phenotypes as well as the full correlation. The full correlations must be given on the diagonal, while the off-diagonal entries must hold the correlation between phenotypes. All correlations must be between -1 and 1.

pid

A string holding the name of the column in family and threshs that hold the personal identifier(s). Defaults to "PID".

fam_id

A string holding the name of the column in family that holds the family identifier. Defaults to "fam_ID".

out

A character or numeric vector indicating whether the genetic component of the full liability, the full liability or both should be returned. If out = c(1) or out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or out = c("full"), the full liability is estimated and returned. If out = c(1,2) or out = c("genetic", "full"), both components are estimated and returned. Defaults to c(1).

tol

A number that is used as the convergence criterion for the Gibbs sampler. Equals the standard error of the mean. That is, a tolerance of 0.2 means that the standard error of the mean is below 0.2. Defaults to 0.01.

parallel

A logical scalar indicating whether computations should be performed parallel. In order for this to be possible, the user must install the library "future.apply" and create a plan (see future_apply). Defaults to FALSE.

Value

If family and threshs are two matrices, lists or data frames that can be converted into tibbles, if family has two columns named like the strings represented in pid and fam_id, if threshs has a column named like the string given in pid as well as a column named "lower" and a column named "upper" and if the heritabilities, corrmat, out and tol are of the required form, then the function returns a tibble with at least six columns (depending on the length of out). The first two columns correspond to the columns fam_id and pid from family. If out is equal to c(1) or c("genetic"), the third and fourth columns hold the estimated genetic liability as well as the corresponding standard error for the first phenotype, respectively. If out equals c(2) or c("full"), the third and fourth columns hold the estimated full liability as well as the corresponding standard error for the first phenotype, respectively. If out is equal to c(1,2) or c("genetic","full"), the third and fourth columns hold the estimated genetic liability as well as the corresponding standard error for the first phenotype, respectively, while the fifth and sixth columns hold the estimated full liability as well as the corresponding standard error for the first phenotype, respectively. The remaining columns hold the estimated genetic liabilities and/or the estimated full liabilities as well as the corresponding standard errors for the remaining phenotypes.

See also