Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g       o      m       f      s1    mgm     mgf    pgm     pgf
#>    <chr>      <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 fam_ID_1 -0.297   1.27   -0.165 -2.18   -0.579   0.678 -0.652   0.431 -1.59  
#>  2 fam_ID_2  0.584   1.17    0.575  1.01   -1.06   -0.424 -0.0579  1.56  -0.872 
#>  3 fam_ID_3 -0.142  -0.660  -1.03   0.473   0.333   0.726 -1.30   -1.91   1.69  
#>  4 fam_ID_4 -0.119   0.455   1.10   0.0454  0.0385 -0.450  0.790   0.667 -1.84  
#>  5 fam_ID_5 -0.612  -1.48   -0.915  0.0255 -1.79   -0.991 -0.526  -1.15  -1.13  
#>  6 fam_ID_6  0.616   0.289   0.434 -1.58    0.730  -0.552  0.195  -0.275 -0.0513
#>  7 fam_ID_7  0.0320 -0.761   0.171 -0.430  -0.162   0.196  1.00   -0.182  2.44  
#>  8 fam_ID_8  0.423  -0.241   0.909  0.639   1.46   -0.794 -2.14    0.379 -1.15  
#>  9 fam_ID_9  0.406   0.0808  1.01   1.02   -0.0886  1.37  -0.0316  1.39  -0.0735
#> 10 fam_ID_… -1.31   -0.460  -1.32  -0.263  -1.75   -2.19  -0.0309  0.277 -0.0850
#> # ℹ 990 more rows
#> # ℹ 16 more variables: o_status <lgl>, m_status <lgl>, f_status <lgl>,
#> #   s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, pgm_status <lgl>,
#> #   pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.76
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  2.43
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.59
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.42
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.03
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.24
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.87
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.24
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.38
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  2.63
#> # ℹ 7,990 more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID          g       o      m     mgm     mgf   mhs1    mhs2 o_status
#>    <chr>       <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>   <dbl> <lgl>   
#>  1 fam_ID_1  -0.442  -0.629  -1.23   0.766  -0.415   0.325 -0.0404 FALSE   
#>  2 fam_ID_2  -0.149   0.0900  0.353  0.0410 -0.390   0.709  0.941  FALSE   
#>  3 fam_ID_3   1.05    0.879   0.325 -1.48    1.49    0.117 -0.429  FALSE   
#>  4 fam_ID_4   1.55    1.64    1.08   0.278   0.208  -0.877  0.409  TRUE    
#>  5 fam_ID_5   0.867   0.731  -0.226  1.02    1.00   -0.367 -0.0972 FALSE   
#>  6 fam_ID_6  -0.110  -0.0557 -1.23  -0.273  -1.20   -1.22  -1.22   FALSE   
#>  7 fam_ID_7  -1.20   -1.94   -1.50   0.135  -0.535  -1.72   0.411  FALSE   
#>  8 fam_ID_8  -1.48   -2.48   -1.14  -1.77    1.86   -0.631 -0.785  FALSE   
#>  9 fam_ID_9  -0.0812 -0.0568 -0.110 -1.59    1.70   -1.70   0.207  FALSE   
#> 10 fam_ID_10 -0.287  -1.29   -0.124 -0.0170  0.0166  1.28   0.799  FALSE   
#> # ℹ 990 more rows
#> # ℹ 11 more variables: m_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.31
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.95
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.55
#>  4 fam_ID_4  fam_ID_4_1  o        1.64  1.64
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.10
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     2.55
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.79
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.14
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     2.63
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.76
#> # ℹ 5,990 more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID        m      f      s1 m_status f_status s1_status m_aoo f_aoo s1_aoo
#>    <chr>     <dbl>  <dbl>   <dbl> <lgl>    <lgl>    <lgl>     <dbl> <dbl>  <dbl>
#>  1 fam_ID_1 -0.412 -0.535 -1.05   FALSE    FALSE    FALSE        56    61     32
#>  2 fam_ID_2  0.226 -1.55  -0.538  FALSE    FALSE    FALSE        52    59     34
#>  3 fam_ID_3  0.601  0.220  2.16   FALSE    FALSE    TRUE         61    55     53
#>  4 fam_ID_4 -0.425  1.93  -0.0310 FALSE    TRUE     FALSE        58    61     31
#>  5 fam_ID_5  0.278  1.12   1.40   FALSE    FALSE    FALSE        43    49     22
#>  6 fam_ID_6  2.00  -0.777  1.66   TRUE     FALSE    TRUE         59    41     86
#>  7 fam_ID_7 -0.947 -0.557  0.369  FALSE    FALSE    FALSE        60    58     38
#>  8 fam_ID_8 -0.644  1.49   0.286  FALSE    FALSE    FALSE        46    42     21
#>  9 fam_ID_9  1.78   1.20  -0.0720 TRUE     FALSE    FALSE        69    65     40
#> 10 fam_ID_… -0.314  2.13  -0.633  FALSE    TRUE     FALSE        53    55     24
#> # ℹ 490 more rows
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     2.08
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     2.21
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     1.93
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     2.02
#>  5 fam_ID_5  fam_ID_5_1  m     -Inf     2.55
#>  6 fam_ID_6  fam_ID_6_1  m        1.99  1.99
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     1.96
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     2.44
#>  9 fam_ID_9  fam_ID_9_1  m        1.78  1.78
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     2.18
#> # ℹ 1,490 more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID         g       o o_status o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -0.564  0.192  FALSE       28
#>  2 fam_ID_2   0.648  0.0833 FALSE       29
#>  3 fam_ID_3  -0.469  0.234  FALSE       28
#>  4 fam_ID_4  -1.32  -1.02   FALSE       36
#>  5 fam_ID_5   1.27   0.774  FALSE       15
#>  6 fam_ID_6  -0.753 -0.153  FALSE       11
#>  7 fam_ID_7  -1.45  -1.84   FALSE       34
#>  8 fam_ID_8  -0.186  0.424  FALSE       20
#>  9 fam_ID_9  -0.247  0.483  FALSE       40
#> 10 fam_ID_10 -0.685 -0.528  FALSE       13
#> # ℹ 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.12
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.09
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.12
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.82
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.57
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.70
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.90
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.40
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  2.67
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.63
#> # ℹ 190 more rows
#>