Estimating the genetic or full liability for multiple phenotypes

estimate_liability_multi estimates the genetic component of the full liability and/or the full liability for a number of individuals based on their family history for a variable number of phenotypes.

estimate_liability_multi(
  .tbl = NULL,
  family = NULL,
  threshs = NULL,
  h2_vec,
  genetic_corrmat,
  full_corrmat,
  phen_names = NULL,
  pid = "PID",
  fam_id = "fam_ID",
  role = "role",
  out = c(1),
  tol = 0.01
)

Arguments

.tbl

A matrix, list or data frame that can be converted into a tibble. Must have at least seven columns that hold the family identifier, the personal identifier, the role and the lower and upper thresholds for all phenotypes of interest. Note that the role must be one of the following abbreviations

g (Genetic component of full liability)

o (Full liability)

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to NULL.

family

A matrix, list or data frame that can be converted into a tibble. Must have at least two columns that hold the family identifier and the corresponding personal identifiers, respectively. That is, for each family in fam_id there should be a list holding all individuals belonging to that family in pid. Note that the personal identifiers for all individuals must have a special format. It must be end with _?, where ? is one of the following abbreviations

g (Genetic component of full liability)

o (Full liability)

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). See also construct_covmat. Defaults to NULL.

threshs

A matrix, list or data frame that can be converted into a tibble. Must have at least five columns; one holding the personal identifier for all individuals, and the remaining four holding the lower and upper thresholds for the first and second phenotype, respectively. It must be possible to tie each pair of lower and upper thresholds to a specific phenotype uniquely. This is done easily by adding _name_of_phenotype to the column names lower and upper, e.g. lower_p1 and upper_p1 for the lower and upper thresholds corresponding to the first phenotype. Defaults to NULL.

h2_vec

A numeric vector representing the liability-scale heritabilities for all phenotypes. All entries in h2_vec must be non-negative and at most 1.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric.

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric.

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

pid

A string holding the name of the column in family and threshs that hold the personal identifier(s). Defaults to "PID".

fam_id

A string holding the name of the column in family that holds the family identifier. Defaults to "fam_ID".

role

A string holding the name of the column in .tbl that holds the role.Each role must be chosen from the following list of abbreviations

g (Genetic component of full liability)

o (Full liability)

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to "role".

out

A character or numeric vector indicating whether the genetic component of the full liability, the full liability or both should be returned. If out = c(1) or out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or out = c("full"), the full liability is estimated and returned. If out = c(1,2) or out = c("genetic", "full"), both components are estimated and returned. Defaults to c(1).

tol

A number that is used as the convergence criterion for the Gibbs sampler. Equals the standard error of the mean. That is, a tolerance of 0.2 means that the standard error of the mean is below 0.2. Defaults to 0.01.

Value

If family and threshs are two matrices, lists or data frames that can be converted into tibbles, if family has two columns named like the strings represented in pid and fam_id, if threshs has a column named like the string given in pid as well as a column named "lower"

and a column named "upper" and if the liability-scale heritabilities in h2_vec, genetic_corrmat, full_corrmat, out and tol are of the required form, then the function returns a tibble with at least six columns (depending on the length of out). The first two columns correspond to the columns fam_id and pid present in the tibble family. If out is equal to c(1) or c("genetic"), the third and fourth columns hold the estimated genetic liability as well as the corresponding standard error for the first phenotype, respectively. If out equals c(2) or c("full"), the third and fourth columns hold the estimated full liability as well as the corresponding standard error for the first phenotype, respectively. If out is equal to c(1,2) or c("genetic","full"), the third and fourth columns hold the estimated genetic liability as well as the corresponding standard error for the first phenotype, respectively, while the fifth and sixth columns hold the estimated full liability as well as the corresponding standard error for the first phenotype, respectively. The remaining columns hold the estimated genetic liabilities and/or the estimated full liabilities as well as the corresponding standard errors for the remaining phenotypes.

Details

This function can be used to estimate either the genetic component of the full liability, the full liability or both for a variable number of traits.

See also

Examples

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1
#
sims <- simulate_under_LTM(fam_vec = c("m","f"), n_fam = NULL, add_ind = TRUE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, h2 = rep(.5,3), 
n_sim = 10, pop_prev = rep(.1,3))
#
full_covmat = lapply(sims[-which(names(sims) == "thresholds")], function(x) {
   dplyr::select(x[[1]], dplyr::starts_with("o") & dplyr::ends_with("status"))
   })
full_covmat = cov(dplyr::bind_cols(full_covmat))
diag(full_covmat) = 1
estimate_liability_multi(.tbl = sims$thresholds, h2_vec = rep(.5,3), 
genetic_corrmat = genetic_corrmat, full_corrmat = full_covmat,
pid = "indiv_ID", fam_id = "fam_ID", role = "role", out = c(1), 
tol = 0.01)
#> The number of workers is 1
#> # A tibble: 10 × 8
#>    fam_ID    indiv_ID    genetic_pheno…¹ genet…² genet…³ genet…⁴ genet…⁵ genet…⁶
#>    <chr>     <chr>       <chr>           <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 fam_ID_1  fam_ID_1_1  0.729169158677… 0.0038… 0.8308… 0.0022… 0.7169… 0.0036…
#>  2 fam_ID_2  fam_ID_2_1  0.872699036041… 0.0037… 1.1394… 0.0021… 0.8739… 0.0037…
#>  3 fam_ID_3  fam_ID_3_1  -0.01684288177… 0.0054… -0.015… 0.0052… -0.016… 0.0051…
#>  4 fam_ID_4  fam_ID_4_1  -0.08258063915… 0.0047… -0.079… 0.0049… -0.081… 0.0047…
#>  5 fam_ID_5  fam_ID_5_1  0.088600040859… 0.0043… 0.3765… 0.0041… 0.0901… 0.0045…
#>  6 fam_ID_6  fam_ID_6_1  -0.07834531374… 0.0046… -0.080… 0.0046… -0.081… 0.0044…
#>  7 fam_ID_7  fam_ID_7_1  1.441316607292… 0.0014… 1.3494… 0.0015… 1.2945… 0.0015…
#>  8 fam_ID_8  fam_ID_8_1  -0.09663442700… 0.0046… -0.098… 0.0049… -0.103… 0.0048…
#>  9 fam_ID_9  fam_ID_9_1  -0.11592552911… 0.0044… -0.118… 0.0046… -0.116… 0.0044…
#> 10 fam_ID_10 fam_ID_10_1 -0.04830915666… 0.0051… -0.044… 0.0048… -0.045… 0.0049…
#> # … with abbreviated variable names ¹​genetic_phenotype1_est,
#> #   ²​genetic_phenotype1_se, ³​genetic_phenotype2_est, ⁴​genetic_phenotype2_se,
#> #   ⁵​genetic_phenotype3_est, ⁶​genetic_phenotype3_se