Estimating the genetic or full liability

estimate_liability_single estimates the genetic component of the full liability and/or the full liability for a number of individuals based on their family history.

estimate_liability_single(
  .tbl = NULL,
  family = NULL,
  threshs = NULL,
  h2 = 0.5,
  pid = "PID",
  fam_id = "fam_ID",
  role = "role",
  out = c(1),
  tol = 0.01
)

Arguments

.tbl

A matrix, list or data frame that can be converted into a tibble. Must have at least five columns that hold the family identifier, the personal identifier, the role and the lower and upper thresholds. Note that the role must be one of the following abbreviations

g (Genetic component of full liability)

o (Full liability)

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to NULL.

family

A matrix, list or data frame that can be converted into a tibble. Must have at least two columns that hold the family identifier and the corresponding personal identifiers, respectively. That is, for each family in fam_id there should be a list holding all individuals belonging to that family in pid. Note that the personal identifiers for all individuals must have a special format. It must end on _?, where ? is one of the following abbreviations

g (Genetic component of full liability)

o (Full liability)

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). See also construct_covmat. Defaults to NULL.

threshs

A matrix, list or data frame that can be converted into a tibble. Must have at least three columns, one holding the personal identifier for all individuals, and the remaining two holding the lower and upper thresholds, respectively. The latter should be called "lower" and "upper", respectively. Defaults to NULL.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

pid

A string holding the name of the column in .tbl (or family and threshs) that hold the personal identifier(s). Defaults to "PID".

fam_id

A string holding the name of the column in .tbl or family that holds the family identifier. Defaults to "fam_ID".

role

A string holding the name of the column in .tbl that holds the role. Each role must be chosen from the following list of abbreviations

g (Genetic component of full liability)

o (Full liability)

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to "role".

out

A character or numeric vector indicating whether the genetic component of the full liability, the full liability or both should be returned. If out = c(1) or out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or out = c("full"), the full liability is estimated and returned. If out = c(1,2) or out = c("genetic", "full"), both components are estimated and returned. Defaults to c(1).

tol

A number that is used as the convergence criterion for the Gibbs sampler. Equals the standard error of the mean. That is, a tolerance of 0.2 means that the standard error of the mean is below 0.2. Defaults to 0.01.

Value

If family and threshs are two matrices, lists or data frames that can be converted into tibbles, if family has two columns named like the strings represented in pid and fam_id, if threshs has a column named like the string given in pid as well as a column named "lower" and a column named "upper" and if the liability-scale heritability h2, out, tol and always_add are of the required form, then the function returns a tibble with either four or six columns (depending on the length of out). The first two columns correspond to the columns fam_id and pid ' present in family. If out is equal to c(1) or c("genetic"), the third and fourth column hold the estimated genetic liability as well as the corresponding standard error, respectively. If out equals c(2) or c("full"), the third and fourth column hold the estimated full liability as well as the corresponding standard error, respectively. If out is equal to c(1,2) or c("genetic","full"), the third and fourth column hold the estimated genetic liability as well as the corresponding standard error, respectively, while the fifth and sixth column hold the estimated full liability as well as the corresponding standard error, respectively.

Details

This function can be used to estimate either the genetic component of the full liability, the full liability or both. It is possible to input either

See also

Examples

sims <- simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, 
add_ind = TRUE, h2 = 0.5, n_sim=10, pop_prev = .05)
#
estimate_liability_single(.tbl = sims$thresholds, 
h2 = 0.5, pid = "indiv_ID", fam_id = "fam_ID", role = "role", out = c(1), 
tol = 0.01)
#> The number of workers is 1
#> # A tibble: 10 × 4
#>    fam_ID    indiv_ID    genetic_est           genetic_se         
#>    <chr>     <chr>       <chr>                 <chr>              
#>  1 fam_ID_1  fam_ID_1_1  0.00505443674057094   0.00421100510517694
#>  2 fam_ID_2  fam_ID_2_1  0.469968821490876     0.00367676389745846
#>  3 fam_ID_3  fam_ID_3_1  -0.0337087205929126   0.00414079149581881
#>  4 fam_ID_4  fam_ID_4_1  -0.0118644613443411   0.00417119462190505
#>  5 fam_ID_5  fam_ID_5_1  -0.0293431188287592   0.00439244106786504
#>  6 fam_ID_6  fam_ID_6_1  -0.0134597430247327   0.00410643292831094
#>  7 fam_ID_7  fam_ID_7_1  -0.0365623754336157   0.00421198835040333
#>  8 fam_ID_8  fam_ID_8_1  -0.0330629908484133   0.00464612627560725
#>  9 fam_ID_9  fam_ID_9_1  -0.000154253638899973 0.00418233713734622
#> 10 fam_ID_10 fam_ID_10_1 -0.0327984271351036   0.004078823450613  
# 
sims <- simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, 
h2 = 0.5, n_sim=10, pop_prev = .05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#
estimate_liability_single(.tbl = sims$thresholds, 
h2 = 0.5, pid = "indiv_ID", fam_id = "fam_ID", role = "role",
out = c("genetic"), tol = 0.01)
#> The number of workers is 1
#> # A tibble: 10 × 4
#>    fam_ID    indiv_ID    genetic_est           genetic_se         
#>    <chr>     <chr>       <chr>                 <chr>              
#>  1 fam_ID_1  fam_ID_1_1  -0.00158578147944848  0.00402958731201032
#>  2 fam_ID_2  fam_ID_2_1  0.000298345606234469  0.00375684172639474
#>  3 fam_ID_3  fam_ID_3_1  -0.000349183930627193 0.00397644675614783
#>  4 fam_ID_4  fam_ID_4_1  0.000473282054658652  0.00378926103880689
#>  5 fam_ID_5  fam_ID_5_1  0.000345428670880936  0.00424465926219562
#>  6 fam_ID_6  fam_ID_6_1  -0.000979739502746221 0.0040094954917013 
#>  7 fam_ID_7  fam_ID_7_1  -0.00146408182261283  0.00378962075706592
#>  8 fam_ID_8  fam_ID_8_1  -0.00113193564436767  0.00377796129676452
#>  9 fam_ID_9  fam_ID_9_1  -0.00766287179262568  0.0037769918938736 
#> 10 fam_ID_10 fam_ID_10_1 0.00252500962858401   0.00377403219678656