Prepares input for estimate_liability

Prepares input for estimate_liability

prepare_LTFHPlus_input(
  family,
  CIP,
  CIP_columns = c("sex", "birth_year", "age"),
  status_col = "status",
  use_fixed_case_thr = F,
  fam_id_col = "fam_id",
  personal_id_col = "pid",
  role_col = "role"
)

Arguments

family

contains family and personal ids and role with a family.

CIP

tibble with population representative cumulative incidence proportions has the interpretation of "proportion of people that has experienced the trait subset by CIP_columns.

CIP_columns

the columns the CIPs are subset by, e.g. CIPs by birth_year, sex.

status_col

Column that contains the status of each family member

use_fixed_case_thr

Should the threshold be fixed for cases? Can be used if CIPs are detailed, e.g. stratified by birth_year and sex.

fam_id_col

Column that contains the family ID

personal_id_col

Column that contains the personal ID

role_col

Column that cotnains the role of each individual