Simulate under the liability threshold model.

simulate_under_LTM simulates thresholds under the liability threshold model for a given family structure.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if h2 is a number satisfying 0 <= h2, n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing three tibbles. The first tibble, sim_obs, holds the disease status and the current age/age of onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. The last tibble, fam_ID, connects the personal identifiers for all individuals to the family identifiers. As thresholds, fam_ID has the format required in estimate_liability. Note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age of onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called 'o'). If both fam_vec and n_fam are defined, the user is asked to decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 x 34
#>    indiv_ID      g      o       m      f     s1    mgm    mgf     pgm    pgf
#>    <chr>     <dbl>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 ID1      -0.810 -0.689  0.101   0.285 -0.653 -0.768  0.119  0.565  -0.823
#>  2 ID2       1.36   2.68   0.571   1.04  -1.16   0.381 -0.915  0.407   0.281
#>  3 ID3       0.548  0.522  0.224  -0.396 -1.86   1.05   0.700  1.19   -0.486
#>  4 ID4      -0.989 -0.366 -0.607  -0.418 -0.668 -1.21  -0.699  1.46   -1.19 
#>  5 ID5       0.887 -0.192  0.507   0.773  0.756 -0.143  0.893 -0.0917 -0.170
#>  6 ID6      -0.932 -1.88  -0.540  -1.23  -0.386 -2.00   0.718 -0.691  -2.21 
#>  7 ID7       0.207 -0.905 -0.0734 -0.102  0.331 -0.107  0.233  0.569  -1.65 
#>  8 ID8       0.622  0.531  0.208   1.73   0.131  0.123 -2.20   1.81    0.334
#>  9 ID9      -1.87  -1.77   0.950  -0.798  0.557 -0.961 -0.911 -0.696  -0.374
#> 10 ID10     -0.210 -1.22  -0.829  -0.163 -2.08   1.79  -0.785 -1.62    0.361
#> # ... with 990 more rows, and 24 more variables: o_status <lgl>,
#> #   m_status <lgl>, f_status <lgl>, s1_status <lgl>, mgm_status <lgl>,
#> #   mgf_status <lgl>, pgm_status <lgl>, pgf_status <lgl>, o_age <int>,
#> #   s1_age <int>, m_age <int>, f_age <int>, mgm_age <int>, mgf_age <int>,
#> #   pgm_age <dbl>, pgf_age <dbl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 x 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     3.24
#>  2 ID2_o     2.68  2.68
#>  3 ID3_o  -Inf     3.24
#>  4 ID4_o  -Inf     3.48
#>  5 ID5_o  -Inf     3.38
#>  6 ID6_o  -Inf     3.38
#>  7 ID7_o  -Inf     2.95
#>  8 ID8_o  -Inf     3.14
#>  9 ID9_o  -Inf     3.03
#> 10 ID10_o -Inf     2.83
#> # ... with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # ... with 990 more rows
#> 
simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> Error in mutate(., fam_ID = paste0(indiv_ID, "_", i), upper = convert_age_to_thresh(!!as.symbol(paste0(i,     "_aoo")), dist = "logistic", pop_prev = pop_prev, mid_point = 60,     slope = 1/8), lower = ifelse(!!as.symbol(paste0(i, "_status")),     convert_age_to_thresh(!!as.symbol(paste0(i, "_aoo")), dist = "logistic",         pop_prev = pop_prev, mid_point = 60, slope = 1/8), -Inf)): Problem while computing `upper = convert_age_to_thresh(...)`.
#> i The error occurred in row 149.
#> Caused by error in `convert_age_to_thresh()`:
#> ! The age must be non-negative!
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 x 13
#>    indiv_ID      m       f     s1 m_status f_status s1_status s1_age m_age f_age
#>    <chr>     <dbl>   <dbl>  <dbl> <lgl>    <lgl>    <lgl>      <int> <int> <int>
#>  1 ID1       0.386  0.685  -0.219 FALSE    FALSE    FALSE         25    48    50
#>  2 ID2      -0.145 -1.02    0.397 FALSE    FALSE    FALSE         40    62    65
#>  3 ID3      -0.908 -0.868   0.981 FALSE    FALSE    FALSE         32    56    57
#>  4 ID4       0.696 -0.845   0.449 FALSE    FALSE    FALSE         15    44    35
#>  5 ID5      -0.294 -0.903  -0.628 FALSE    FALSE    FALSE         35    63    55
#>  6 ID6      -1.01  -0.399   0.415 FALSE    FALSE    FALSE         26    49    53
#>  7 ID7       0.111  0.0210  1.02  FALSE    FALSE    FALSE         18    36    39
#>  8 ID8      -0.287  1.67    1.07  FALSE    TRUE     FALSE         21    48    50
#>  9 ID9      -0.546 -0.431  -0.167 FALSE    FALSE    FALSE         33    57    61
#> 10 ID10      0.424 -0.310  -0.463 FALSE    FALSE    FALSE         40    60    60
#> # ... with 490 more rows, and 3 more variables: m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 1,500 x 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_m   -Inf  2.36
#>  2 ID2_m   -Inf  1.91
#>  3 ID3_m   -Inf  2.08
#>  4 ID4_m   -Inf  2.51
#>  5 ID5_m   -Inf  1.89
#>  6 ID6_m   -Inf  2.32
#>  7 ID7_m   -Inf  2.82
#>  8 ID8_m   -Inf  2.36
#>  9 ID9_m   -Inf  2.05
#> 10 ID10_m  -Inf  1.96
#> # ... with 1,490 more rows
#> 
#> $fam_ID
#> # A tibble: 500 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [3]>
#>  2 ID2      <chr [3]>
#>  3 ID3      <chr [3]>
#>  4 ID4      <chr [3]>
#>  5 ID5      <chr [3]>
#>  6 ID6      <chr [3]>
#>  7 ID7      <chr [3]>
#>  8 ID8      <chr [3]>
#>  9 ID9      <chr [3]>
#> 10 ID10     <chr [3]>
#> # ... with 490 more rows
#> 
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> $sim_obs
#> # A tibble: 200 x 6
#>    indiv_ID       g     o o_status o_age o_aoo
#>    <chr>      <dbl> <dbl> <lgl>    <int> <dbl>
#>  1 ID1       0.0283 0.738 FALSE       18    18
#>  2 ID2      -0.318  0.737 FALSE       19    19
#>  3 ID3      -0.428  0.465 FALSE       33    33
#>  4 ID4       0.248  0.639 FALSE       15    15
#>  5 ID5       0.138  0.237 FALSE       13    13
#>  6 ID6      -0.643  0.380 FALSE       10    10
#>  7 ID7       0.185  0.670 FALSE       19    19
#>  8 ID8       0.710  1.86  TRUE        17    64
#>  9 ID9      -0.162  1.08  FALSE       29    29
#> 10 ID10      1.02   0.617 FALSE       21    21
#> # ... with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 x 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     3.47
#>  2 ID2_o  -Inf     3.44
#>  3 ID3_o  -Inf     2.94
#>  4 ID4_o  -Inf     3.57
#>  5 ID5_o  -Inf     3.63
#>  6 ID6_o  -Inf     3.73
#>  7 ID7_o  -Inf     3.44
#>  8 ID8_o     1.86  1.86
#>  9 ID9_o  -Inf     3.09
#> 10 ID10_o -Inf     3.37
#> # ... with 190 more rows
#> 
#> $fam_ID
#> # A tibble: 200 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [1]>
#>  2 ID2      <chr [1]>
#>  3 ID3      <chr [1]>
#>  4 ID4      <chr [1]>
#>  5 ID5      <chr [1]>
#>  6 ID6      <chr [1]>
#>  7 ID7      <chr [1]>
#>  8 ID8      <chr [1]>
#>  9 ID9      <chr [1]>
#> 10 ID10     <chr [1]>
#> # ... with 190 more rows
#>