Simulate under the liability threshold model.

simulate_under_LTM simulates families and thresholds under the liability threshold model for a given family structure and a variable number of phenotypes.Please note that it is not possible to simulate different family structures.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  genetic_corrmat = NULL,
  full_corrmat = NULL,
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

Either a number or a numeric vector holding the liability-scale heritability(ies) for one or more phenotypes. All entries in h2 must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to 0.5.

genetic_corrmat

Either NULL or a numeric matrix holding the genetic correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in genetic_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

full_corrmat

Either NULL or a numeric matrix holding the full correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in full_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

phen_names

Either NULL or character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. Must be specified, if length(h2) \(> 0\), and will be ignored if h2 is a number. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

Either a number or a numeric vector holding the population prevalence(s), i.e. the overall prevalence(s) in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing the following lists. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

Details

This function can be used to simulate the case-control status, the current age and age-of-onset as well as the lower and upper thresholds for a variable number of phenotypes for all family members in each of the n_sim families. If h2 is a number, simulate_under_LTM simulates the case- control status, the current age and age-of-onset as well as thresholds for a single phenotype. However, if h2 is a numeric vector, if genetic_corrmat and full_corrmat are two symmetric correlation matrices, and if phen_names and pop_prev are to numeric vectors holding the phenotype names and the population prevalences, respectively, then simulate_under_LTM simulates the case-control status, the current age and age-of-onset as well as thresholds for two or more (correlated) phenotypes. The family members can be specified using one of two possible formats.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID       g       o      m      f     s1     mgm     mgf      pgm      pgf
#>    <chr>    <dbl>   <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>    <dbl>    <dbl>
#>  1 fam_I… -0.806  -1.62    0.662 -1.80  -0.448 -0.0462 -0.775  -0.533   -2.10   
#>  2 fam_I…  1.23    0.611   0.808  1.19   1.29  -1.27    1.56   -0.110   -0.0405 
#>  3 fam_I…  0.181   0.577  -0.535  1.32   0.207 -0.331   0.858  -0.316   -0.174  
#>  4 fam_I… -0.564  -0.260   0.803 -0.182  0.187 -1.01   -0.562  -0.172    0.743  
#>  5 fam_I… -0.282  -1.92    0.127 -0.763  1.24   1.21   -0.272  -0.250    0.988  
#>  6 fam_I…  0.119  -0.102   1.78  -0.287  0.906  0.923  -0.779  -0.993   -0.0804 
#>  7 fam_I… -0.0839  0.0841  0.378 -0.427  0.587 -0.857   1.86   -0.00791  0.907  
#>  8 fam_I… -0.311   0.296   0.125  0.885 -0.657  0.552  -0.0966  0.274   -0.395  
#>  9 fam_I… -0.427  -0.963  -0.510 -0.335  0.305  0.333   0.868   0.493   -0.925  
#> 10 fam_I…  1.17    0.390   1.38   1.82   1.39  -0.0410  0.513   0.633    0.00820
#> # ℹ 990 more rows
#> # ℹ 16 more variables: o_status <lgl>, m_status <lgl>, f_status <lgl>,
#> #   s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, pgm_status <lgl>,
#> #   pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.99
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.17
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.24
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.55
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.28
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.68
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.14
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.52
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.14
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.10
#> # ℹ 7,990 more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID          g        o      m     mgm    mgf     s1     s2   mhs1    mhs2
#>    <chr>       <dbl>    <dbl>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1 fam_ID_1  -0.240  -0.998   -0.287  0.252   0.585 -1.02  -2.32  -1.30  -0.775 
#>  2 fam_ID_2   0.600   0.473    1.67  -0.840   2.16   2.25   0.717  1.03   2.66  
#>  3 fam_ID_3   0.0496  0.217   -0.833 -0.630   0.810  0.438 -1.30  -1.01   0.852 
#>  4 fam_ID_4   0.0678  0.286   -1.20  -0.527  -1.31   0.969  2.00  -0.340  0.617 
#>  5 fam_ID_5   0.289   0.856    1.35  -1.08    0.375 -0.357 -0.194 -0.502 -0.740 
#>  6 fam_ID_6   0.659   0.00457  0.396 -0.425   0.984  1.36   1.67   2.05   0.556 
#>  7 fam_ID_7  -0.459   0.256   -1.26  -0.0849 -0.454  0.708 -2.73  -0.738 -0.475 
#>  8 fam_ID_8   0.450   1.09     0.461  2.20   -0.126  1.44  -0.894  0.493  0.0352
#>  9 fam_ID_9   1.26    0.0626   0.921  0.295  -0.360  0.611  0.737 -0.247 -0.373 
#> 10 fam_ID_10  1.19    0.268   -0.286  0.371   1.31   0.336  0.136  0.343  0.0839
#> # ℹ 990 more rows
#> # ℹ 16 more variables: o_status <lgl>, m_status <lgl>, mgm_status <lgl>,
#> #   mgf_status <lgl>, s1_status <lgl>, s2_status <lgl>, mhs1_status <lgl>,
#> #   mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>,
#> #   s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.99
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.21
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.28
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.38
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.95
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.45
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.06
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.45
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  2.91
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  2.72
#> # ℹ 7,990 more rows
#> 

simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 200)
#> $sim_obs
#> # A tibble: 200 × 10
#>    fam_ID      m       f       s1 m_status f_status s1_status m_aoo f_aoo s1_aoo
#>    <chr>   <dbl>   <dbl>    <dbl> <lgl>    <lgl>    <lgl>     <dbl> <dbl>  <dbl>
#>  1 fam_I…  1.25   0.282   0.272   FALSE    FALSE    FALSE        53    41     23
#>  2 fam_I… -0.628 -0.510  -1.42    FALSE    FALSE    FALSE        36    38     10
#>  3 fam_I…  0.528  0.532   0.883   FALSE    FALSE    FALSE        30    34     10
#>  4 fam_I…  1.13   0.924  -1.88    FALSE    FALSE    FALSE        43    48     24
#>  5 fam_I…  1.46   0.818  -0.451   TRUE     FALSE    FALSE        68    46     16
#>  6 fam_I… -0.357  1.72    0.502   FALSE    TRUE     FALSE        56    58     33
#>  7 fam_I… -1.14   1.90    0.672   FALSE    TRUE     FALSE        62    53     38
#>  8 fam_I… -0.933 -0.209   0.00902 FALSE    FALSE    FALSE        63    62     40
#>  9 fam_I… -0.237 -1.08    0.421   FALSE    FALSE    FALSE        62    63     37
#> 10 fam_I…  0.240  0.0153  0.0532  FALSE    FALSE    FALSE        54    47     24
#> # ℹ 190 more rows
#> 
#> $thresholds
#> # A tibble: 600 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     1.89
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     2.59
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     2.83
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     2.30
#>  5 fam_ID_5  fam_ID_5_1  m        1.45  1.45
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     1.78
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     1.59
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     1.56
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     1.59
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     1.85
#> # ℹ 590 more rows
#> 

simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID          g      o o_status o_aoo
#>    <chr>       <dbl>  <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -1.23   -1.53  FALSE       36
#>  2 fam_ID_2   0.0461  0.742 FALSE       10
#>  3 fam_ID_3  -0.652  -0.754 FALSE       30
#>  4 fam_ID_4   0.703  -0.950 FALSE       29
#>  5 fam_ID_5   0.0253 -0.339 FALSE       21
#>  6 fam_ID_6   0.981   0.318 FALSE       36
#>  7 fam_ID_7  -0.305   0.103 FALSE       18
#>  8 fam_ID_8  -0.346   0.201 FALSE       36
#>  9 fam_ID_9   1.36    1.33  FALSE       27
#> 10 fam_ID_10 -0.673  -0.869 FALSE       17
#> # ℹ 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.82
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.73
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.05
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.09
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.37
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.82
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.47
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  2.82
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.16
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.50
#> # ℹ 190 more rows
#>