Simulate under the liability threshold model.

simulate_under_LTM simulates thresholds under the liability threshold model for a given family structure.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if h2 is a number satisfying 0 <= h2, n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing three tibbles. The first tibble, sim_obs, holds the disease status and the current age/age of onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. The last tibble, fam_ID, connects the personal identifiers for all individuals to the family identifiers. As thresholds, fam_ID has the format required in estimate_liability. Note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age of onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called 'o'). If both fam_vec and n_fam are defined, the user is asked to decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 34
#>    indiv_ID      g       o       m      f      s1     mgm     mgf     pgm
#>    <chr>     <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ID1       0.230 -0.0240  0.865   1.53  -0.0398  0.569  -1.66   -0.0315
#>  2 ID2      -1.38  -1.02   -1.67    1.19  -0.0301  0.179  -0.128  -0.811 
#>  3 ID3      -0.894 -0.685   0.0792 -0.759  0.0675  0.0560  0.337  -1.61  
#>  4 ID4       0.609  0.444  -0.539  -0.199 -0.344   0.0822  0.0840  0.111 
#>  5 ID5       0.914  1.57    0.556   0.656  1.33   -0.722  -0.515  -1.15  
#>  6 ID6       0.940  1.80    0.102   1.68   0.522  -0.670   0.807  -0.316 
#>  7 ID7       0.210 -0.0389  1.73   -0.664  1.89    0.0877  1.54    0.683 
#>  8 ID8      -0.678  0.259   0.298   0.860 -1.30   -0.0839  0.822   0.573 
#>  9 ID9       1.14   1.22    0.189   1.40  -0.0585 -0.492   0.463   0.531 
#> 10 ID10     -0.155 -1.07    0.568   0.992  0.479  -0.648  -0.457   2.70  
#> # … with 990 more rows, and 25 more variables: pgf <dbl>, o_status <lgl>,
#> #   m_status <lgl>, f_status <lgl>, s1_status <lgl>, mgm_status <lgl>,
#> #   mgf_status <lgl>, pgm_status <lgl>, pgf_status <lgl>, o_age <int>,
#> #   s1_age <int>, m_age <int>, f_age <int>, mgm_age <int>, mgf_age <int>,
#> #   pgm_age <dbl>, pgf_age <dbl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     3.45
#>  2 ID2_o  -Inf     3.03
#>  3 ID3_o  -Inf     2.47
#>  4 ID4_o  -Inf     3.06
#>  5 ID5_o     1.56  1.56
#>  6 ID6_o     1.81  1.81
#>  7 ID7_o  -Inf     2.72
#>  8 ID8_o  -Inf     3.14
#>  9 ID9_o  -Inf     3.24
#> 10 ID10_o -Inf     3.10
#> # … with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # … with 990 more rows
#> 
simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 34
#>    indiv_ID       g      o      m    mgm     mgf      s1     s2   mhs1    mhs2
#>    <chr>      <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl>
#>  1 ID1       0.160   0.546 -0.114 -0.397 -0.0815  0.317   0.790  1.58   1.41  
#>  2 ID2       0.313   0.157  1.59   0.614 -0.462   0.855  -1.01   0.910  1.99  
#>  3 ID3      -1.32   -2.02  -1.36  -0.331 -0.174  -1.18   -1.80   1.32  -0.431 
#>  4 ID4      -1.13    0.220  0.557  0.586 -1.17   -1.53    0.235  2.13   0.241 
#>  5 ID5       0.0912 -0.305  0.771 -0.272  1.01    0.364   0.163  0.332  1.34  
#>  6 ID6       0.749  -0.722  0.911  0.316  0.0850 -0.0698 -0.174 -0.615  0.288 
#>  7 ID7      -0.870  -0.867  1.13  -0.761  1.16   -1.56    1.52   0.644  0.667 
#>  8 ID8       0.480   0.581  0.324 -0.373 -0.773  -0.328  -0.336  0.343  0.0619
#>  9 ID9      -0.290   0.460  0.499 -0.926  0.836   0.984   2.52   0.657  1.77  
#> 10 ID10      0.689   0.960  0.819  2.66  -0.969   1.16    0.646  2.00   0.288 
#> # … with 990 more rows, and 24 more variables: o_status <lgl>, m_status <lgl>,
#> #   mgm_status <lgl>, mgf_status <lgl>, s1_status <lgl>, s2_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_age <int>, s1_age <int>,
#> #   s2_age <int>, mhs1_age <int>, mhs2_age <int>, m_age <int>, mgm_age <int>,
#> #   mgf_age <int>, o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>,
#> #   s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_o   -Inf  2.51
#>  2 ID2_o   -Inf  3.24
#>  3 ID3_o   -Inf  2.68
#>  4 ID4_o   -Inf  3.06
#>  5 ID5_o   -Inf  3.55
#>  6 ID6_o   -Inf  3.52
#>  7 ID7_o   -Inf  3.06
#>  8 ID8_o   -Inf  2.91
#>  9 ID9_o   -Inf  2.99
#> 10 ID10_o  -Inf  2.79
#> # … with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # … with 990 more rows
#> 
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 13
#>    indiv_ID       m       f     s1 m_status f_status s1_status s1_age m_age
#>    <chr>      <dbl>   <dbl>  <dbl> <lgl>    <lgl>    <lgl>      <int> <int>
#>  1 ID1      -0.836  -0.400  -0.277 FALSE    FALSE    FALSE         16    36
#>  2 ID2       0.832  -1.15   -0.901 FALSE    FALSE    FALSE         11    37
#>  3 ID3      -1.35   -0.463   1.71  FALSE    FALSE    TRUE          20    42
#>  4 ID4      -1.07   -0.229  -1.44  FALSE    FALSE    FALSE         31    59
#>  5 ID5       0.793  -2.51   -1.09  FALSE    FALSE    FALSE         14    34
#>  6 ID6      -1.09   -0.457  -2.50  FALSE    FALSE    FALSE         30    53
#>  7 ID7      -0.999   0.0263 -1.95  FALSE    FALSE    FALSE         25    53
#>  8 ID8      -0.0649  0.636   1.82  FALSE    FALSE    TRUE          22    47
#>  9 ID9       0.240   0.150  -0.997 FALSE    FALSE    FALSE         39    67
#> 10 ID10     -0.821  -0.290  -2.20  FALSE    FALSE    FALSE         35    62
#> # … with 490 more rows, and 4 more variables: f_age <int>, m_aoo <dbl>,
#> #   f_aoo <dbl>, s1_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 1,500 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_m   -Inf  2.82
#>  2 ID2_m   -Inf  2.79
#>  3 ID3_m   -Inf  2.59
#>  4 ID4_m   -Inf  1.99
#>  5 ID5_m   -Inf  2.90
#>  6 ID6_m   -Inf  2.18
#>  7 ID7_m   -Inf  2.18
#>  8 ID8_m   -Inf  2.40
#>  9 ID9_m   -Inf  1.81
#> 10 ID10_m  -Inf  1.91
#> # … with 1,490 more rows
#> 
#> $fam_ID
#> # A tibble: 500 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [3]>
#>  2 ID2      <chr [3]>
#>  3 ID3      <chr [3]>
#>  4 ID4      <chr [3]>
#>  5 ID5      <chr [3]>
#>  6 ID6      <chr [3]>
#>  7 ID7      <chr [3]>
#>  8 ID8      <chr [3]>
#>  9 ID9      <chr [3]>
#> 10 ID10     <chr [3]>
#> # … with 490 more rows
#> 
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> $sim_obs
#> # A tibble: 200 × 6
#>    indiv_ID       g      o o_status o_age o_aoo
#>    <chr>      <dbl>  <dbl> <lgl>    <int> <dbl>
#>  1 ID1      -0.0223 -0.415 FALSE       16    16
#>  2 ID2       0.929  -0.421 FALSE       13    13
#>  3 ID3       0.585   0.243 FALSE       39    39
#>  4 ID4       0.266   0.734 FALSE       22    22
#>  5 ID5      -0.332  -0.938 FALSE       36    36
#>  6 ID6       1.38    0.576 FALSE       40    40
#>  7 ID7      -0.497   0.159 FALSE       11    11
#>  8 ID8       1.13    1.36  FALSE       27    27
#>  9 ID9      -0.181  -0.171 FALSE       39    39
#> 10 ID10      1.48    1.18  FALSE       28    28
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_o   -Inf  3.54
#>  2 ID2_o   -Inf  3.63
#>  3 ID3_o   -Inf  2.71
#>  4 ID4_o   -Inf  3.33
#>  5 ID5_o   -Inf  2.82
#>  6 ID6_o   -Inf  2.67
#>  7 ID7_o   -Inf  3.70
#>  8 ID8_o   -Inf  3.16
#>  9 ID9_o   -Inf  2.71
#> 10 ID10_o  -Inf  3.12
#> # … with 190 more rows
#> 
#> $fam_ID
#> # A tibble: 200 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [1]>
#>  2 ID2      <chr [1]>
#>  3 ID3      <chr [1]>
#>  4 ID4      <chr [1]>
#>  5 ID5      <chr [1]>
#>  6 ID6      <chr [1]>
#>  7 ID7      <chr [1]>
#>  8 ID8      <chr [1]>
#>  9 ID9      <chr [1]>
#> 10 ID10     <chr [1]>
#> # … with 190 more rows
#>