Simulate under the liability threshold model.

simulate_under_LTM simulates thresholds under the liability threshold model for a given family structure.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if h2 is a number satisfying 0 <= h2, n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing three tibbles. The first tibble, sim_obs, holds the disease status and the current age/age of onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. The last tibble, fam_ID, connects the personal identifiers for all individuals to the family identifiers. As thresholds, fam_ID has the format required in estimate_liability. Note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age of onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called 'o'). If both fam_vec and n_fam are defined, the user is asked to decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 34
#>    indiv_ID       g      o         m      f      s1    mgm    mgf     pgm    pgf
#>    <chr>      <dbl>  <dbl>     <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 ID1       0.0262 -1.28  -0.0793   -0.766  2.18    0.793 -0.128  1.55   -1.24 
#>  2 ID2      -0.767   0.121  0.710    -1.73  -0.815   0.485  0.290 -0.967  -0.216
#>  3 ID3      -1.27   -1.02   0.143    -2.14   0.406   0.235  0.967 -1.38   -0.655
#>  4 ID4      -0.630  -1.99   0.378     0.857  0.235   0.755 -1.08  -0.802   2.26 
#>  5 ID5       0.375   0.532 -1.64      0.903 -0.145  -0.216 -0.583  0.0689  1.67 
#>  6 ID6       0.0884  0.677  0.525     0.231  0.386   1.74   1.14  -0.0308  0.654
#>  7 ID7      -0.114   0.333 -0.200     1.02  -0.270  -0.906 -0.464 -0.593   0.603
#>  8 ID8      -1.86   -1.84  -1.15      0.677 -1.25   -0.270 -0.580 -1.29    0.332
#>  9 ID9      -0.533   0.481  0.000619  1.66   0.0392 -1.39   0.311  0.469   0.511
#> 10 ID10      0.954   0.425  0.178     0.653 -1.23   -0.251  0.840 -0.0952 -0.852
#> # … with 990 more rows, and 24 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_age <int>, s1_age <int>, m_age <int>,
#> #   f_age <int>, mgm_age <int>, mgf_age <int>, pgm_age <dbl>, pgf_age <dbl>,
#> #   o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>, mgm_aoo <dbl>,
#> #   mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_o   -Inf  2.79
#>  2 ID2_o   -Inf  2.55
#>  3 ID3_o   -Inf  2.87
#>  4 ID4_o   -Inf  3.17
#>  5 ID5_o   -Inf  3.06
#>  6 ID6_o   -Inf  2.47
#>  7 ID7_o   -Inf  2.63
#>  8 ID8_o   -Inf  2.79
#>  9 ID9_o   -Inf  3.28
#> 10 ID10_o  -Inf  2.72
#> # … with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # … with 990 more rows
#> 
simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> Error in mutate(., fam_ID = paste0(indiv_ID, "_", i), upper = convert_age_to_thresh(!!as.symbol(paste0(i,     "_aoo")), dist = "logistic", pop_prev = pop_prev, mid_point = 60,     slope = 1/8), lower = ifelse(!!as.symbol(paste0(i, "_status")),     convert_age_to_thresh(!!as.symbol(paste0(i, "_aoo")), dist = "logistic",         pop_prev = pop_prev, mid_point = 60, slope = 1/8), -Inf)): Problem while computing `upper = convert_age_to_thresh(...)`.
#> ℹ The error occurred in row 128.
#> Caused by error in `convert_age_to_thresh()`:
#> ! The age must be non-negative!
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 13
#>    indiv_ID        m      f      s1 m_status f_status s1_status s1_age m_age
#>    <chr>       <dbl>  <dbl>   <dbl> <lgl>    <lgl>    <lgl>      <int> <int>
#>  1 ID1       0.00497 -2.03  -0.211  FALSE    FALSE    FALSE         34    64
#>  2 ID2       1.40     1.05  -0.657  FALSE    FALSE    FALSE         38    58
#>  3 ID3       0.119   -0.105  0.637  FALSE    FALSE    FALSE         37    62
#>  4 ID4       0.854   -0.309  0.375  FALSE    FALSE    FALSE         11    31
#>  5 ID5       0.194    1.12   1.34   FALSE    FALSE    FALSE         37    63
#>  6 ID6       0.348    0.600  0.0216 FALSE    FALSE    FALSE         13    34
#>  7 ID7      -1.30    -1.67  -1.05   FALSE    FALSE    FALSE         36    60
#>  8 ID8      -0.752    0.492 -0.774  FALSE    FALSE    FALSE         28    58
#>  9 ID9       0.0211  -1.26  -0.396  FALSE    FALSE    FALSE         30    49
#> 10 ID10     -0.397    0.502  0.807  FALSE    FALSE    FALSE         20    38
#> # … with 490 more rows, and 4 more variables: f_age <int>, m_aoo <dbl>,
#> #   f_aoo <dbl>, s1_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 1,500 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_m   -Inf  1.86
#>  2 ID2_m   -Inf  2.02
#>  3 ID3_m   -Inf  1.91
#>  4 ID4_m   -Inf  3.01
#>  5 ID5_m   -Inf  1.89
#>  6 ID6_m   -Inf  2.90
#>  7 ID7_m   -Inf  1.96
#>  8 ID8_m   -Inf  2.02
#>  9 ID9_m   -Inf  2.32
#> 10 ID10_m  -Inf  2.75
#> # … with 1,490 more rows
#> 
#> $fam_ID
#> # A tibble: 500 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [3]>
#>  2 ID2      <chr [3]>
#>  3 ID3      <chr [3]>
#>  4 ID4      <chr [3]>
#>  5 ID5      <chr [3]>
#>  6 ID6      <chr [3]>
#>  7 ID7      <chr [3]>
#>  8 ID8      <chr [3]>
#>  9 ID9      <chr [3]>
#> 10 ID10     <chr [3]>
#> # … with 490 more rows
#> 
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> $sim_obs
#> # A tibble: 200 × 6
#>    indiv_ID       g       o o_status o_age o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <int> <dbl>
#>  1 ID1      -0.392  -0.0360 FALSE       10    10
#>  2 ID2      -0.303  -1.61   FALSE       24    24
#>  3 ID3      -0.285  -0.322  FALSE       16    16
#>  4 ID4      -0.0218 -0.473  FALSE       15    15
#>  5 ID5      -0.140  -0.729  FALSE       40    40
#>  6 ID6      -0.616  -0.905  FALSE       32    32
#>  7 ID7      -0.749  -0.573  FALSE       12    12
#>  8 ID8       0.811   2.91   TRUE        24    34
#>  9 ID9      -0.0574 -0.0785 FALSE       24    24
#> 10 ID10     -0.881  -0.128  FALSE       16    16
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     3.73
#>  2 ID2_o  -Inf     3.26
#>  3 ID3_o  -Inf     3.54
#>  4 ID4_o  -Inf     3.57
#>  5 ID5_o  -Inf     2.67
#>  6 ID6_o  -Inf     2.97
#>  7 ID7_o  -Inf     3.67
#>  8 ID8_o     2.90  2.90
#>  9 ID9_o  -Inf     3.26
#> 10 ID10_o -Inf     3.54
#> # … with 190 more rows
#> 
#> $fam_ID
#> # A tibble: 200 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [1]>
#>  2 ID2      <chr [1]>
#>  3 ID3      <chr [1]>
#>  4 ID4      <chr [1]>
#>  5 ID5      <chr [1]>
#>  6 ID6      <chr [1]>
#>  7 ID7      <chr [1]>
#>  8 ID8      <chr [1]>
#>  9 ID9      <chr [1]>
#> 10 ID10     <chr [1]>
#> # … with 190 more rows
#>