Simulate under the liability threshold model.

simulate_under_LTM simulates thresholds under the liability threshold model for a given family structure.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if h2 is a number satisfying 0 <= h2, n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing three tibbles. The first tibble, sim_obs, holds the disease status and the current age/age of onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. The last tibble, fam_ID, connects the personal identifiers for all individuals to the family identifiers. As thresholds, fam_ID has the format required in estimate_liability. Note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age of onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called 'o'). If both fam_vec and n_fam are defined, the user is asked to decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 34
#>    indiv_ID      g       o       m       f     s1     mgm     mgf     pgm    pgf
#>    <chr>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>  <dbl>
#>  1 ID1       0.639  0.280  -0.360   0.170  -0.771 -0.351  -1.10    0.0713  0.521
#>  2 ID2      -0.642 -1.39   -0.910  -1.24    1.51   1.04   -1.26   -0.367  -0.581
#>  3 ID3      -0.839 -0.809  -0.497  -0.698  -0.137 -0.234  -0.0805  0.0369 -1.33 
#>  4 ID4       1.40   1.07    0.507   0.219  -1.30   0.188   1.63    2.79   -0.552
#>  5 ID5       0.143 -0.0638  0.179  -0.0154 -0.613  0.210   0.527   0.308  -1.73 
#>  6 ID6       0.517  0.619   0.0308 -0.333   0.388  0.397   0.809   1.20    0.530
#>  7 ID7      -0.626  0.0101 -1.75   -0.808  -3.02   0.0692  0.180   1.58    0.236
#>  8 ID8      -1.75  -1.01   -1.33   -1.69   -1.06  -0.885   0.365   0.161   0.390
#>  9 ID9      -1.30  -1.82    0.193   2.26   -0.526 -0.976  -1.58    2.72    0.492
#> 10 ID10     -0.157  0.248   1.19   -0.100  -0.336  0.993  -0.812  -0.303   0.346
#> # … with 990 more rows, and 24 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_age <int>, s1_age <int>, m_age <int>,
#> #   f_age <int>, mgm_age <int>, mgf_age <int>, pgm_age <dbl>, pgf_age <dbl>,
#> #   o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>, mgm_aoo <dbl>,
#> #   mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_o   -Inf  2.87
#>  2 ID2_o   -Inf  2.76
#>  3 ID3_o   -Inf  3.28
#>  4 ID4_o   -Inf  2.72
#>  5 ID5_o   -Inf  3.42
#>  6 ID6_o   -Inf  2.99
#>  7 ID7_o   -Inf  3.48
#>  8 ID8_o   -Inf  2.79
#>  9 ID9_o   -Inf  2.99
#> 10 ID10_o  -Inf  2.91
#> # … with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # … with 990 more rows
#> 
simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 34
#>    indiv_ID        g        o       m      mgm     mgf     s1      s2   mhs1
#>    <chr>       <dbl>    <dbl>   <dbl>    <dbl>   <dbl>  <dbl>   <dbl>  <dbl>
#>  1 ID1       0.251   -0.208   -0.0466  1.58    -0.758  -0.752 -0.450  -1.66 
#>  2 ID2      -0.292    0.0253   0.514  -0.754    0.632   2.76  -0.949   0.793
#>  3 ID3       0.00251  0.00920  0.473   0.141    0.992   0.517  1.19   -0.492
#>  4 ID4       0.108   -0.681   -1.01   -0.0731  -2.96   -1.53  -0.0370  0.992
#>  5 ID5      -0.221   -0.124    1.09    0.188   -0.326   0.474  1.41    0.212
#>  6 ID6      -0.0329  -0.0461   0.990  -2.55     0.678   0.851 -1.50   -0.642
#>  7 ID7      -0.0782   0.791   -2.32    0.00466 -1.97    0.532  0.223  -2.09 
#>  8 ID8      -0.297    0.0266   0.588   0.218   -0.102   0.419  1.56    1.08 
#>  9 ID9       0.334    1.07     1.05    2.89     0.0986  1.38   0.690  -0.114
#> 10 ID10      0.0930  -0.236   -0.499   0.0354  -0.877  -0.832 -1.53   -1.04 
#> # … with 990 more rows, and 25 more variables: mhs2 <dbl>, o_status <lgl>,
#> #   m_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, s1_status <lgl>,
#> #   s2_status <lgl>, mhs1_status <lgl>, mhs2_status <lgl>, o_age <int>,
#> #   s1_age <int>, s2_age <int>, mhs1_age <int>, mhs2_age <int>, m_age <int>,
#> #   mgm_age <int>, mgf_age <int>, o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>,
#> #   mgf_aoo <dbl>, s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_o   -Inf  2.91
#>  2 ID2_o   -Inf  2.59
#>  3 ID3_o   -Inf  2.72
#>  4 ID4_o   -Inf  2.72
#>  5 ID5_o   -Inf  3.17
#>  6 ID6_o   -Inf  3.28
#>  7 ID7_o   -Inf  3.35
#>  8 ID8_o   -Inf  2.59
#>  9 ID9_o   -Inf  3.52
#> 10 ID10_o  -Inf  3.14
#> # … with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # … with 990 more rows
#> 
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> Error in mutate(., fam_ID = paste0(indiv_ID, "_", i), upper = convert_age_to_thresh(!!as.symbol(paste0(i,     "_aoo")), dist = "logistic", pop_prev = pop_prev, mid_point = 60,     slope = 1/8), lower = ifelse(!!as.symbol(paste0(i, "_status")),     convert_age_to_thresh(!!as.symbol(paste0(i, "_aoo")), dist = "logistic",         pop_prev = pop_prev, mid_point = 60, slope = 1/8), -Inf)): Problem while computing `upper = convert_age_to_thresh(...)`.
#> ℹ The error occurred in row 294.
#> Caused by error in `convert_age_to_thresh()`:
#> ! The age must be non-negative!
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> $sim_obs
#> # A tibble: 200 × 6
#>    indiv_ID      g      o o_status o_age o_aoo
#>    <chr>     <dbl>  <dbl> <lgl>    <int> <dbl>
#>  1 ID1      -0.910 -0.580 FALSE       38    38
#>  2 ID2      -0.704 -0.418 FALSE       12    12
#>  3 ID3       1.18   2.44  TRUE        29    46
#>  4 ID4       0.274  1.03  FALSE       21    21
#>  5 ID5      -0.908 -0.796 FALSE       28    28
#>  6 ID6       0.232  0.510 FALSE       16    16
#>  7 ID7      -0.106 -0.645 FALSE       13    13
#>  8 ID8       0.279  1.22  FALSE       19    19
#>  9 ID9       0.607  1.09  FALSE       40    40
#> 10 ID10     -0.155 -0.555 FALSE       15    15
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     2.75
#>  2 ID2_o  -Inf     3.67
#>  3 ID3_o     2.44  2.44
#>  4 ID4_o  -Inf     3.37
#>  5 ID5_o  -Inf     3.12
#>  6 ID6_o  -Inf     3.54
#>  7 ID7_o  -Inf     3.63
#>  8 ID8_o  -Inf     3.44
#>  9 ID9_o  -Inf     2.67
#> 10 ID10_o -Inf     3.57
#> # … with 190 more rows
#> 
#> $fam_ID
#> # A tibble: 200 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [1]>
#>  2 ID2      <chr [1]>
#>  3 ID3      <chr [1]>
#>  4 ID4      <chr [1]>
#>  5 ID5      <chr [1]>
#>  6 ID6      <chr [1]>
#>  7 ID7      <chr [1]>
#>  8 ID8      <chr [1]>
#>  9 ID9      <chr [1]>
#> 10 ID10     <chr [1]>
#> # … with 190 more rows
#>