Simulate under the liability threshold model.

simulate_under_LTM simulates families and thresholds under the liability threshold model for a given family structure and a variable number of phenotypes.Please note that it is not possible to simulate different family structures.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  genetic_corrmat = NULL,
  full_corrmat = NULL,
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

Either a number or a numeric vector holding the liability-scale heritability(ies) for one or more phenotypes. All entries in h2 must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to 0.5.

genetic_corrmat

Either NULL or a numeric matrix holding the genetic correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in genetic_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

full_corrmat

Either NULL or a numeric matrix holding the full correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in full_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

phen_names

Either NULL or character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. Must be specified, if length(h2) \(> 0\), and will be ignored if h2 is a number. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

Either a number or a numeric vector holding the population prevalence(s), i.e. the overall prevalence(s) in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing the following lists. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

Details

This function can be used to simulate the case-control status, the current age and age-of-onset as well as the lower and upper thresholds for a variable number of phenotypes for all family members in each of the n_sim families. If h2 is a number, simulate_under_LTM simulates the case- control status, the current age and age-of-onset as well as thresholds for a single phenotype. However, if h2 is a numeric vector, if genetic_corrmat and full_corrmat are two symmetric correlation matrices, and if phen_names and pop_prev are to numeric vectors holding the phenotype names and the population prevalences, respectively, then simulate_under_LTM simulates the case-control status, the current age and age-of-onset as well as thresholds for two or more (correlated) phenotypes. The family members can be specified using one of two possible formats.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID        g       o       m      f      s1     mgm    mgf     pgm     pgf
#>    <chr>     <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>
#>  1 fam_ID… -0.837  -0.761   1.63   -1.12   0.0441  1.20   -1.26   0.657   0.468 
#>  2 fam_ID…  0.205   0.547  -1.59    0.527  0.433  -0.266   0.706  1.94    1.49  
#>  3 fam_ID…  0.0344  0.0817  0.186   0.681 -0.163  -0.189  -1.32  -1.55   -0.203 
#>  4 fam_ID…  1.04    0.778  -0.426   0.855  1.15    0.0357 -0.838 -0.488   1.18  
#>  5 fam_ID… -1.26   -1.80   -0.0794 -1.62  -1.80   -0.0520 -1.26  -0.401  -2.12  
#>  6 fam_ID… -0.815   0.221  -0.437   0.962  0.204   0.384   0.648 -0.406  -0.880 
#>  7 fam_ID… -0.118  -1.50   -0.342  -0.470  1.77   -1.91    2.37  -0.0912  1.67  
#>  8 fam_ID… -0.668  -1.70    0.528   0.353 -0.0896 -0.537  -1.52   0.155   1.77  
#>  9 fam_ID… -0.0925 -0.0624  0.744  -0.253 -0.695   1.04   -1.35   0.458   0.158 
#> 10 fam_ID…  0.694   0.861   0.0853  0.480  1.18    0.814   1.39  -1.15    0.0820
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.99
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.52
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.99
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.55
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.28
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.06
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.99
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  2.59
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  2.83
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  2.51
#> # … with 7,990 more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g       o      m     mgm     mgf       s1     s2   mhs1   mhs2
#>    <chr>      <dbl>   <dbl>  <dbl>   <dbl>   <dbl>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1 fam_ID_1  -0.398 -0.541  -0.229 -1.39    0.236  -1.33     0.167 -0.990 -0.439
#>  2 fam_ID_2   0.492  1.39    1.14   0.980   0.0763 -0.125    1.30  -0.429  0.629
#>  3 fam_ID_3  -0.767 -0.578  -0.113  1.28   -0.279   0.674    0.827 -1.13  -1.08 
#>  4 fam_ID_4  -0.900 -1.96    0.479  0.544  -0.161   0.534   -0.591  0.323 -1.10 
#>  5 fam_ID_5  -0.177  0.0107  0.666 -0.0401  0.124   1.17     1.54   0.846  0.478
#>  6 fam_ID_6  -0.849 -0.785  -0.739 -0.654   0.843  -0.437   -2.94  -0.447  0.667
#>  7 fam_ID_7   0.740 -0.751   0.830 -0.235   0.0242  1.10    -0.150 -1.09  -0.217
#>  8 fam_ID_8  -0.736 -0.965  -0.901  0.449  -1.42    0.00930 -0.313 -0.244 -1.05 
#>  9 fam_ID_9  -0.254 -0.287   0.284  0.499   0.251   0.218   -0.407 -0.209 -0.668
#> 10 fam_ID_10  0.188  0.135   0.135  1.20   -0.331  -1.61    -1.26   1.50  -0.745
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   mgm_status <lgl>, mgf_status <lgl>, s1_status <lgl>, s2_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>,
#> #   mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.91
#>  2 fam_ID_2  fam_ID_2_1  o        1.39  1.39
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.24
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     2.55
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.76
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     2.59
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.99
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.55
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     2.59
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.76
#> # … with 7,990 more rows
#> 

simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 200)
#> $sim_obs
#> # A tibble: 200 × 10
#>    fam_ID          m       f      s1 m_status f_sta…¹ s1_st…² m_aoo f_aoo s1_aoo
#>    <chr>       <dbl>   <dbl>   <dbl> <lgl>    <lgl>   <lgl>   <dbl> <dbl>  <dbl>
#>  1 fam_ID_1   0.0303  0.311  -0.510  FALSE    FALSE   FALSE      50    49     29
#>  2 fam_ID_2   0.871   0.863  -0.156  FALSE    FALSE   FALSE      38    38     13
#>  3 fam_ID_3  -0.124  -0.379  -1.28   FALSE    FALSE   FALSE      31    31     12
#>  4 fam_ID_4   1.36    0.135   0.688  TRUE     FALSE   FALSE      76    38     13
#>  5 fam_ID_5  -0.936  -0.566   0.487  FALSE    FALSE   FALSE      52    52     30
#>  6 fam_ID_6  -0.666   0.139  -0.122  FALSE    FALSE   FALSE      53    56     32
#>  7 fam_ID_7   1.00    0.430   1.30   FALSE    FALSE   TRUE       30    33     86
#>  8 fam_ID_8   0.886   0.731   0.242  FALSE    FALSE   FALSE      39    38     17
#>  9 fam_ID_9   0.673   0.0433 -1.04   FALSE    FALSE   FALSE      54    63     35
#> 10 fam_ID_10 -2.04    0.306   0.0785 FALSE    FALSE   FALSE      43    45     16
#> # … with 190 more rows, and abbreviated variable names ¹​f_status, ²​s1_status
#> 
#> $thresholds
#> # A tibble: 600 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     2.01
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     2.51
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     2.79
#>  4 fam_ID_4  fam_ID_4_1  m        1.35  1.35
#>  5 fam_ID_5  fam_ID_5_1  m     -Inf     1.93
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     1.89
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     2.83
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     2.47
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     1.85
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     2.30
#> # … with 590 more rows
#> 

simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID          g       o o_status o_aoo
#>    <chr>       <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -1.45   -0.917  FALSE       25
#>  2 fam_ID_2   0.700   1.59   FALSE       20
#>  3 fam_ID_3  -0.468  -2.44   FALSE       20
#>  4 fam_ID_4  -0.344  -0.631  FALSE       15
#>  5 fam_ID_5  -0.545  -1.56   FALSE       22
#>  6 fam_ID_6   0.849   0.392  FALSE       15
#>  7 fam_ID_7   0.0430  0.0715 FALSE       15
#>  8 fam_ID_8   1.02    2.08   TRUE        56
#>  9 fam_ID_9  -0.477  -0.830  FALSE       31
#> 10 fam_ID_10  0.196   0.620  FALSE       10
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.23
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.40
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.40
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.57
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.33
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.57
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     3.57
#>  8 fam_ID_8  fam_ID_8_1  o        2.08  2.08
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.01
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.73
#> # … with 190 more rows
#>