Simulate under the liability threshold model.

simulate_under_LTM simulates families and thresholds under the liability threshold model for a given family structure and a variable number of phenotypes.Please note that it is not possible to simulate different family structures.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  genetic_corrmat = NULL,
  full_corrmat = NULL,
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

Either a number or a numeric vector holding the liability-scale heritability(ies) for one or more phenotypes. All entries in h2 must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to 0.5.

genetic_corrmat

Either NULL or a numeric matrix holding the genetic correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in genetic_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

full_corrmat

Either NULL or a numeric matrix holding the full correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in full_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

phen_names

Either NULL or character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. Must be specified, if length(h2) \(> 0\), and will be ignored if h2 is a number. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

Either a number or a numeric vector holding the population prevalence(s), i.e. the overall prevalence(s) in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing the following lists. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

Details

This function can be used to simulate the case-control status, the current age and age-of-onset as well as the lower and upper thresholds for a variable number of phenotypes for all family members in each of the n_sim families. If h2 is a number, simulate_under_LTM simulates the case- control status, the current age and age-of-onset as well as thresholds for a single phenotype. However, if h2 is a numeric vector, if genetic_corrmat and full_corrmat are two symmetric correlation matrices, and if phen_names and pop_prev are to numeric vectors holding the phenotype names and the population prevalences, respectively, then simulate_under_LTM simulates the case-control status, the current age and age-of-onset as well as thresholds for two or more (correlated) phenotypes. The family members can be specified using one of two possible formats.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g       o       m      f      s1     mgm    mgf    pgm     pgf
#>    <chr>      <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl>
#>  1 fam_ID_1   0.245  0.777  -0.997   0.812 -0.434   0.0165  1.38   1.05  -0.0532
#>  2 fam_ID_2  -0.358 -1.32   -1.20   -1.10  -1.97   -1.12    1.42  -1.22  -1.34  
#>  3 fam_ID_3  -0.375  1.07    0.0547 -0.577  0.0238  0.0197  0.365 -0.879  1.28  
#>  4 fam_ID_4   0.182  0.345  -1.12   -0.870 -0.597  -1.06   -0.384 -2.19  -0.361 
#>  5 fam_ID_5  -0.615 -0.0831  0.345  -1.00   0.0263  1.19   -0.899 -0.147 -1.23  
#>  6 fam_ID_6  -0.132 -0.395   0.995   0.655  1.88    0.514   1.18  -0.687  0.729 
#>  7 fam_ID_7  -0.236  0.401   1.60   -1.78  -0.146   0.703   0.482 -1.95   0.824 
#>  8 fam_ID_8  -0.996 -1.64   -0.155  -0.615 -1.06    1.03    1.89   1.01  -0.409 
#>  9 fam_ID_9  -0.288 -0.682   1.32    0.487 -0.418  -0.299   0.550 -2.14   0.881 
#> 10 fam_ID_10 -0.244 -0.355  -0.854   0.930 -0.509   1.30   -1.47  -0.416  1.05  
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.17
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.17
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.14
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.76
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.47
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.55
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.21
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  2.63
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.24
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  2.43
#> # … with 7,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g        o      m    mgm    mgf     s1      s2   mhs1   mhs2
#>    <chr>      <dbl>    <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1 fam_ID_1  -0.181 -0.251   -0.193 -1.11   0.990 -0.277  1.12    0.996  2.69 
#>  2 fam_ID_2   1.62   1.91     0.375  1.38   0.110 -0.198  1.51   -0.135  2.19 
#>  3 fam_ID_3   0.474 -0.00478  0.215 -0.571  0.584 -0.754  0.186   1.92   0.408
#>  4 fam_ID_4  -0.233  0.0656   1.39   0.266  0.592 -1.04  -0.130  -0.412 -0.544
#>  5 fam_ID_5   0.464  1.72     1.06  -0.714  1.79  -0.317 -0.324  -0.492 -0.489
#>  6 fam_ID_6   0.707 -0.220   -0.463 -0.808  0.712  0.704  0.0929  0.654 -0.109
#>  7 fam_ID_7  -1.24  -1.49    -0.774  0.553  0.782 -1.02   2.57   -2.19  -0.464
#>  8 fam_ID_8  -0.408  1.06     0.140  1.33  -0.484 -0.131 -1.91   -0.763  0.467
#>  9 fam_ID_9   0.199 -0.0271   1.27  -0.698  0.409 -1.25   0.839   0.598 -0.284
#> 10 fam_ID_10  0.287  0.0946   0.867  1.49   1.33   0.455 -0.0541  0.170  1.02 
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   mgm_status <lgl>, mgf_status <lgl>, s1_status <lgl>, s2_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>,
#> #   mhs2_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.42
#>  2 fam_ID_2  fam_ID_2_1  o        1.93  1.93
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     2.95
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     2.76
#>  5 fam_ID_5  fam_ID_5_1  o        1.71  1.71
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.38
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.47
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.43
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     2.91
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.59
#> # … with 7,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 200)
#> $sim_obs
#> # A tibble: 200 × 10
#>    fam_ID         m      f     s1 m_status f_status s1_status m_aoo f_aoo s1_aoo
#>    <chr>      <dbl>  <dbl>  <dbl> <lgl>    <lgl>    <lgl>     <dbl> <dbl>  <dbl>
#>  1 fam_ID_1  -0.108  0.427 -0.621 FALSE    FALSE    FALSE        38    34     10
#>  2 fam_ID_2   0.386 -0.729  0.964 FALSE    FALSE    FALSE        59    58     34
#>  3 fam_ID_3  -0.192 -0.433  1.50  FALSE    FALSE    TRUE         51    58     65
#>  4 fam_ID_4  -1.02  -1.18   0.349 FALSE    FALSE    FALSE        55    55     36
#>  5 fam_ID_5   1.83   0.717  0.737 TRUE     FALSE    FALSE        54    68     40
#>  6 fam_ID_6   1.03  -0.537  0.651 FALSE    FALSE    FALSE        42    51     21
#>  7 fam_ID_7   1.29   1.19   1.21  TRUE     FALSE    FALSE        90    36     15
#>  8 fam_ID_8  -0.735  0.775  0.163 FALSE    FALSE    FALSE        44    41     22
#>  9 fam_ID_9   0.146  0.576  1.16  FALSE    FALSE    FALSE        46    34     16
#> 10 fam_ID_10  0.536  0.884 -0.186 FALSE    FALSE    FALSE        63    67     37
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 600 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     2.51
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     1.68
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     1.97
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     1.81
#>  5 fam_ID_5  fam_ID_5_1  m        1.85  1.85
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     2.34
#>  7 fam_ID_7  fam_ID_7_1  m        1.29  1.29
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     2.26
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     2.18
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     1.56
#> # … with 590 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID         g       o o_status o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -0.373  1.11   FALSE       38
#>  2 fam_ID_2  -1.01  -1.07   FALSE       14
#>  3 fam_ID_3  -1.13  -1.97   FALSE       33
#>  4 fam_ID_4  -1.46  -1.42   FALSE       19
#>  5 fam_ID_5  -0.546 -0.0687 FALSE       18
#>  6 fam_ID_6  -0.294 -1.88   FALSE       29
#>  7 fam_ID_7  -0.496 -0.0420 FALSE       10
#>  8 fam_ID_8   0.882  1.20   FALSE       30
#>  9 fam_ID_9  -0.859 -0.827  FALSE       29
#> 10 fam_ID_10 -0.339 -0.165  FALSE       31
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.75
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.60
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.94
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.44
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.47
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.09
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.73
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.05
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.09
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.01
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>