Simulate under the liability threshold model.

simulate_under_LTM simulates families and thresholds under the liability threshold model for a given family structure and a variable number of phenotypes.Please note that it is not possible to simulate different family structures.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  genetic_corrmat = NULL,
  full_corrmat = NULL,
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

Either a number or a numeric vector holding the liability-scale heritability(ies) for one or more phenotypes. All entries in h2 must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to 0.5.

genetic_corrmat

Either NULL or a numeric matrix holding the genetic correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in genetic_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

full_corrmat

Either NULL or a numeric matrix holding the full correlations between the desired phenotypes. Must be specified, if length(h2)\(>0\), and will be ignored if h2 is a number. All diagonal entries in full_corrmat must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to NULL.

phen_names

Either NULL or character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. Must be specified, if length(h2) \(> 0\), and will be ignored if h2 is a number. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

Either a number or a numeric vector holding the population prevalence(s), i.e. the overall prevalence(s) in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing the following lists. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

Details

This function can be used to simulate the case-control status, the current age and age-of-onset as well as the lower and upper thresholds for a variable number of phenotypes for all family members in each of the n_sim families. If h2 is a number, simulate_under_LTM simulates the case- control status, the current age and age-of-onset as well as thresholds for a single phenotype. However, if h2 is a numeric vector, if genetic_corrmat and full_corrmat are two symmetric correlation matrices, and if phen_names and pop_prev are to numeric vectors holding the phenotype names and the population prevalences, respectively, then simulate_under_LTM simulates the case-control status, the current age and age-of-onset as well as thresholds for two or more (correlated) phenotypes. The family members can be specified using one of two possible formats.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g       o        m      f     s1    mgm     mgf    pgm     pgf
#>    <chr>      <dbl>   <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 fam_ID_1  -1.27  -0.309   1.16     0.604  0.945 -0.541 -0.657  -0.956 -0.745 
#>  2 fam_ID_2   0.426 -0.729  -0.463    0.491  0.286  0.778  0.577  -0.911  1.60  
#>  3 fam_ID_3  -0.186 -0.660   0.393   -1.75  -1.02   0.564 -0.0731 -1.18  -0.0261
#>  4 fam_ID_4   0.433  1.04   -0.00120  1.33   0.665 -0.591  0.532  -0.478 -0.667 
#>  5 fam_ID_5  -0.227  0.176  -1.08    -0.951 -0.694 -0.385 -0.290  -1.57  -1.25  
#>  6 fam_ID_6   0.396  0.183  -0.421   -0.163 -0.783 -0.872  1.41    1.09   0.884 
#>  7 fam_ID_7  -0.681  0.0281 -0.248   -2.38  -0.652 -0.802  0.298   0.695 -0.637 
#>  8 fam_ID_8   0.592  1.11   -0.810   -0.431 -1.25  -1.16   1.28    0.123 -0.364 
#>  9 fam_ID_9   2.02   0.969   2.05     1.30   2.25   0.431  2.44    0.676  1.62  
#> 10 fam_ID_10  1.09   0.789  -0.265    1.83   2.57  -0.470  0.721  -0.424 -0.222 
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.38
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.38
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.03
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.68
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.72
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.63
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.17
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.14
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.42
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.31
#> # … with 7,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID          g      o      m    mgm     mgf     s1      s2     mhs1   mhs2
#>    <chr>       <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>    <dbl>  <dbl>
#>  1 fam_ID_1   0.183   0.781 -0.110  0.715  0.646  -0.143  0.202   0.382   -1.18 
#>  2 fam_ID_2  -1.34   -0.182 -1.13   0.551  0.168  -1.81  -1.80   -1.66    -0.732
#>  3 fam_ID_3   1.03    1.45   1.72   2.26   2.00    0.881  0.583   2.77     0.360
#>  4 fam_ID_4  -1.11   -1.87  -1.74  -1.46  -0.460  -0.111 -0.591  -1.59    -1.83 
#>  5 fam_ID_5   0.345   0.612  0.710  0.321 -0.0419 -0.499 -0.212  -0.793   -0.361
#>  6 fam_ID_6   0.559   0.277  0.930 -0.393 -0.807  -1.45  -0.186  -0.588    0.336
#>  7 fam_ID_7  -0.0794  0.705 -0.512 -1.66   0.764  -0.215  1.08   -0.00445  1.07 
#>  8 fam_ID_8   0.143   0.734 -0.807 -1.28  -1.20   -0.708  1.83   -0.520    2.34 
#>  9 fam_ID_9  -0.404   1.43   1.60   0.332 -0.843   0.700  0.0952  1.22    -0.278
#> 10 fam_ID_10  0.332   0.266  1.39  -0.726 -1.16   -0.141 -0.0991 -0.356   -0.706
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   mgm_status <lgl>, mgf_status <lgl>, s1_status <lgl>, s2_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>,
#> #   mhs2_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.51
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.59
#>  3 fam_ID_3  fam_ID_3_1  o        1.45  1.45
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     2.43
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.45
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.42
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.95
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.52
#>  9 fam_ID_9  fam_ID_9_1  o        1.44  1.44
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.47
#> # … with 7,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 200)
#> $sim_obs
#> # A tibble: 200 × 10
#>    fam_ID         m       f      s1 m_status f_status s1_st…¹ m_aoo f_aoo s1_aoo
#>    <chr>      <dbl>   <dbl>   <dbl> <lgl>    <lgl>    <lgl>   <dbl> <dbl>  <dbl>
#>  1 fam_ID_1  -0.626 -1.77   -0.562  FALSE    FALSE    FALSE      51    56     30
#>  2 fam_ID_2  -0.450 -0.406   0.0334 FALSE    FALSE    FALSE      35    28     10
#>  3 fam_ID_3   0.700  0.351  -1.61   FALSE    FALSE    FALSE      63    57     34
#>  4 fam_ID_4   0.832  0.454   0.425  FALSE    FALSE    FALSE      68    57     39
#>  5 fam_ID_5   1.31   0.0135  0.251  TRUE     FALSE    FALSE      84    53     24
#>  6 fam_ID_6   0.419 -0.0396 -0.499  FALSE    FALSE    FALSE      35    47     17
#>  7 fam_ID_7   0.249  1.86   -0.611  FALSE    TRUE     FALSE      35    54     12
#>  8 fam_ID_8   1.45  -0.484   1.34   TRUE     FALSE    TRUE       68    54     78
#>  9 fam_ID_9  -1.43  -1.07   -2.18   FALSE    FALSE    FALSE      44    37     19
#> 10 fam_ID_10  0.239 -1.22   -0.463  FALSE    FALSE    FALSE      44    42     20
#> # … with 190 more rows, and abbreviated variable name ¹​s1_status
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 600 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     1.97
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     2.63
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     1.56
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     1.45
#>  5 fam_ID_5  fam_ID_5_1  m        1.31  1.31
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     2.63
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     2.63
#>  8 fam_ID_8  fam_ID_8_1  m        1.45  1.45
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     2.26
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     2.26
#> # … with 590 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID          g       o o_status o_aoo
#>    <chr>       <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -0.548  -0.340  FALSE       31
#>  2 fam_ID_2   1.55    3.08   TRUE        29
#>  3 fam_ID_3  -1.36   -1.70   FALSE       31
#>  4 fam_ID_4  -0.253   0.467  FALSE       18
#>  5 fam_ID_5  -0.0936 -0.202  FALSE       33
#>  6 fam_ID_6   0.553   0.935  FALSE       10
#>  7 fam_ID_7   0.240   0.895  FALSE       26
#>  8 fam_ID_8  -0.848  -0.0947 FALSE       33
#>  9 fam_ID_9   0.465   0.346  FALSE       22
#> 10 fam_ID_10  0.299   0.108  FALSE       11
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.01
#>  2 fam_ID_2  fam_ID_2_1  o        3.09  3.09
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.01
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.47
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.94
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.73
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     3.19
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.94
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.33
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.70
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>