Simulate under the liability threshold model.

simulate_under_LTM simulates thresholds under the liability threshold model for a given family structure.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if h2 is a number satisfying 0 <= h2, n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing three tibbles. The first tibble, sim_obs, holds the disease status and the current age/age of onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. The last tibble, fam_ID, connects the personal identifiers for all individuals to the family identifiers. As thresholds, fam_ID has the format required in estimate_liability. Note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age of onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called 'o'). If both fam_vec and n_fam are defined, the user is asked to decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 x 34
#>    indiv_ID      g      o      m      f      s1    mgm    mgf     pgm     pgf
#>    <chr>     <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>   <dbl>
#>  1 ID1       0.774  1.46  -0.264 -0.113  0.750  -0.767 -0.164 -0.302  -1.21  
#>  2 ID2       0.286 -0.479  2.06   1.73   2.78    0.729  0.779  0.0285  0.859 
#>  3 ID3      -0.103 -0.297  0.321 -2.08  -0.837   0.375 -0.216 -2.04   -0.0839
#>  4 ID4       1.62   0.693  0.297  0.434  0.399   0.325 -1.06   0.910   0.399 
#>  5 ID5       1.51   0.486  1.13  -1.04   0.0481 -0.327  0.807 -0.133  -0.460 
#>  6 ID6       0.755  0.212  0.707 -0.392  0.382   0.376  0.797  0.245  -0.0321
#>  7 ID7      -0.496 -0.896 -0.858  0.557  0.165   0.431 -2.89  -0.229   0.0757
#>  8 ID8       1.56   1.56   3.23   1.45   2.45    2.81  -0.129 -0.349   0.602 
#>  9 ID9      -0.527 -0.578  0.553  0.198  0.483   2.36   1.20  -1.81   -1.60  
#> 10 ID10      0.659  0.559  0.762 -0.716  1.73    0.143 -0.497 -0.0929 -0.663 
#> # ... with 990 more rows, and 24 more variables: o_status <lgl>,
#> #   m_status <lgl>, f_status <lgl>, s1_status <lgl>, mgm_status <lgl>,
#> #   mgf_status <lgl>, pgm_status <lgl>, pgf_status <lgl>, o_age <int>,
#> #   s1_age <int>, m_age <int>, f_age <int>, mgm_age <int>, mgf_age <int>,
#> #   pgm_age <dbl>, pgf_age <dbl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 x 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o     1.47  1.47
#>  2 ID2_o  -Inf     2.68
#>  3 ID3_o  -Inf     2.76
#>  4 ID4_o  -Inf     3.03
#>  5 ID5_o  -Inf     2.43
#>  6 ID6_o  -Inf     2.83
#>  7 ID7_o  -Inf     3.31
#>  8 ID8_o     1.56  1.56
#>  9 ID9_o  -Inf     3.28
#> 10 ID10_o -Inf     2.55
#> # ... with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # ... with 990 more rows
#> 
simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $sim_obs
#> # A tibble: 1,000 x 34
#>    indiv_ID        g      o       m     mgm    mgf      s1      s2    mhs1
#>    <chr>       <dbl>  <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ID1       0.259    0.830 -0.822  -0.0468  1.81   0.564   1.19    0.483 
#>  2 ID2       0.205    0.396  0.516   0.209   1.39   0.551   1.56   -0.0817
#>  3 ID3      -0.896   -0.671  0.479   0.184  -0.853  0.0123 -1.02    1.53  
#>  4 ID4      -0.421   -0.386 -0.0383 -0.427  -1.03  -1.17    0.416  -0.937 
#>  5 ID5       0.467    0.448 -1.44    0.228  -2.11   0.395   1.03   -2.53  
#>  6 ID6       0.787    2.00   0.918   0.0491  0.475  0.843   1.83    1.12  
#>  7 ID7      -0.0697  -0.427 -0.0138  0.365   1.10   0.222  -0.479  -0.638 
#>  8 ID8       1.29     1.31   1.65   -0.470   0.185  1.71    1.03   -0.236 
#>  9 ID9       0.00926  1.22  -0.811  -0.439   0.363  0.780   0.0558  0.286 
#> 10 ID10     -1.12    -1.42  -0.714  -0.476  -0.446 -2.37    0.763   1.07  
#> # ... with 990 more rows, and 25 more variables: mhs2 <dbl>, o_status <lgl>,
#> #   m_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, s1_status <lgl>,
#> #   s2_status <lgl>, mhs1_status <lgl>, mhs2_status <lgl>, o_age <int>,
#> #   s1_age <int>, s2_age <int>, mhs1_age <int>, mhs2_age <int>, m_age <int>,
#> #   mgm_age <int>, mgf_age <int>, o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>,
#> #   mgf_aoo <dbl>, s1_aoo <dbl>, s2_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 x 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     2.43
#>  2 ID2_o  -Inf     3.03
#>  3 ID3_o  -Inf     3.48
#>  4 ID4_o  -Inf     2.76
#>  5 ID5_o  -Inf     3.03
#>  6 ID6_o     2.01  2.01
#>  7 ID7_o  -Inf     2.47
#>  8 ID8_o     1.31  1.31
#>  9 ID9_o  -Inf     2.91
#> 10 ID10_o -Inf     3.14
#> # ... with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # ... with 990 more rows
#> 
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 x 13
#>    indiv_ID       m      f     s1 m_status f_status s1_status s1_age m_age f_age
#>    <chr>      <dbl>  <dbl>  <dbl> <lgl>    <lgl>    <lgl>      <int> <int> <int>
#>  1 ID1       0.534  -0.304  1.76  FALSE    FALSE    TRUE          32    61    59
#>  2 ID2       1.41    0.540  0.336 FALSE    FALSE    FALSE         32    58    52
#>  3 ID3       0.186   0.134  0.104 FALSE    FALSE    FALSE         25    43    52
#>  4 ID4       0.0928 -2.75  -1.69  FALSE    FALSE    FALSE         30    51    48
#>  5 ID5      -0.943   0.186 -0.607 FALSE    FALSE    FALSE         24    47    43
#>  6 ID6      -0.887  -1.23  -0.934 FALSE    FALSE    FALSE         21    44    51
#>  7 ID7      -0.251  -0.210 -0.371 FALSE    FALSE    FALSE         29    56    49
#>  8 ID8      -0.894   1.78   0.705 FALSE    TRUE     FALSE         36    65    61
#>  9 ID9      -0.517  -0.369 -1.18  FALSE    FALSE    FALSE         14    32    44
#> 10 ID10     -0.674   1.04  -0.751 FALSE    FALSE    FALSE         40    70    60
#> # ... with 490 more rows, and 3 more variables: m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 1,500 x 3
#>    fam_ID lower upper
#>    <chr>  <dbl> <dbl>
#>  1 ID1_m   -Inf  1.93
#>  2 ID2_m   -Inf  2.02
#>  3 ID3_m   -Inf  2.55
#>  4 ID4_m   -Inf  2.25
#>  5 ID5_m   -Inf  2.40
#>  6 ID6_m   -Inf  2.51
#>  7 ID7_m   -Inf  2.08
#>  8 ID8_m   -Inf  1.84
#>  9 ID9_m   -Inf  2.97
#> 10 ID10_m  -Inf  1.76
#> # ... with 1,490 more rows
#> 
#> $fam_ID
#> # A tibble: 500 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [3]>
#>  2 ID2      <chr [3]>
#>  3 ID3      <chr [3]>
#>  4 ID4      <chr [3]>
#>  5 ID5      <chr [3]>
#>  6 ID6      <chr [3]>
#>  7 ID7      <chr [3]>
#>  8 ID8      <chr [3]>
#>  9 ID9      <chr [3]>
#> 10 ID10     <chr [3]>
#> # ... with 490 more rows
#> 
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> $sim_obs
#> # A tibble: 200 x 6
#>    indiv_ID       g       o o_status o_age o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <int> <dbl>
#>  1 ID1       1.22    0.494  FALSE       27    27
#>  2 ID2      -0.785  -0.703  FALSE       24    24
#>  3 ID3      -0.917  -1.53   FALSE       37    37
#>  4 ID4      -0.286  -1.23   FALSE       16    16
#>  5 ID5       0.789   2.30   TRUE        36    49
#>  6 ID6      -0.437  -1.33   FALSE       26    26
#>  7 ID7       0.0470  1.04   FALSE       11    11
#>  8 ID8       0.391   0.764  FALSE       17    17
#>  9 ID9      -0.137  -0.445  FALSE       30    30
#> 10 ID10     -0.193   0.0342 FALSE       38    38
#> # ... with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 x 3
#>    fam_ID   lower upper
#>    <chr>    <dbl> <dbl>
#>  1 ID1_o  -Inf     3.16
#>  2 ID2_o  -Inf     3.26
#>  3 ID3_o  -Inf     2.79
#>  4 ID4_o  -Inf     3.54
#>  5 ID5_o     2.32  2.32
#>  6 ID6_o  -Inf     3.19
#>  7 ID7_o  -Inf     3.70
#>  8 ID8_o  -Inf     3.50
#>  9 ID9_o  -Inf     3.05
#> 10 ID10_o -Inf     2.75
#> # ... with 190 more rows
#> 
#> $fam_ID
#> # A tibble: 200 x 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [1]>
#>  2 ID2      <chr [1]>
#>  3 ID3      <chr [1]>
#>  4 ID4      <chr [1]>
#>  5 ID5      <chr [1]>
#>  6 ID6      <chr [1]>
#>  7 ID7      <chr [1]>
#>  8 ID8      <chr [1]>
#>  9 ID9      <chr [1]>
#> 10 ID10     <chr [1]>
#> # ... with 190 more rows
#>