Simulate under the liability threshold model.

simulate_under_LTM simulates thresholds under the liability threshold model for a given family structure.

simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

mgp[1-2]* (Maternal grandparent)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

pgp[1-2]* (Paternal grandparent)

s[0-9]* (Full siblings)

c[0-9]* (children)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the heritability on liability scale for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if h2 is a number satisfying 0 <= h2, n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list containing three tibbles. The first tibble, sim_obs, holds the disease status and the current age/age of onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. The last tibble, fam_ID, connects the personal identifiers for all individuals to the family identifiers. As thresholds, fam_ID has the format required in estimate_liability. Note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age of onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called 'o'). If both fam_vec and n_fam are defined, the user is asked to decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM()
#> $sim_obs
#> # A tibble: 1,000 × 34
#>    indiv_ID      g       o       m       f     s1     mgm      mgf    pgm
#>    <chr>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>    <dbl>  <dbl>
#>  1 ID1      -0.358 -0.784  -0.603  -0.121  -1.79  -1.48   -0.629    0.845
#>  2 ID2       0.289 -0.193  -0.134   1.05    1.12   0.169  -0.137    1.90 
#>  3 ID3      -0.647  0.0391 -0.0940 -0.349  -0.314  0.764   0.176   -1.70 
#>  4 ID4       0.480  1.04    0.974   0.603  -1.52  -0.0314 -1.55     1.15 
#>  5 ID5      -0.658 -0.343  -0.576  -1.95   -1.18  -0.559  -0.613   -1.91 
#>  6 ID6      -0.937 -1.28    0.139  -0.185  -0.535 -1.86   -0.199   -0.195
#>  7 ID7       0.555  1.31    0.300  -0.974  -0.546 -1.19    0.0180   0.351
#>  8 ID8      -0.255 -0.615  -2.37    1.08   -0.484  1.06   -1.07    -1.02 
#>  9 ID9      -0.608 -0.524   0.980  -0.0450  0.992  1.38    0.597   -0.393
#> 10 ID10     -0.407  0.179  -0.373  -0.307  -1.96  -0.511   0.00160  0.314
#> # … with 990 more rows, and 25 more variables: pgf <dbl>, o_status <lgl>,
#> #   m_status <lgl>, f_status <lgl>, s1_status <lgl>, mgm_status <lgl>,
#> #   mgf_status <lgl>, pgm_status <lgl>, pgf_status <lgl>, o_age <int>,
#> #   s1_age <int>, m_age <int>, f_age <int>, mgm_age <int>, mgf_age <int>,
#> #   pgm_age <int>, pgf_age <int>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 3
#>    fam_ID         lower upper
#>    <chr>          <dbl> <dbl>
#>  1 ID1_member1  -Inf     2.68
#>  2 ID2_member1  -Inf     2.95
#>  3 ID3_member1  -Inf     2.91
#>  4 ID4_member1  -Inf     3.45
#>  5 ID5_member1  -Inf     2.95
#>  6 ID6_member1  -Inf     2.91
#>  7 ID7_member1     1.31  1.31
#>  8 ID8_member1  -Inf     2.79
#>  9 ID9_member1  -Inf     3.03
#> 10 ID10_member1 -Inf     3.03
#> # … with 7,990 more rows
#> 
#> $fam_ID
#> # A tibble: 1,000 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [8]>
#>  2 ID2      <chr [8]>
#>  3 ID3      <chr [8]>
#>  4 ID4      <chr [8]>
#>  5 ID5      <chr [8]>
#>  6 ID6      <chr [8]>
#>  7 ID7      <chr [8]>
#>  8 ID8      <chr [8]>
#>  9 ID9      <chr [8]>
#> 10 ID10     <chr [8]>
#> # … with 990 more rows
#> 
simulate_under_LTM(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> Error in mutate(., fam_ID = paste0(indiv_ID, "_member", nbr), upper = convert_age_to_thresh(!!as.symbol(paste0(i,     "_aoo")), dist = "logistic", pop_prev = pop_prev, mid_point = 60,     slope = 1/8), lower = ifelse(!!as.symbol(paste0(i, "_status")),     convert_age_to_thresh(!!as.symbol(paste0(i, "_aoo")), dist = "logistic",         pop_prev = pop_prev, mid_point = 60, slope = 1/8), -Inf)): Problem while computing `upper = convert_age_to_thresh(...)`.
#> ℹ The error occurred in row 23.
#> Caused by error in `convert_age_to_thresh()`:
#> ! The age must be non-negative!
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 13
#>    indiv_ID      m       f      s1 m_status f_status s1_status s1_age m_age
#>    <chr>     <dbl>   <dbl>   <dbl> <lgl>    <lgl>    <lgl>      <int> <int>
#>  1 ID1      -0.759  0.878  -0.0955 FALSE    FALSE    FALSE         33    51
#>  2 ID2       0.744 -0.167   0.0139 FALSE    FALSE    FALSE         17    44
#>  3 ID3       0.248 -1.61    0.0735 FALSE    FALSE    FALSE         25    54
#>  4 ID4       1.25  -1.46    0.0352 FALSE    FALSE    FALSE         31    49
#>  5 ID5       2.00  -0.0512  0.0782 TRUE     FALSE    FALSE         33    62
#>  6 ID6      -1.17  -1.52   -0.223  FALSE    FALSE    FALSE         11    35
#>  7 ID7      -0.928  0.219  -1.42   FALSE    FALSE    FALSE         30    53
#>  8 ID8       1.20  -0.594  -0.704  FALSE    FALSE    FALSE         38    64
#>  9 ID9      -1.36   0.919  -0.736  FALSE    FALSE    FALSE         13    37
#> 10 ID10      0.499 -0.824   0.0411 FALSE    FALSE    FALSE         27    48
#> # … with 490 more rows, and 4 more variables: f_age <int>, m_aoo <dbl>,
#> #   f_aoo <dbl>, s1_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 1,500 × 3
#>    fam_ID         lower upper
#>    <chr>          <dbl> <dbl>
#>  1 ID1_member1  -Inf     2.25
#>  2 ID2_member1  -Inf     2.51
#>  3 ID3_member1  -Inf     2.14
#>  4 ID4_member1  -Inf     2.32
#>  5 ID5_member1     1.99  1.99
#>  6 ID6_member1  -Inf     2.86
#>  7 ID7_member1  -Inf     2.18
#>  8 ID8_member1  -Inf     1.86
#>  9 ID9_member1  -Inf     2.79
#> 10 ID10_member1 -Inf     2.36
#> # … with 1,490 more rows
#> 
#> $fam_ID
#> # A tibble: 500 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [3]>
#>  2 ID2      <chr [3]>
#>  3 ID3      <chr [3]>
#>  4 ID4      <chr [3]>
#>  5 ID5      <chr [3]>
#>  6 ID6      <chr [3]>
#>  7 ID7      <chr [3]>
#>  8 ID8      <chr [3]>
#>  9 ID9      <chr [3]>
#> 10 ID10     <chr [3]>
#> # … with 490 more rows
#> 
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> $sim_obs
#> # A tibble: 200 × 6
#>    indiv_ID       g       o o_status o_age o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <int> <dbl>
#>  1 ID1       0.723   1.00   FALSE       32    32
#>  2 ID2       0.0696 -0.908  FALSE       16    16
#>  3 ID3       1.08    0.926  FALSE       32    32
#>  4 ID4       0.103  -0.0424 FALSE       25    25
#>  5 ID5      -0.0858  0.370  FALSE       29    29
#>  6 ID6       0.843   1.52   FALSE       35    35
#>  7 ID7       0.0847  0.534  FALSE       25    25
#>  8 ID8      -1.37   -1.96   FALSE       31    31
#>  9 ID9      -0.0597  0.236  FALSE       24    24
#> 10 ID10     -0.842  -2.39   FALSE       28    28
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 3
#>    fam_ID       lower upper
#>    <chr>        <dbl> <dbl>
#>  1 ID1_member1   -Inf  2.97
#>  2 ID2_member1   -Inf  3.54
#>  3 ID3_member1   -Inf  2.97
#>  4 ID4_member1   -Inf  3.23
#>  5 ID5_member1   -Inf  3.09
#>  6 ID6_member1   -Inf  2.86
#>  7 ID7_member1   -Inf  3.23
#>  8 ID8_member1   -Inf  3.01
#>  9 ID9_member1   -Inf  3.26
#> 10 ID10_member1  -Inf  3.12
#> # … with 190 more rows
#> 
#> $fam_ID
#> # A tibble: 200 × 2
#>    indiv_ID fam_ID   
#>    <chr>    <list>   
#>  1 ID1      <chr [1]>
#>  2 ID2      <chr [1]>
#>  3 ID3      <chr [1]>
#>  4 ID4      <chr [1]>
#>  5 ID5      <chr [1]>
#>  6 ID6      <chr [1]>
#>  7 ID7      <chr [1]>
#>  8 ID8      <chr [1]>
#>  9 ID9      <chr [1]>
#> 10 ID10     <chr [1]>
#> # … with 190 more rows
#>