Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype1 o_phenotype1 m_phenotype1 f_phenotype1 s1_phenotype1
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        0.924        0.0789      -0.785        0.0173        0.0609
#>  2 fam_ID_2        0.0220       0.412       -0.0198      -1.09          0.595 
#>  3 fam_ID_3       -0.241        0.616        0.266       -0.436         0.248 
#>  4 fam_ID_4        0.648        0.458        0.811        1.15          2.06  
#>  5 fam_ID_5       -0.140        0.313        0.536        0.642         0.928 
#>  6 fam_ID_6       -0.255        0.171        0.0701       0.437         0.525 
#>  7 fam_ID_7        0.728        0.599        1.74        -0.541         0.873 
#>  8 fam_ID_8       -0.262       -1.69         0.910       -0.378         0.0209
#>  9 fam_ID_9        0.205        0.894        1.52        -0.409         1.25  
#> 10 fam_ID_10       0.886       -0.286       -0.275       -0.609         2.03  
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype1 <dbl>, mgf_phenotype1 <dbl>,
#> #   pgm_phenotype1 <dbl>, pgf_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, f_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, pgm_phenotype1_status <lgl>,
#> #   pgf_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype2 o_phenotype2 m_phenotype2 f_phenotype2 s1_phenotype2
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1       -0.169        0.243        0.802       -0.172         1.31  
#>  2 fam_ID_2       -0.107       -0.505        0.321       -0.479        -0.481 
#>  3 fam_ID_3       -0.296       -0.341        0.574       -1.39          0.731 
#>  4 fam_ID_4        0.436       -0.0511      -0.273        0.871        -0.0458
#>  5 fam_ID_5        0.782        1.31         1.04        -0.0304        1.40  
#>  6 fam_ID_6        0.0749       0.305        0.544       -0.938         0.493 
#>  7 fam_ID_7        1.32         0.161       -0.0624      -0.147        -1.77  
#>  8 fam_ID_8        0.652        0.108       -0.317        0.291         1.11  
#>  9 fam_ID_9       -0.896       -1.12        -0.705       -0.324        -0.358 
#> 10 fam_ID_10       0.0349       0.316       -0.897        0.606        -0.133 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype2 <dbl>, mgf_phenotype2 <dbl>,
#> #   pgm_phenotype2 <dbl>, pgf_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, f_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, pgm_phenotype2_status <lgl>,
#> #   pgf_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype3 o_phenotype3 m_phenotype3 f_phenotype3 s1_phenotype3
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        -1.03       -1.60         1.64        -0.370        0.980  
#>  2 fam_ID_2        -0.428      -0.0118       0.0445      -0.0400      -0.194  
#>  3 fam_ID_3         0.766       1.95         0.125        1.92         1.39   
#>  4 fam_ID_4         0.691       0.192        0.412        0.926       -0.657  
#>  5 fam_ID_5         0.816       0.864       -0.214        0.427        1.11   
#>  6 fam_ID_6        -0.176      -0.972       -0.182       -0.848        0.540  
#>  7 fam_ID_7         1.36        1.92         0.178        3.01         0.737  
#>  8 fam_ID_8         0.523       0.870        0.997        0.149       -0.00371
#>  9 fam_ID_9         0.468      -1.96         1.86         0.751        0.598  
#> 10 fam_ID_10        0.870       0.927        0.135        1.18        -0.346  
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype3 <dbl>, mgf_phenotype3 <dbl>,
#> #   pgm_phenotype3 <dbl>, pgf_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, f_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, pgm_phenotype3_status <lgl>,
#> #   pgf_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o                 -Inf             2.72          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 -Inf             3.10          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o                 -Inf             2.95          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 -Inf             2.72          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o                 -Inf             2.76             1.31
#>  6 fam_ID_6  fam_ID_6_1 o                 -Inf             3.24          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 -Inf             3.24          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o                 -Inf             3.14          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o                 -Inf             3.52          -Inf   
#> 10 fam_ID_10 fam_ID_10… o                 -Inf             3.35          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype1 o_phenotype1 m_phenotype1 mgm_phenotype1 mgf_phenotype1
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1       -0.350        0.337        0.539        -0.482          -0.467
#>  2 fam_ID_2        0.755        1.61        -0.183        -1.53           -0.429
#>  3 fam_ID_3        0.969        1.40         0.881         1.58            0.630
#>  4 fam_ID_4        0.977        2.37        -0.102         0.821          -0.325
#>  5 fam_ID_5        0.106        0.373        0.561        -0.580           0.164
#>  6 fam_ID_6        0.241       -0.760       -0.305        -0.652          -0.218
#>  7 fam_ID_7        0.241        0.131        0.771         2.41            0.918
#>  8 fam_ID_8       -0.568       -1.43        -1.09          0.238           0.342
#>  9 fam_ID_9       -1.19        -1.40        -1.69         -0.405           1.29 
#> 10 fam_ID_…        0.401        0.916       -0.464        -0.0809         -0.655
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype1 <dbl>, s2_phenotype1 <dbl>,
#> #   mhs1_phenotype1 <dbl>, mhs2_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   s2_phenotype1_status <lgl>, mhs1_phenotype1_status <lgl>,
#> #   mhs2_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype2 o_phenotype2 m_phenotype2 mgm_phenotype2 mgf_phenotype2
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1        0.192       -0.958      -0.121           0.880        -1.20  
#>  2 fam_ID_2        0.240       -0.392       0.634          -2.38         -0.303 
#>  3 fam_ID_3        0.591       -0.502       1.03           -0.283        -0.301 
#>  4 fam_ID_4       -0.283       -1.23       -1.63           -0.352         0.0572
#>  5 fam_ID_5       -1.11        -0.265      -0.470           0.321        -0.793 
#>  6 fam_ID_6       -0.755       -1.64       -1.07           -0.863        -0.454 
#>  7 fam_ID_7        0.643        0.949       1.19            1.39          0.868 
#>  8 fam_ID_8        0.759        1.96       -0.373          -0.840         0.0660
#>  9 fam_ID_9        0.599        1.08        0.447           2.07         -2.72  
#> 10 fam_ID_…       -0.180       -0.331      -0.0997         -0.257         0.704 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype2 <dbl>, s2_phenotype2 <dbl>,
#> #   mhs1_phenotype2 <dbl>, mhs2_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   s2_phenotype2_status <lgl>, mhs1_phenotype2_status <lgl>,
#> #   mhs2_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype3 o_phenotype3 m_phenotype3 mgm_phenotype3 mgf_phenotype3
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1      -0.219       -0.0355        0.178          1.45          0.0809
#>  2 fam_ID_2       0.232       -0.759         0.539          0.595        -1.62  
#>  3 fam_ID_3       0.857       -0.496         0.481          0.349        -1.63  
#>  4 fam_ID_4      -0.675       -0.661        -0.967         -0.192        -0.471 
#>  5 fam_ID_5       0.682        1.63          0.447         -0.247         0.917 
#>  6 fam_ID_6      -0.594       -0.102        -0.257         -0.265         0.0773
#>  7 fam_ID_7       0.382        1.07         -0.933         -1.44         -0.133 
#>  8 fam_ID_8      -1.22        -0.508         0.440         -0.170        -1.41  
#>  9 fam_ID_9      -0.0539       0.789        -1.35          -0.430        -1.05  
#> 10 fam_ID_…       0.0882      -0.913        -1.13           1.48          0.625 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype3 <dbl>, s2_phenotype3 <dbl>,
#> #   mhs1_phenotype3 <dbl>, mhs2_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   s2_phenotype3_status <lgl>, mhs1_phenotype3_status <lgl>,
#> #   mhs2_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                3.06          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 1.62             1.62          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o                 1.39             1.39          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 2.39             2.39          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                2.55          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                3.17          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                3.14          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                2.63             1.97
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                2.83          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                2.59          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype1 f_phenotype1 s1_phenotype1 m_phenotype1_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        -1.73       -0.180        -0.193  FALSE              
#>  2 fam_ID_2         0.940       1.40          1.03   FALSE              
#>  3 fam_ID_3        -0.139       0.649         1.90   FALSE              
#>  4 fam_ID_4         0.692       0.840         2.42   FALSE              
#>  5 fam_ID_5        -0.231      -1.46          0.519  FALSE              
#>  6 fam_ID_6         0.531       0.0210        0.356  FALSE              
#>  7 fam_ID_7         2.04        0.191         0.0817 TRUE               
#>  8 fam_ID_8         0.899       0.554         0.722  FALSE              
#>  9 fam_ID_9        -0.163       0.184         0.274  FALSE              
#> 10 fam_ID_10        0.484      -0.697         1.74   FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype2 f_phenotype2 s1_phenotype2 m_phenotype2_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1         0.826       0.957         0.858  FALSE              
#>  2 fam_ID_2         0.228       1.05          1.45   FALSE              
#>  3 fam_ID_3         0.591      -0.0131        1.39   FALSE              
#>  4 fam_ID_4        -0.322       0.333         0.651  FALSE              
#>  5 fam_ID_5         1.10        0.0884        1.31   FALSE              
#>  6 fam_ID_6        -0.481       0.0744        1.12   FALSE              
#>  7 fam_ID_7         0.887      -0.0138       -0.0612 FALSE              
#>  8 fam_ID_8         0.160      -0.304         0.293  FALSE              
#>  9 fam_ID_9         0.880      -0.211         0.765  FALSE              
#> 10 fam_ID_10       -0.164      -1.32          0.474  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype3 f_phenotype3 s1_phenotype3 m_phenotype3_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1         0.688       -0.394         0.774 FALSE              
#>  2 fam_ID_2         0.368        2.01          2.21  FALSE              
#>  3 fam_ID_3         0.125        0.833         1.76  FALSE              
#>  4 fam_ID_4        -1.06         1.18          1.13  FALSE              
#>  5 fam_ID_5         0.559       -1.75         -0.126 FALSE              
#>  6 fam_ID_6         1.06        -0.221         0.319 FALSE              
#>  7 fam_ID_7         0.617       -0.714        -0.634 FALSE              
#>  8 fam_ID_8         1.10         1.43          2.64  FALSE              
#>  9 fam_ID_9        -0.576        1.21          0.615 FALSE              
#> 10 fam_ID_10        1.01        -1.93          0.402 FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 m              -Inf                2.43             -Inf
#>  2 fam_ID_2  fam_ID_2_1 m              -Inf                1.97             -Inf
#>  3 fam_ID_3  fam_ID_3_1 m              -Inf                2.39             -Inf
#>  4 fam_ID_4  fam_ID_4_1 m              -Inf                2.51             -Inf
#>  5 fam_ID_5  fam_ID_5_1 m              -Inf                1.56             -Inf
#>  6 fam_ID_6  fam_ID_6_1 m              -Inf                1.89             -Inf
#>  7 fam_ID_7  fam_ID_7_1 m                 2.05             2.05             -Inf
#>  8 fam_ID_8  fam_ID_8_1 m              -Inf                1.81             -Inf
#>  9 fam_ID_9  fam_ID_9_1 m              -Inf                2.22             -Inf
#> 10 fam_ID_10 fam_ID_10… m              -Inf                2.05             -Inf
#> # ℹ 290 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.811        -0.796 FALSE                             39
#>  2 fam_ID_2        0.765         0.927 FALSE                             20
#>  3 fam_ID_3        0.586         0.783 FALSE                             11
#>  4 fam_ID_4        0.380         0.687 FALSE                             32
#>  5 fam_ID_5        0.0505        1.22  FALSE                             24
#>  6 fam_ID_6       -0.169         0.833 FALSE                             32
#>  7 fam_ID_7        0.630         1.35  TRUE                              77
#>  8 fam_ID_8       -0.695         1.47  TRUE                              67
#>  9 fam_ID_9       -0.491        -0.157 FALSE                             27
#> 10 fam_ID_10      -0.999        -0.696 FALSE                             15
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        1.67         1.96   TRUE                              51
#>  2 fam_ID_2       -0.875       -1.73   FALSE                             20
#>  3 fam_ID_3        0.891        2.22   TRUE                              45
#>  4 fam_ID_4        0.348       -0.946  FALSE                             32
#>  5 fam_ID_5        0.0107      -0.0545 FALSE                             24
#>  6 fam_ID_6        0.0881       0.112  FALSE                             32
#>  7 fam_ID_7        0.276        0.279  FALSE                             31
#>  8 fam_ID_8        0.100        0.444  FALSE                             24
#>  9 fam_ID_9        0.199        0.136  FALSE                             27
#> 10 fam_ID_10       1.22         0.448  FALSE                             15
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.356       0.00344 FALSE                             39
#>  2 fam_ID_2       -0.381      -1.03    FALSE                             20
#>  3 fam_ID_3        0.134      -1.28    FALSE                             11
#>  4 fam_ID_4       -0.0383     -0.611   FALSE                             32
#>  5 fam_ID_5        0.173       0.302   FALSE                             24
#>  6 fam_ID_6       -0.774      -1.25    FALSE                             32
#>  7 fam_ID_7       -0.565      -1.59    FALSE                             31
#>  8 fam_ID_8       -0.353       0.964   FALSE                             24
#>  9 fam_ID_9        0.922       0.498   FALSE                             27
#> 10 fam_ID_10       0.434      -0.0496  FALSE                             15
#> # ℹ 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                2.47             1.97
#>  2 fam_ID_2  fam_ID_2_1 o              -Inf                3.21          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                3.52             2.22
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                2.76          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                3.06          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                2.76          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 1.34             1.34          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o                 1.47             1.47          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                2.95          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                3.38          -Inf   
#> # ℹ 140 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#>