Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1        0.235  1.01     0.990  0.124   0.800    1.11   0.561   0.372 
#>  2 fam_ID_2       -0.990  0.0453  -1.24  -1.53    0.144    0.522  0.115  -1.47  
#>  3 fam_ID_3        0.142 -0.139    0.731 -0.572   0.717    1.07   0.540  -2.17  
#>  4 fam_ID_4        0.594  1.45     0.105  0.434   0.0151  -0.967 -1.69   -0.0351
#>  5 fam_ID_5        1.06   1.67     0.960  1.52    1.12    -0.826 -0.0183  1.66  
#>  6 fam_ID_6       -0.746 -1.72    -0.248 -0.0722 -1.06     0.983 -0.284  -0.677 
#>  7 fam_ID_7       -0.678 -0.302    0.249 -0.147  -0.538   -0.433 -0.957  -1.15  
#>  8 fam_ID_8        0.350  0.503   -0.455  0.716  -0.641    0.576 -0.682  -0.275 
#>  9 fam_ID_9       -0.263 -0.523   -0.285 -0.209  -0.0619   0.593 -0.361   2.01  
#> 10 fam_ID_10       0.184 -0.238   -1.31   1.06    0.200    0.520 -0.691  -1.03  
#> # … with 990 more rows, 17 more variables: pgf_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   pgm_phenotype1_status <lgl>, pgf_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>,
#> #   s1_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_ph…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1 fam_ID_1      0.344  -0.895   0.0873  2.02    1.01    -1.16   0.953   0.318  
#>  2 fam_ID_2     -0.377   0.0543  0.285  -1.65    0.472    1.31  -0.0862 -0.799  
#>  3 fam_ID_3     -0.0103  0.704  -0.0123 -0.253  -0.397    0.330  0.0827  1.64   
#>  4 fam_ID_4      0.883   0.871   0.783   0.0359  0.260    0.755  0.345   0.224  
#>  5 fam_ID_5      0.261  -0.0346  0.446   0.0999  0.463    1.93   0.0841 -1.06   
#>  6 fam_ID_6     -0.254  -0.463   1.45   -0.995  -0.632    1.06   0.0725 -0.145  
#>  7 fam_ID_7      1.13    1.42    0.263   1.49    1.91     0.697  0.0799 -0.214  
#>  8 fam_ID_8      1.02    0.728  -0.246   1.02    0.249    0.658  0.171  -0.00460
#>  9 fam_ID_9     -0.324  -0.0216  1.29    0.519  -0.202    0.554 -0.837  -0.966  
#> 10 fam_ID_10    -0.626  -0.356  -0.254   0.725   0.0856   0.340 -0.580   0.0587 
#> # … with 990 more rows, 17 more variables: pgf_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   pgm_phenotype2_status <lgl>, pgf_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>,
#> #   s1_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      0.320   -0.0571  0.206   -0.525   0.960  -1.03    0.547  -0.411
#>  2 fam_ID_2      0.113    0.173  -0.237    0.920   0.137  -0.337   0.802  -0.183
#>  3 fam_ID_3     -0.00587  1.05    0.522    0.258  -0.950   0.277  -1.80    0.106
#>  4 fam_ID_4      0.956    1.28   -0.535    1.83   -0.187  -0.345  -0.937   0.312
#>  5 fam_ID_5     -1.22    -0.560   0.840   -0.358  -1.17    0.653  -0.389  -0.921
#>  6 fam_ID_6      0.161   -0.324   0.990   -0.704   0.351  -0.441  -0.821  -0.667
#>  7 fam_ID_7      0.0534  -0.861  -0.0159   0.506   0.220   0.609   0.737  -1.60 
#>  8 fam_ID_8     -0.991   -0.482   0.0643  -0.137  -0.827   0.470   0.871   0.270
#>  9 fam_ID_9     -0.0275   1.16    0.170    0.566  -0.582   0.344   0.319   0.527
#> 10 fam_ID_10    -0.733   -1.23    0.372   -0.330  -0.612   0.354  -1.33    1.33 
#> # … with 990 more rows, 17 more variables: pgf_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   pgm_phenotype3_status <lgl>, pgf_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>,
#> #   s1_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       3.06 -Inf       3.06    -Inf    3.06
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.83 -Inf       2.83    -Inf    2.83
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.45 -Inf       3.45    -Inf    3.45
#>  4 fam_ID_4  fam_ID_4_1  o          1.45    1.45 -Inf       3.21    -Inf    3.21
#>  5 fam_ID_5  fam_ID_5_1  o          1.68    1.68 -Inf       3.42    -Inf    3.42
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       2.99 -Inf       2.99    -Inf    2.99
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       2.59    1.42    1.42    -Inf    2.59
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.52 -Inf       3.52    -Inf    3.52
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.83 -Inf       2.83    -Inf    2.83
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.87 -Inf       2.87    -Inf    2.87
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_pheno…¹ o_phe…² m_phe…³ mgm_ph…⁴ mgf_p…⁵ s1_phe…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>         <dbl>   <dbl>   <dbl>    <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1    -1.31   -0.758  -0.739  -0.509    0.0409 -0.791    -1.49   1.10  
#>  2 fam_ID_2     0.261   0.273   1.24   -0.376    0.375   0.324     0.295  0.0380
#>  3 fam_ID_3     1.11    1.51    1.32    0.0210  -1.70   -0.294     1.25   2.45  
#>  4 fam_ID_4    -0.0256  0.0944 -0.0386 -0.205    0.237  -0.298    -0.187 -0.0747
#>  5 fam_ID_5     0.628   0.579   0.771   0.00589  1.24    0.509     0.748 -0.865 
#>  6 fam_ID_6    -0.0694  0.334  -0.0148 -0.645    0.548  -0.343     0.923 -0.190 
#>  7 fam_ID_7    -0.232   0.870  -1.22   -0.933   -0.330  -2.31      0.474  0.341 
#>  8 fam_ID_8     0.0151 -1.64    0.889  -0.412   -0.561   0.00447   0.310  0.938 
#>  9 fam_ID_9    -0.697  -1.08   -1.38   -1.49     1.62   -0.200    -0.687 -1.48  
#> 10 fam_ID_10    0.101   1.39    0.519  -0.275    1.57    0.0731   -0.340  0.497 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, s2_phenotype1_status <lgl>,
#> #   mhs1_phenotype1_status <lgl>, mhs2_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>,
#> #   mgf_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ mgm_ph…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       2.26   1.44     1.06   1.55    -0.886   1.19     0.305  -0.538
#>  2 fam_ID_2       1.87   2.08     0.117 -0.00674  1.01    0.745    1.12    1.24 
#>  3 fam_ID_3      -0.247  0.0659  -1.04  -0.762    1.25   -0.138   -0.280   0.755
#>  4 fam_ID_4       1.07  -0.295    0.408  0.578    1.27    0.628    1.78    0.626
#>  5 fam_ID_5       0.657  0.957   -0.129 -1.80     0.266   0.985   -0.319  -1.44 
#>  6 fam_ID_6       0.202 -0.208   -0.161  0.130    1.60   -1.09     1.36    1.17 
#>  7 fam_ID_7      -0.466 -0.0350  -0.776 -0.0563  -0.130   0.592   -1.71   -0.350
#>  8 fam_ID_8      -0.204  0.378    1.76   0.944   -1.80   -0.156    0.793   0.440
#>  9 fam_ID_9       0.282  0.0608   2.45   0.790    1.34    0.0750   0.242   0.348
#> 10 fam_ID_10     -0.960  0.549   -0.348  0.314    0.0984  0.893   -1.36   -0.538
#> # … with 990 more rows, 17 more variables: mhs2_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, s2_phenotype2_status <lgl>,
#> #   mhs1_phenotype2_status <lgl>, mhs2_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>,
#> #   mgf_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       -0.308  -0.330  -0.235 -0.282  -0.0786  -0.861  -0.825  -0.739
#>  2 fam_ID_2        0.330   0.710  -0.674 -0.959   0.461    0.617   0.819  -0.367
#>  3 fam_ID_3       -1.07   -0.657   0.636  1.50    0.574   -2.66   -1.11    3.31 
#>  4 fam_ID_4        0.543   1.50    1.18   0.849  -0.0304  -0.989   1.84   -0.413
#>  5 fam_ID_5       -0.374  -0.190   1.05   0.210   0.290   -0.659  -0.174   0.365
#>  6 fam_ID_6        0.175   1.25   -0.415 -0.155   0.555    0.742   1.00   -1.50 
#>  7 fam_ID_7        1.40    0.952   1.14   0.701  -0.330    0.682   1.56    1.43 
#>  8 fam_ID_8        0.420   0.208   1.25  -0.0500  0.909    2.15    0.692   0.600
#>  9 fam_ID_9       -0.824  -0.779  -0.376 -1.13    1.59     1.36   -1.16   -1.86 
#> 10 fam_ID_10       1.31    1.91    0.519 -1.24    0.595    1.05    1.08   -0.596
#> # … with 990 more rows, 17 more variables: mhs2_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, s2_phenotype3_status <lgl>,
#> #   mhs1_phenotype3_status <lgl>, mhs2_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>,
#> #   mgf_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       3.35    1.44    1.44 -Inf       3.35
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.99    2.09    2.09 -Inf       2.99
#>  3 fam_ID_3  fam_ID_3_1  o          1.51    1.51 -Inf       3.21 -Inf       3.21
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       3.35 -Inf       3.35    1.51    1.51
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       2.47 -Inf       2.47 -Inf       2.47
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       2.51 -Inf       2.51 -Inf       2.51
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       2.55 -Inf       2.55 -Inf       2.55
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       2.59 -Inf       2.59 -Inf       2.59
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.51 -Inf       2.51 -Inf       2.51
#> 10 fam_ID_10 fam_ID_10_1 o          1.38    1.38 -Inf       2.99    1.93    1.93
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       -0.806   1.17    0.880 FALSE   FALSE   FALSE        43      42
#>  2 fam_ID_2       -1.02   -1.69    0.169 FALSE   FALSE   FALSE        38      30
#>  3 fam_ID_3       -0.482  -0.504  -0.273 FALSE   FALSE   FALSE        51      52
#>  4 fam_ID_4       -0.548   1.91   -0.453 FALSE   TRUE    FALSE        50      52
#>  5 fam_ID_5       -0.650   0.798   0.487 FALSE   FALSE   FALSE        41      39
#>  6 fam_ID_6       -0.175  -0.848   1.05  FALSE   FALSE   FALSE        33      29
#>  7 fam_ID_7        1.07    1.65    1.87  FALSE   TRUE    TRUE         49      60
#>  8 fam_ID_8       -0.217   0.267  -0.421 FALSE   FALSE   FALSE        42      48
#>  9 fam_ID_9        0.742  -0.241  -0.230 FALSE   FALSE   FALSE        46      44
#> 10 fam_ID_10      -0.967   0.188  -0.380 FALSE   FALSE   FALSE        46      48
#> # … with 90 more rows, 1 more variable: s1_phenotype1_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype1, ²​f_phenotype1, ³​s1_phenotype1,
#> #   ⁴​m_phenotype1_status, ⁵​f_phenotype1_status, ⁶​s1_phenotype1_status,
#> #   ⁷​m_phenotype1_aoo, ⁸​f_phenotype1_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       -1.83   0.643   0.705  FALSE   FALSE   FALSE        43      42
#>  2 fam_ID_2       -1.73  -0.684  -0.0359 FALSE   FALSE   FALSE        38      30
#>  3 fam_ID_3       -0.433 -2.27   -1.53   FALSE   FALSE   FALSE        51      52
#>  4 fam_ID_4        1.46   0.775  -0.692  TRUE    FALSE   FALSE        68      59
#>  5 fam_ID_5       -0.511  1.05    0.744  FALSE   FALSE   FALSE        41      39
#>  6 fam_ID_6        0.828 -0.0253  1.66   FALSE   FALSE   TRUE         33      29
#>  7 fam_ID_7        0.877  1.22    0.579  FALSE   FALSE   FALSE        49      41
#>  8 fam_ID_8       -0.125  0.808   0.0917 FALSE   FALSE   FALSE        42      48
#>  9 fam_ID_9        1.15   0.871   0.0907 FALSE   FALSE   FALSE        46      44
#> 10 fam_ID_10      -1.97  -0.922  -1.36   FALSE   FALSE   FALSE        46      48
#> # … with 90 more rows, 1 more variable: s1_phenotype2_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype2, ²​f_phenotype2, ³​s1_phenotype2,
#> #   ⁴​m_phenotype2_status, ⁵​f_phenotype2_status, ⁶​s1_phenotype2_status,
#> #   ⁷​m_phenotype2_aoo, ⁸​f_phenotype2_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       0.0937  -0.793   1.43  FALSE   FALSE   TRUE         43      42
#>  2 fam_ID_2      -0.337   -0.900  -0.263 FALSE   FALSE   FALSE        38      30
#>  3 fam_ID_3      -1.21     0.242  -0.678 FALSE   FALSE   FALSE        51      52
#>  4 fam_ID_4      -0.522    0.710  -0.320 FALSE   FALSE   FALSE        50      59
#>  5 fam_ID_5       0.251    1.72    0.567 FALSE   TRUE    FALSE        41      58
#>  6 fam_ID_6       1.09     0.357   1.17  FALSE   FALSE   FALSE        33      29
#>  7 fam_ID_7      -0.159    0.126  -1.09  FALSE   FALSE   FALSE        49      41
#>  8 fam_ID_8       0.196    0.934  -1.11  FALSE   FALSE   FALSE        42      48
#>  9 fam_ID_9      -1.59     0.329  -0.784 FALSE   FALSE   FALSE        46      44
#> 10 fam_ID_10      0.699   -0.216   0.410 FALSE   FALSE   FALSE        46      48
#> # … with 90 more rows, 1 more variable: s1_phenotype3_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype3, ²​f_phenotype3, ³​s1_phenotype3,
#> #   ⁴​m_phenotype3_status, ⁵​f_phenotype3_status, ⁶​s1_phenotype3_status,
#> #   ⁷​m_phenotype3_aoo, ⁸​f_phenotype3_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m          -Inf    2.30 -Inf       2.30    -Inf    2.30
#>  2 fam_ID_2  fam_ID_2_1  m          -Inf    2.51 -Inf       2.51    -Inf    2.51
#>  3 fam_ID_3  fam_ID_3_1  m          -Inf    1.97 -Inf       1.97    -Inf    1.97
#>  4 fam_ID_4  fam_ID_4_1  m          -Inf    2.01    1.45    1.45    -Inf    2.01
#>  5 fam_ID_5  fam_ID_5_1  m          -Inf    2.39 -Inf       2.39    -Inf    2.39
#>  6 fam_ID_6  fam_ID_6_1  m          -Inf    2.72 -Inf       2.72    -Inf    2.72
#>  7 fam_ID_7  fam_ID_7_1  m          -Inf    2.05 -Inf       2.05    -Inf    2.05
#>  8 fam_ID_8  fam_ID_8_1  m          -Inf    2.34 -Inf       2.34    -Inf    2.34
#>  9 fam_ID_9  fam_ID_9_1  m          -Inf    2.18 -Inf       2.18    -Inf    2.18
#> 10 fam_ID_10 fam_ID_10_1 m          -Inf    2.18 -Inf       2.18    -Inf    2.18
#> # … with 290 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1         0.109        0.637 FALSE                             19
#>  2 fam_ID_2        -0.887       -2.89  FALSE                             35
#>  3 fam_ID_3         0.477        1.45  TRUE                              68
#>  4 fam_ID_4        -0.377        0.756 FALSE                             30
#>  5 fam_ID_5         0.334       -0.486 FALSE                             32
#>  6 fam_ID_6        -0.575        0.934 FALSE                             38
#>  7 fam_ID_7        -0.313       -0.199 FALSE                             40
#>  8 fam_ID_8        -0.477       -0.795 FALSE                             10
#>  9 fam_ID_9        -0.966       -1.54  FALSE                             35
#> 10 fam_ID_10        0.599       -0.640 FALSE                             22
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.172        -0.158 FALSE                             19
#>  2 fam_ID_2        0.409         0.157 FALSE                             35
#>  3 fam_ID_3       -0.487        -2.51  FALSE                             19
#>  4 fam_ID_4       -0.673        -0.502 FALSE                             30
#>  5 fam_ID_5       -1.10         -0.471 FALSE                             32
#>  6 fam_ID_6       -0.274        -0.101 FALSE                             38
#>  7 fam_ID_7        0.0409        0.909 FALSE                             40
#>  8 fam_ID_8        0.268         1.45  TRUE                              68
#>  9 fam_ID_9        0.898        -0.131 FALSE                             35
#> 10 fam_ID_10       0.366        -0.173 FALSE                             22
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.929        -0.777 FALSE                             19
#>  2 fam_ID_2       -0.484        -0.634 FALSE                             35
#>  3 fam_ID_3       -0.923        -1.09  FALSE                             19
#>  4 fam_ID_4       -1.19         -0.708 FALSE                             30
#>  5 fam_ID_5        0.202         0.192 FALSE                             32
#>  6 fam_ID_6       -0.0184       -1.10  FALSE                             38
#>  7 fam_ID_7       -2.52         -2.83  FALSE                             40
#>  8 fam_ID_8        0.158         0.673 FALSE                             10
#>  9 fam_ID_9       -0.0798        1.02  FALSE                             35
#> 10 fam_ID_10      -1.00          0.556 FALSE                             22
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       3.24 -Inf       3.24    -Inf    3.24
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.63 -Inf       2.63    -Inf    2.63
#>  3 fam_ID_3  fam_ID_3_1  o          1.45    1.45 -Inf       3.24    -Inf    3.24
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       2.83 -Inf       2.83    -Inf    2.83
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       2.76 -Inf       2.76    -Inf    2.76
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       2.51 -Inf       2.51    -Inf    2.51
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       2.43 -Inf       2.43    -Inf    2.43
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.55    1.45    1.45    -Inf    3.55
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.63 -Inf       2.63    -Inf    2.63
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       3.14 -Inf       3.14    -Inf    3.14
#> # … with 140 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#>