Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       1.30    3.12    1.42     0.839 -0.137    1.33    0.446   0.151
#>  2 fam_ID_2      -0.352  -0.855   0.643   -0.818  0.0211  -0.464  -1.11   -0.357
#>  3 fam_ID_3       0.0823 -0.929  -0.312   -0.214 -0.448   -0.974  -2.00    0.715
#>  4 fam_ID_4      -1.22   -1.69   -1.19    -0.510 -0.996    0.687   1.20    0.101
#>  5 fam_ID_5      -0.244  -0.0947  0.117    0.560 -0.425   -0.259  -0.642  -0.742
#>  6 fam_ID_6       1.08    1.56    0.310    0.325 -0.504    0.777   0.784  -1.43 
#>  7 fam_ID_7      -0.703  -0.325  -0.0235  -0.973 -0.725   -0.515  -0.279  -0.563
#>  8 fam_ID_8      -0.697  -1.79   -0.507    0.543 -0.291   -0.349  -1.49   -0.434
#>  9 fam_ID_9       1.19    1.87    2.00     1.67   0.836    2.09   -0.110   0.433
#> 10 fam_ID_10      0.457   0.555  -1.08    -0.708 -0.522   -0.343   1.26    1.02 
#> # … with 990 more rows, 17 more variables: pgf_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   pgm_phenotype1_status <lgl>, pgf_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>,
#> #   s1_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.399  -0.193    1.17   0.0782   0.627   0.683   0.557 -1.44  
#>  2 fam_ID_2       0.197   1.04     0.466  0.590   -0.463   0.629  -0.145 -0.754 
#>  3 fam_ID_3      -0.0251  0.151    0.244  0.415    0.462   0.621   0.856 -0.338 
#>  4 fam_ID_4       0.429  -0.0616   1.78   0.474   -1.25    1.83   -0.671  0.372 
#>  5 fam_ID_5      -1.44   -1.61    -2.56  -0.224   -1.58    1.00   -2.13  -2.13  
#>  6 fam_ID_6       0.165   0.137   -0.375 -0.395   -1.27    1.31   -0.539 -0.0954
#>  7 fam_ID_7       0.271   0.0123  -0.397 -1.15    -0.252   0.410  -0.489  1.06  
#>  8 fam_ID_8       0.661   1.19    -0.219  1.70     0.416  -0.914  -0.855 -1.60  
#>  9 fam_ID_9       0.258   0.137   -0.376  1.66    -0.438   0.791   2.20   0.421 
#> 10 fam_ID_10     -0.829  -1.08    -0.425 -0.785   -0.953  -0.228  -1.22  -1.04  
#> # … with 990 more rows, 17 more variables: pgf_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   pgm_phenotype2_status <lgl>, pgf_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>,
#> #   s1_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_ph…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1 fam_ID_1    -0.0476   -1.19  -0.202   -0.876   1.22  -0.835   0.233   0.00687
#>  2 fam_ID_2    -0.401    -0.888  0.0172   0.617   0.419 -1.20    0.375  -0.462  
#>  3 fam_ID_3    -1.07     -0.467 -1.40     1.19   -0.875 -0.784  -0.148   0.753  
#>  4 fam_ID_4    -0.332    -1.04  -1.12     0.485  -2.21  -1.46    0.243  -0.664  
#>  5 fam_ID_5    -0.478    -0.613 -0.152   -0.563   1.15   0.0308  0.0698  0.373  
#>  6 fam_ID_6    -0.00248  -0.911  1.12    -0.530  -1.49   0.203  -0.0174 -0.826  
#>  7 fam_ID_7    -0.636    -1.25   1.05    -0.686  -1.95   1.36   -1.00    0.667  
#>  8 fam_ID_8    -1.17     -1.39  -0.439   -0.506  -1.02   1.78   -0.822   0.408  
#>  9 fam_ID_9    -0.104    -1.76   0.0772  -0.333   0.372  0.0141 -1.50    0.725  
#> 10 fam_ID_10    0.460     0.442  0.895   -0.845   1.41   1.76    0.835   0.0239 
#> # … with 990 more rows, 17 more variables: pgf_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   pgm_phenotype3_status <lgl>, pgf_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>,
#> #   s1_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          3.14    3.14    -Inf    2.68    -Inf    2.68
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.91    -Inf    2.91    -Inf    2.91
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       2.72    -Inf    2.72    -Inf    2.72
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       2.95    -Inf    2.95    -Inf    2.95
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       2.87    -Inf    2.87    -Inf    2.87
#>  6 fam_ID_6  fam_ID_6_1  o          1.56    1.56    -Inf    3.38    -Inf    3.38
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.10    -Inf    3.10    -Inf    3.10
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.52    -Inf    3.52    -Inf    3.52
#>  9 fam_ID_9  fam_ID_9_1  o          1.89    1.89    -Inf    3.06    -Inf    3.06
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.91    -Inf    2.91    -Inf    2.91
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_pheno…¹ o_phen…² m_phe…³ mgm_p…⁴ mgf_ph…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>         <dbl>    <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1    -0.873  -8.81e-4  -0.730 -0.289   0.00774  0.0200  -1.30    0.776
#>  2 fam_ID_2     1.45    8.25e-1   1.14  -0.274   0.200    0.562    1.29    1.13 
#>  3 fam_ID_3     0.348  -3.71e-1  -1.28   0.420   0.270    0.408   -0.176  -1.60 
#>  4 fam_ID_4     0.572   4.96e-1   0.346  0.375   1.96     2.15     0.716   0.961
#>  5 fam_ID_5     0.0894  1.06e+0   0.805  1.38   -0.0932   0.385   -0.469   0.550
#>  6 fam_ID_6    -0.765  -1.43e+0  -0.855 -0.0876 -0.628   -1.11    -0.263  -0.779
#>  7 fam_ID_7    -1.51   -1.20e+0  -0.758  0.554  -1.08    -1.89    -1.30   -1.32 
#>  8 fam_ID_8    -0.337  -2.08e-1  -0.557 -0.278   0.302   -0.170   -0.345   0.461
#>  9 fam_ID_9    -0.814  -9.99e-1  -2.22   0.817  -0.611   -1.90     0.364  -0.376
#> 10 fam_ID_10   -0.227  -3.47e-1  -0.913  0.122  -2.39     0.0658  -0.548   0.241
#> # … with 990 more rows, 17 more variables: mhs2_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, s2_phenotype1_status <lgl>,
#> #   mhs1_phenotype1_status <lgl>, mhs2_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>,
#> #   mgf_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phen…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      -0.106  -0.287  0.00405  -0.170  -0.286   0.135   1.04    0.616
#>  2 fam_ID_2       0.367   0.383 -1.50     -0.558  -0.316   0.172   1.64   -0.335
#>  3 fam_ID_3      -0.214   0.460 -1.94     -0.280  -1.10    0.972  -0.455  -0.120
#>  4 fam_ID_4       0.188  -0.342  0.445     1.33   -0.701  -0.265   0.162   2.02 
#>  5 fam_ID_5      -0.579  -0.422  0.429     0.533  -0.764   2.03    1.29    0.548
#>  6 fam_ID_6      -0.483   0.272  0.320    -0.947   1.47    0.208  -0.301  -0.868
#>  7 fam_ID_7       1.41    0.565 -1.70     -0.208  -0.426  -1.31    1.82   -0.230
#>  8 fam_ID_8       0.799   0.518 -1.53      0.868  -2.14    1.27    0.870  -1.48 
#>  9 fam_ID_9      -0.479  -0.350  1.19      0.914  -0.515  -1.04    0.426   1.23 
#> 10 fam_ID_10     -0.925  -1.55  -0.450    -0.671  -0.178  -0.341  -0.384  -1.06 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, s2_phenotype2_status <lgl>,
#> #   mhs1_phenotype2_status <lgl>, mhs2_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>,
#> #   mgf_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.0775 -0.903   -0.481 -1.05    0.0178 -1.29   -0.792   -1.90 
#>  2 fam_ID_2      -0.496   0.352    0.209 -1.15    0.0348 -0.822   0.0294  -1.50 
#>  3 fam_ID_3      -0.0818  0.0222  -1.35  -0.755   0.733  -0.397  -0.0666  -0.960
#>  4 fam_ID_4       0.862   2.86     0.356 -0.663   1.02    1.93    0.868    0.294
#>  5 fam_ID_5       0.587   0.997    2.31   1.10   -0.0859  0.352   0.365    0.712
#>  6 fam_ID_6       0.136  -0.971   -0.590  0.157  -0.581   0.520   0.174   -0.450
#>  7 fam_ID_7       1.22    2.60     0.651  1.26   -1.00    0.0931  0.141    0.727
#>  8 fam_ID_8       0.276   0.287    1.26  -0.0352 -1.87    1.61    1.27    -1.38 
#>  9 fam_ID_9      -0.146   0.0533  -0.734  0.186   0.179  -0.641   0.469    0.185
#> 10 fam_ID_10      0.224   1.81     1.65   0.615  -1.50    1.72    1.96     1.62 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, s2_phenotype3_status <lgl>,
#> #   mhs1_phenotype3_status <lgl>, mhs2_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>,
#> #   mgf_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          -Inf    2.51    -Inf    2.51 -Inf       2.51
#>  2 fam_ID_2  fam_ID_2_1  o          -Inf    3.10    -Inf    3.10 -Inf       3.10
#>  3 fam_ID_3  fam_ID_3_1  o          -Inf    2.59    -Inf    2.59 -Inf       2.59
#>  4 fam_ID_4  fam_ID_4_1  o          -Inf    3.52    -Inf    3.52    2.87    2.87
#>  5 fam_ID_5  fam_ID_5_1  o          -Inf    2.83    -Inf    2.83 -Inf       2.83
#>  6 fam_ID_6  fam_ID_6_1  o          -Inf    3.03    -Inf    3.03 -Inf       3.03
#>  7 fam_ID_7  fam_ID_7_1  o          -Inf    2.63    -Inf    2.63    2.59    2.59
#>  8 fam_ID_8  fam_ID_8_1  o          -Inf    2.72    -Inf    2.72 -Inf       2.72
#>  9 fam_ID_9  fam_ID_9_1  o          -Inf    2.76    -Inf    2.76 -Inf       2.76
#> 10 fam_ID_10 fam_ID_10_1 o          -Inf    3.21    -Inf    3.21    1.81    1.81
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenot…¹ f_phen…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>          <dbl>    <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1      -0.145 -0.945    -0.705 FALSE   FALSE   FALSE        58      58
#>  2 fam_ID_2      -1.05  -0.513    -0.731 FALSE   FALSE   FALSE        54      56
#>  3 fam_ID_3      -0.712 -0.945    -0.326 FALSE   FALSE   FALSE        43      40
#>  4 fam_ID_4       0.147 -1.65     -0.972 FALSE   FALSE   FALSE        56      52
#>  5 fam_ID_5       1.13   2.78      0.946 FALSE   TRUE    FALSE        46      31
#>  6 fam_ID_6      -0.338 -0.0401   -0.209 FALSE   FALSE   FALSE        48      50
#>  7 fam_ID_7       0.204 -0.00999   0.946 FALSE   FALSE   FALSE        57      50
#>  8 fam_ID_8      -0.228  0.160    -0.795 FALSE   FALSE   FALSE        39      39
#>  9 fam_ID_9      -0.398 -0.252    -0.149 FALSE   FALSE   FALSE        48      46
#> 10 fam_ID_10     -0.154 -0.539     0.478 FALSE   FALSE   FALSE        46      49
#> # … with 90 more rows, 1 more variable: s1_phenotype1_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype1, ²​f_phenotype1, ³​s1_phenotype1,
#> #   ⁴​m_phenotype1_status, ⁵​f_phenotype1_status, ⁶​s1_phenotype1_status,
#> #   ⁷​m_phenotype1_aoo, ⁸​f_phenotype1_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       0.443   -0.222  0.0206 FALSE   FALSE   FALSE        58      58
#>  2 fam_ID_2      -0.295   -0.834 -0.385  FALSE   FALSE   FALSE        54      56
#>  3 fam_ID_3      -2.06    -1.09  -0.704  FALSE   FALSE   FALSE        43      40
#>  4 fam_ID_4       0.0753  -0.834 -0.248  FALSE   FALSE   FALSE        56      52
#>  5 fam_ID_5      -0.249    1.81   0.830  FALSE   TRUE    FALSE        46      55
#>  6 fam_ID_6       0.106    0.272  1.55   FALSE   FALSE   TRUE         48      50
#>  7 fam_ID_7       0.170   -1.27   0.0933 FALSE   FALSE   FALSE        57      50
#>  8 fam_ID_8       0.372    1.64  -0.590  FALSE   TRUE    FALSE        39      60
#>  9 fam_ID_9      -0.998    0.188  1.13   FALSE   FALSE   FALSE        48      46
#> 10 fam_ID_10     -0.734   -0.930  0.450  FALSE   FALSE   FALSE        46      49
#> # … with 90 more rows, 1 more variable: s1_phenotype2_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype2, ²​f_phenotype2, ³​s1_phenotype2,
#> #   ⁴​m_phenotype2_status, ⁵​f_phenotype2_status, ⁶​s1_phenotype2_status,
#> #   ⁷​m_phenotype2_aoo, ⁸​f_phenotype2_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenot…¹ f_phe…² s1_phe…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>          <dbl>   <dbl>    <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1     -0.137  -1.80   -1.13    FALSE   FALSE   FALSE        58      58
#>  2 fam_ID_2      0.317  -0.0373  0.384   FALSE   FALSE   FALSE        54      56
#>  3 fam_ID_3      0.586  -1.36    0.00187 FALSE   FALSE   FALSE        43      40
#>  4 fam_ID_4     -0.525   0.454  -0.875   FALSE   FALSE   FALSE        56      52
#>  5 fam_ID_5     -0.0428  2.15    0.960   FALSE   TRUE    FALSE        46      47
#>  6 fam_ID_6      1.37    0.121  -0.395   TRUE    FALSE   FALSE        74      50
#>  7 fam_ID_7     -0.118  -1.10   -0.315   FALSE   FALSE   FALSE        57      50
#>  8 fam_ID_8     -0.700   1.20   -0.0595  FALSE   FALSE   FALSE        39      39
#>  9 fam_ID_9      1.28   -0.823   0.118   FALSE   FALSE   FALSE        48      46
#> 10 fam_ID_10     0.864   1.02    1.11    FALSE   FALSE   FALSE        46      49
#> # … with 90 more rows, 1 more variable: s1_phenotype3_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype3, ²​f_phenotype3, ³​s1_phenotype3,
#> #   ⁴​m_phenotype3_status, ⁵​f_phenotype3_status, ⁶​s1_phenotype3_status,
#> #   ⁷​m_phenotype3_aoo, ⁸​f_phenotype3_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m          -Inf    1.71    -Inf    1.71 -Inf       1.71
#>  2 fam_ID_2  fam_ID_2_1  m          -Inf    1.85    -Inf    1.85 -Inf       1.85
#>  3 fam_ID_3  fam_ID_3_1  m          -Inf    2.30    -Inf    2.30 -Inf       2.30
#>  4 fam_ID_4  fam_ID_4_1  m          -Inf    1.78    -Inf    1.78 -Inf       1.78
#>  5 fam_ID_5  fam_ID_5_1  m          -Inf    2.18    -Inf    2.18 -Inf       2.18
#>  6 fam_ID_6  fam_ID_6_1  m          -Inf    2.09    -Inf    2.09    1.37    1.37
#>  7 fam_ID_7  fam_ID_7_1  m          -Inf    1.74    -Inf    1.74 -Inf       1.74
#>  8 fam_ID_8  fam_ID_8_1  m          -Inf    2.47    -Inf    2.47 -Inf       2.47
#>  9 fam_ID_9  fam_ID_9_1  m          -Inf    2.09    -Inf    2.09 -Inf       2.09
#> 10 fam_ID_10 fam_ID_10_1 m          -Inf    2.18    -Inf    2.18 -Inf       2.18
#> # … with 290 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.843        1.38   TRUE                              73
#>  2 fam_ID_2        0.0184      -0.259  FALSE                             10
#>  3 fam_ID_3        1.24         1.24   FALSE                             10
#>  4 fam_ID_4       -0.168        0.0291 FALSE                             39
#>  5 fam_ID_5        0.432        1.49   TRUE                              66
#>  6 fam_ID_6       -0.167       -0.180  FALSE                             24
#>  7 fam_ID_7        0.616       -1.19   FALSE                             18
#>  8 fam_ID_8        0.286       -0.175  FALSE                             21
#>  9 fam_ID_9        0.196        1.65   TRUE                              60
#> 10 fam_ID_10      -0.471       -0.397  FALSE                             22
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1      -1.11          -1.38  FALSE                             24
#>  2 fam_ID_2       1.29           1.60  TRUE                              62
#>  3 fam_ID_3      -1.48          -0.712 FALSE                             10
#>  4 fam_ID_4      -0.182         -0.994 FALSE                             39
#>  5 fam_ID_5       0.0852        -0.110 FALSE                             25
#>  6 fam_ID_6       0.00752       -0.179 FALSE                             24
#>  7 fam_ID_7       0.207         -0.489 FALSE                             18
#>  8 fam_ID_8      -0.165         -0.646 FALSE                             21
#>  9 fam_ID_9       1.43           2.09  TRUE                              48
#> 10 fam_ID_10      0.956         -0.311 FALSE                             22
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.687         0.517 FALSE                             24
#>  2 fam_ID_2        0.331        -0.104 FALSE                             10
#>  3 fam_ID_3       -0.165         0.335 FALSE                             10
#>  4 fam_ID_4        0.983         0.292 FALSE                             39
#>  5 fam_ID_5       -0.223        -2.59  FALSE                             25
#>  6 fam_ID_6        0.813         1.74  TRUE                              57
#>  7 fam_ID_7        1.31          1.11  FALSE                             18
#>  8 fam_ID_8        0.0316       -0.662 FALSE                             21
#>  9 fam_ID_9       -1.45         -0.940 FALSE                             11
#> 10 fam_ID_10      -0.620         0.539 FALSE                             22
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          1.38    1.38 -Inf       3.06 -Inf       3.06
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       3.55    1.59    1.59 -Inf       3.55
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.55 -Inf       3.55 -Inf       3.55
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       2.47 -Inf       2.47 -Inf       2.47
#>  5 fam_ID_5  fam_ID_5_1  o          1.49    1.49 -Inf       3.03 -Inf       3.03
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       3.06 -Inf       3.06    1.74    1.74
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.28 -Inf       3.28 -Inf       3.28
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.17 -Inf       3.17 -Inf       3.17
#>  9 fam_ID_9  fam_ID_9_1  o          1.64    1.64    2.09    2.09 -Inf       3.52
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       3.14 -Inf       3.14 -Inf       3.14
#> # … with 140 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#>