Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype1 o_phenotype1 m_phenotype1 f_phenotype1 s1_phenotype1
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1       -0.606       -0.635        0.0453      -0.421        -0.657 
#>  2 fam_ID_2        0.317       -0.104       -0.979       -0.674         0.475 
#>  3 fam_ID_3        0.789        0.512        1.13        -1.85         -0.259 
#>  4 fam_ID_4       -1.31        -0.676       -1.09        -2.13         -0.327 
#>  5 fam_ID_5        0.0251       0.622       -1.35         1.10         -0.187 
#>  6 fam_ID_6        0.370        1.33         0.422        1.18         -1.13  
#>  7 fam_ID_7       -0.371        0.630        0.0943      -1.13         -1.16  
#>  8 fam_ID_8        0.911        0.725        2.40         0.0757        0.0724
#>  9 fam_ID_9        0.0146      -0.0849      -0.0323       0.540        -0.724 
#> 10 fam_ID_10       0.380        1.21         0.442       -0.979        -0.418 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype1 <dbl>, mgf_phenotype1 <dbl>,
#> #   pgm_phenotype1 <dbl>, pgf_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, f_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, pgm_phenotype1_status <lgl>,
#> #   pgf_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype2 o_phenotype2 m_phenotype2 f_phenotype2 s1_phenotype2
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1      -1.21        -0.478         -0.737      -1.50          -1.35 
#>  2 fam_ID_2      -0.359        0.338         -0.232      -0.866         -0.888
#>  3 fam_ID_3      -0.328       -1.43          -0.460       1.10           2.77 
#>  4 fam_ID_4      -1.35        -1.96          -1.38       -1.04          -0.961
#>  5 fam_ID_5       0.547       -0.00671        1.01        0.339          0.468
#>  6 fam_ID_6      -1.01        -0.922         -1.92       -0.0829        -0.761
#>  7 fam_ID_7       1.09        -0.279          0.785       1.40           2.87 
#>  8 fam_ID_8      -0.00288     -0.470         -0.197       0.149          1.11 
#>  9 fam_ID_9       0.562       -0.463         -0.578      -0.310          1.55 
#> 10 fam_ID_10     -0.371       -1.34          -0.704      -1.04          -0.738
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype2 <dbl>, mgf_phenotype2 <dbl>,
#> #   pgm_phenotype2 <dbl>, pgf_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, f_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, pgm_phenotype2_status <lgl>,
#> #   pgf_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype3 o_phenotype3 m_phenotype3 f_phenotype3 s1_phenotype3
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        -0.103       -0.676      -0.537        0.0549      -0.988  
#>  2 fam_ID_2        -1.04        -1.91       -0.495       -0.158       -0.736  
#>  3 fam_ID_3        -0.309       -1.09        0.652       -0.500        1.51   
#>  4 fam_ID_4        -0.615        0.332       0.191       -0.190        0.00154
#>  5 fam_ID_5         0.464       -0.322       1.14        -0.0221      -0.876  
#>  6 fam_ID_6        -0.177       -0.766      -0.340       -0.0266      -1.65   
#>  7 fam_ID_7        -0.936       -0.379      -0.934       -0.0669      -0.965  
#>  8 fam_ID_8        -0.597       -0.600      -1.17        -0.0579       0.0394 
#>  9 fam_ID_9        -0.232       -0.326       0.0627       0.968        0.283  
#> 10 fam_ID_10        0.985        1.28       -0.988       -1.19         0.733  
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype3 <dbl>, mgf_phenotype3 <dbl>,
#> #   pgm_phenotype3 <dbl>, pgf_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, f_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, pgm_phenotype3_status <lgl>,
#> #   pgf_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                3.45             -Inf
#>  2 fam_ID_2  fam_ID_2_1 o              -Inf                2.87             -Inf
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                3.42             -Inf
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                3.21             -Inf
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                3.42             -Inf
#>  6 fam_ID_6  fam_ID_6_1 o                 1.33             1.33             -Inf
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                2.43             -Inf
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                3.42             -Inf
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                2.51             -Inf
#> 10 fam_ID_10 fam_ID_10… o              -Inf                2.68             -Inf
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype1 o_phenotype1 m_phenotype1 mgm_phenotype1 mgf_phenotype1
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1       -1.07       -1.93         -1.68          -2.14          -0.681
#>  2 fam_ID_2       -0.691       1.04          0.719          0.281          0.492
#>  3 fam_ID_3        0.934       1.38          1.46           0.417          0.355
#>  4 fam_ID_4       -0.514      -0.833        -0.807          0.733          0.909
#>  5 fam_ID_5        0.489       0.605        -0.208          0.988         -0.645
#>  6 fam_ID_6       -0.977      -0.0566       -0.703          0.423         -3.25 
#>  7 fam_ID_7        0.251       0.0792        0.111          1.79          -0.783
#>  8 fam_ID_8        0.327       0.456        -0.511          1.53           0.626
#>  9 fam_ID_9        0.127       0.511        -1.74           0.627          0.419
#> 10 fam_ID_…       -0.779      -0.486         1.54           0.650          0.268
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype1 <dbl>, s2_phenotype1 <dbl>,
#> #   mhs1_phenotype1 <dbl>, mhs2_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   s2_phenotype1_status <lgl>, mhs1_phenotype1_status <lgl>,
#> #   mhs2_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype2 o_phenotype2 m_phenotype2 mgm_phenotype2 mgf_phenotype2
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1       0.809        1.49        -0.0791        0.160           0.616 
#>  2 fam_ID_2       1.09         1.54         0.586         1.34            1.87  
#>  3 fam_ID_3      -1.16        -2.36         0.302         0.00633        -0.279 
#>  4 fam_ID_4       0.700        0.0537       0.0361        0.346           0.599 
#>  5 fam_ID_5      -0.504       -0.989        1.17         -0.535          -0.387 
#>  6 fam_ID_6       0.121       -0.750        0.599         0.980           2.43  
#>  7 fam_ID_7      -0.329       -1.10         1.98         -0.495           0.0268
#>  8 fam_ID_8      -0.955       -1.82        -1.27          0.528          -0.720 
#>  9 fam_ID_9      -0.421       -1.32         0.185        -0.476          -0.390 
#> 10 fam_ID_…      -0.0204      -0.996       -1.92         -0.529           0.265 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype2 <dbl>, s2_phenotype2 <dbl>,
#> #   mhs1_phenotype2 <dbl>, mhs2_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   s2_phenotype2_status <lgl>, mhs1_phenotype2_status <lgl>,
#> #   mhs2_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype3 o_phenotype3 m_phenotype3 mgm_phenotype3 mgf_phenotype3
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1      -0.145        0.145        -1.13         -1.23           -1.84 
#>  2 fam_ID_2       1.29         2.11          0.507         0.848          -0.951
#>  3 fam_ID_3      -0.155        0.652         0.699         0.116          -0.994
#>  4 fam_ID_4       2.12         2.60          0.354         2.05           -0.808
#>  5 fam_ID_5      -0.0431      -1.03          1.06         -0.398           0.406
#>  6 fam_ID_6       0.117       -0.0904       -0.613        -0.302           0.442
#>  7 fam_ID_7       0.914        0.377         0.334         1.19            0.626
#>  8 fam_ID_8       0.455        0.999         0.298        -0.0598          0.313
#>  9 fam_ID_9      -0.285       -0.594        -1.86         -0.106           0.721
#> 10 fam_ID_…       0.215        0.271        -1.11         -0.453           1.48 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype3 <dbl>, s2_phenotype3 <dbl>,
#> #   mhs1_phenotype3 <dbl>, mhs2_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   s2_phenotype3_status <lgl>, mhs1_phenotype3_status <lgl>,
#> #   mhs2_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                2.59             1.49
#>  2 fam_ID_2  fam_ID_2_1 o              -Inf                2.63             1.54
#>  3 fam_ID_3  fam_ID_3_1 o                 1.38             1.38          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                2.99          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                3.48          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                3.45          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                2.51          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                3.10          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                3.10          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                3.31          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype1 f_phenotype1 s1_phenotype1 m_phenotype1_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        1.98        -0.861          0.112 TRUE               
#>  2 fam_ID_2        0.484        1.92           0.549 FALSE              
#>  3 fam_ID_3       -0.583        1.56           0.390 FALSE              
#>  4 fam_ID_4       -0.601        1.02          -0.321 FALSE              
#>  5 fam_ID_5        1.43         0.127         -1.07  TRUE               
#>  6 fam_ID_6       -0.617       -0.883          1.76  FALSE              
#>  7 fam_ID_7        0.662        1.02          -0.701 FALSE              
#>  8 fam_ID_8        1.14        -0.559         -0.222 FALSE              
#>  9 fam_ID_9       -0.0815       1.44          -0.652 FALSE              
#> 10 fam_ID_10       1.67        -0.0122         0.757 TRUE               
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype2 f_phenotype2 s1_phenotype2 m_phenotype2_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1       -0.942        -1.48        -0.567  FALSE              
#>  2 fam_ID_2        0.262         1.46         0.359  FALSE              
#>  3 fam_ID_3        0.0549        1.38        -2.05   FALSE              
#>  4 fam_ID_4        0.418         1.64         0.217  FALSE              
#>  5 fam_ID_5        0.239         0.267       -2.15   FALSE              
#>  6 fam_ID_6        2.17         -2.21         0.818  TRUE               
#>  7 fam_ID_7        2.13          0.450        0.417  TRUE               
#>  8 fam_ID_8        0.366        -1.19        -0.262  FALSE              
#>  9 fam_ID_9       -0.570         0.403        0.0299 FALSE              
#> 10 fam_ID_10       0.0716        0.241        0.506  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype3 f_phenotype3 s1_phenotype3 m_phenotype3_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1         1.47        -1.49         -0.807 TRUE               
#>  2 fam_ID_2         0.800        1.57          0.561 FALSE              
#>  3 fam_ID_3        -0.165        1.00         -0.167 FALSE              
#>  4 fam_ID_4        -0.986        1.12          0.739 FALSE              
#>  5 fam_ID_5        -0.137       -0.156        -1.23  FALSE              
#>  6 fam_ID_6         1.68        -1.17          1.25  TRUE               
#>  7 fam_ID_7         0.289        2.00          1.16  FALSE              
#>  8 fam_ID_8         0.155       -0.842        -1.17  FALSE              
#>  9 fam_ID_9        -1.04         1.05         -0.611 FALSE              
#> 10 fam_ID_10        0.344       -0.777        -0.213 FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 m                 1.97             1.97          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 m              -Inf                1.64          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 m              -Inf                2.47          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 m              -Inf                2.43          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 m                 1.44             1.44          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 m              -Inf                2.47             2.18
#>  7 fam_ID_7  fam_ID_7_1 m              -Inf                1.68             2.13
#>  8 fam_ID_8  fam_ID_8_1 m              -Inf                2.09          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 m              -Inf                2.22          -Inf   
#> 10 fam_ID_10 fam_ID_10… m                 1.68             1.68          -Inf   
#> # ℹ 290 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.513        -1.20  FALSE                             26
#>  2 fam_ID_2        0.510         1.34  TRUE                              78
#>  3 fam_ID_3        0.365         0.828 FALSE                             19
#>  4 fam_ID_4       -0.702        -0.570 FALSE                             28
#>  5 fam_ID_5        1.32          2.09  TRUE                              48
#>  6 fam_ID_6        0.836         1.40  TRUE                              72
#>  7 fam_ID_7        1.18          0.602 FALSE                             36
#>  8 fam_ID_8       -0.783        -0.147 FALSE                             18
#>  9 fam_ID_9       -0.300        -0.545 FALSE                             12
#> 10 fam_ID_10      -0.0825        0.288 FALSE                             28
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        -0.291       1.26   FALSE                             26
#>  2 fam_ID_2         0.218       1.89   TRUE                              53
#>  3 fam_ID_3         0.647       0.867  FALSE                             19
#>  4 fam_ID_4        -0.114      -0.450  FALSE                             28
#>  5 fam_ID_5         0.246       1.16   FALSE                             27
#>  6 fam_ID_6         0.370       0.605  FALSE                             25
#>  7 fam_ID_7        -0.131      -0.389  FALSE                             36
#>  8 fam_ID_8         0.298       0.0261 FALSE                             18
#>  9 fam_ID_9         0.107       0.651  FALSE                             12
#> 10 fam_ID_10       -0.358      -0.113  FALSE                             28
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.406        -0.261 FALSE                             26
#>  2 fam_ID_2        0.0299        0.920 FALSE                             30
#>  3 fam_ID_3        0.336         0.585 FALSE                             19
#>  4 fam_ID_4       -0.0680        0.577 FALSE                             28
#>  5 fam_ID_5       -0.620        -0.681 FALSE                             27
#>  6 fam_ID_6        0.994        -0.151 FALSE                             25
#>  7 fam_ID_7       -0.596        -0.789 FALSE                             36
#>  8 fam_ID_8        1.25          1.91  TRUE                              53
#>  9 fam_ID_9        0.336        -0.251 FALSE                             12
#> 10 fam_ID_10       0.0871        0.341 FALSE                             28
#> # ℹ 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                2.99          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 1.34             1.34             1.89
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                3.24          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                2.91          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o                 2.09             2.09          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o                 1.39             1.39          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                2.59          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                3.28          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                3.48          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                2.91          -Inf   
#> # ℹ 140 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#>