Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype1 o_phenotype1 m_phenotype1 f_phenotype1 s1_phenotype1
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1      -0.713         -1.23        -0.220       1.12          0.195 
#>  2 fam_ID_2       1.53           2.14         0.162      -0.765         0.421 
#>  3 fam_ID_3       0.102          0.445       -0.554      -0.230        -0.197 
#>  4 fam_ID_4       0.241         -0.159        2.05        0.493        -0.178 
#>  5 fam_ID_5       0.214          1.12        -1.29        0.833         0.755 
#>  6 fam_ID_6      -0.00500        1.18         0.246       0.208        -0.391 
#>  7 fam_ID_7       0.499          0.470       -1.10       -0.0524        0.0996
#>  8 fam_ID_8      -1.69          -0.709       -0.796      -1.09         -1.72  
#>  9 fam_ID_9      -1.14          -0.797       -0.550       0.952        -0.735 
#> 10 fam_ID_10      0.559          1.13        -1.17        0.483         0.932 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype1 <dbl>, mgf_phenotype1 <dbl>,
#> #   pgm_phenotype1 <dbl>, pgf_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, f_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, pgm_phenotype1_status <lgl>,
#> #   pgf_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype2 o_phenotype2 m_phenotype2 f_phenotype2 s1_phenotype2
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        0.336         0.164      -1.52         0.0821       -0.365 
#>  2 fam_ID_2       -0.0448       -1.50       -0.584        0.578         0.632 
#>  3 fam_ID_3       -0.595         0.228      -0.0197      -0.640        -1.72  
#>  4 fam_ID_4        0.566        -0.108       0.760        1.94          1.12  
#>  5 fam_ID_5        0.971         0.704       0.0830       0.735         0.503 
#>  6 fam_ID_6        0.502        -0.888      -1.08         0.345        -1.23  
#>  7 fam_ID_7       -0.408        -1.16        1.05        -2.02          0.165 
#>  8 fam_ID_8       -0.127        -0.256      -0.465       -1.33         -0.685 
#>  9 fam_ID_9       -0.309         0.340      -0.781        1.46          0.0847
#> 10 fam_ID_10      -0.342        -0.918      -0.355       -0.874         0.987 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype2 <dbl>, mgf_phenotype2 <dbl>,
#> #   pgm_phenotype2 <dbl>, pgf_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, f_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, pgm_phenotype2_status <lgl>,
#> #   pgf_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype3 o_phenotype3 m_phenotype3 f_phenotype3 s1_phenotype3
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1         0.614      -0.0254       1.26       -0.293          1.16  
#>  2 fam_ID_2        -0.757      -1.07         0.812      -0.219         -0.0431
#>  3 fam_ID_3        -1.10       -0.707        0.557       0.671         -0.505 
#>  4 fam_ID_4        -1.23       -0.174       -1.46       -0.00104       -0.213 
#>  5 fam_ID_5        -0.771      -1.23         0.0837      0.511          0.376 
#>  6 fam_ID_6         0.184      -0.480       -0.369       1.82          -0.817 
#>  7 fam_ID_7         0.153       0.436        0.581      -0.256          0.731 
#>  8 fam_ID_8        -0.432      -0.890        1.29        0.803         -2.15  
#>  9 fam_ID_9        -0.817      -0.587        1.15       -0.742          0.0814
#> 10 fam_ID_10       -0.166      -0.521       -0.972      -1.20          -0.925 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype3 <dbl>, mgf_phenotype3 <dbl>,
#> #   pgm_phenotype3 <dbl>, pgf_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, f_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, pgm_phenotype3_status <lgl>,
#> #   pgf_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                3.21             -Inf
#>  2 fam_ID_2  fam_ID_2_1 o                 2.13             2.13             -Inf
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                2.47             -Inf
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                2.51             -Inf
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                3.03             -Inf
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                3.31             -Inf
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                2.59             -Inf
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                3.31             -Inf
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                2.72             -Inf
#> 10 fam_ID_10 fam_ID_10… o              -Inf                3.38             -Inf
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype1 o_phenotype1 m_phenotype1 mgm_phenotype1 mgf_phenotype1
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1        1.20        1.03         1.11            0.241         -0.323
#>  2 fam_ID_2       -0.619      -2.02         0.283           1.35          -0.512
#>  3 fam_ID_3        0.505      -0.0603       0.959           0.930         -0.420
#>  4 fam_ID_4        0.682       1.83         1.31           -1.08           0.163
#>  5 fam_ID_5        0.754       1.26         0.776           0.895         -0.168
#>  6 fam_ID_6        0.343       0.345       -1.31            0.234          0.124
#>  7 fam_ID_7        1.47        0.991        2.03            1.72          -0.375
#>  8 fam_ID_8       -1.85       -1.72         0.306          -0.228         -0.665
#>  9 fam_ID_9        1.83        1.93        -1.31            1.72          -0.859
#> 10 fam_ID_…        0.691       1.11        -0.0815         -0.455         -0.942
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype1 <dbl>, s2_phenotype1 <dbl>,
#> #   mhs1_phenotype1 <dbl>, mhs2_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   s2_phenotype1_status <lgl>, mhs1_phenotype1_status <lgl>,
#> #   mhs2_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype2 o_phenotype2 m_phenotype2 mgm_phenotype2 mgf_phenotype2
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1        0.251       0.614        0.0355         -0.916         0.0940
#>  2 fam_ID_2        1.35        0.783        2.12            3.23          0.876 
#>  3 fam_ID_3        0.430       1.05         1.25           -0.447        -0.622 
#>  4 fam_ID_4        0.138      -0.418       -0.0632          1.06          0.0617
#>  5 fam_ID_5        0.658       2.00        -0.695          -0.973         1.46  
#>  6 fam_ID_6       -0.565      -0.0139       0.778          -0.187         0.478 
#>  7 fam_ID_7       -1.44       -1.55        -0.374          -1.03          0.139 
#>  8 fam_ID_8        0.483       0.476       -1.09            0.822        -1.11  
#>  9 fam_ID_9        0.822       1.18         0.328          -1.69          1.21  
#> 10 fam_ID_…        0.664       0.608        0.458          -0.491         0.113 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype2 <dbl>, s2_phenotype2 <dbl>,
#> #   mhs1_phenotype2 <dbl>, mhs2_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   s2_phenotype2_status <lgl>, mhs1_phenotype2_status <lgl>,
#> #   mhs2_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype3 o_phenotype3 m_phenotype3 mgm_phenotype3 mgf_phenotype3
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1      -0.568       -0.111        0.534          -0.167         -0.807
#>  2 fam_ID_2      -0.899        0.378       -1.27           -1.97          -1.38 
#>  3 fam_ID_3       0.0850      -0.0824      -0.584           1.37           0.960
#>  4 fam_ID_4       0.0579      -1.67        -0.886          -1.74           1.87 
#>  5 fam_ID_5      -0.980       -1.22        -1.74           -0.637         -2.35 
#>  6 fam_ID_6      -0.585       -0.890        0.609          -0.380         -1.34 
#>  7 fam_ID_7      -0.428       -0.116       -0.0706          0.593         -0.400
#>  8 fam_ID_8      -0.216       -0.605       -1.50           -0.113         -0.299
#>  9 fam_ID_9       0.973        0.562        1.85            1.37           0.506
#> 10 fam_ID_…       0.599        0.0764       1.13            0.414          0.646
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype3 <dbl>, s2_phenotype3 <dbl>,
#> #   mhs1_phenotype3 <dbl>, mhs2_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   s2_phenotype3_status <lgl>, mhs1_phenotype3_status <lgl>,
#> #   mhs2_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                2.51          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o              -Inf                2.63          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                3.55          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 1.85             1.85          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                3.52             2.01
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                3.31          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                2.95          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                3.31          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o                 1.93             1.93          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                3.06          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype1 f_phenotype1 s1_phenotype1 m_phenotype1_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1       0.00568       0.0931         0.397 FALSE              
#>  2 fam_ID_2       0.910         0.779          1.30  FALSE              
#>  3 fam_ID_3      -1.43         -0.551         -0.228 FALSE              
#>  4 fam_ID_4      -0.412         0.580          0.910 FALSE              
#>  5 fam_ID_5       0.623        -0.826         -0.947 FALSE              
#>  6 fam_ID_6       0.0250       -0.0195        -1.19  FALSE              
#>  7 fam_ID_7       0.764         1.73           0.391 FALSE              
#>  8 fam_ID_8       0.880         3.55           1.18  FALSE              
#>  9 fam_ID_9      -0.892        -0.864         -0.771 FALSE              
#> 10 fam_ID_10     -1.23         -0.912         -0.843 FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype2 f_phenotype2 s1_phenotype2 m_phenotype2_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        0.450       -0.581        -0.867  FALSE              
#>  2 fam_ID_2        1.52         0.905         1.08   TRUE               
#>  3 fam_ID_3       -0.630        0.693         0.354  FALSE              
#>  4 fam_ID_4       -0.654        0.287        -0.218  FALSE              
#>  5 fam_ID_5       -0.760       -2.27         -1.57   FALSE              
#>  6 fam_ID_6        0.373       -0.0324       -1.09   FALSE              
#>  7 fam_ID_7       -0.287        1.03          0.302  FALSE              
#>  8 fam_ID_8        1.57         2.18         -0.0475 TRUE               
#>  9 fam_ID_9        0.0778      -0.930         0.0532 FALSE              
#> 10 fam_ID_10      -0.996       -0.257        -0.309  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype3 f_phenotype3 s1_phenotype3 m_phenotype3_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1         0.593        1.07         0.670  FALSE              
#>  2 fam_ID_2         0.365        1.08         0.927  FALSE              
#>  3 fam_ID_3        -1.37        -0.238        0.573  FALSE              
#>  4 fam_ID_4         1.13         0.109       -0.680  FALSE              
#>  5 fam_ID_5        -2.38        -0.987       -1.27   FALSE              
#>  6 fam_ID_6        -0.293       -0.961       -0.412  FALSE              
#>  7 fam_ID_7         1.64         3.06         1.12   TRUE               
#>  8 fam_ID_8         0.869        2.21        -0.151  FALSE              
#>  9 fam_ID_9        -3.27        -1.94        -2.44   FALSE              
#> 10 fam_ID_10       -1.56         0.465       -0.0158 FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 m                 -Inf             1.68          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 m                 -Inf             1.85             1.51
#>  3 fam_ID_3  fam_ID_3_1 m                 -Inf             1.85          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 m                 -Inf             2.05          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 m                 -Inf             1.59          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 m                 -Inf             2.18          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 m                 -Inf             2.68          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 m                 -Inf             2.43             1.56
#>  9 fam_ID_9  fam_ID_9_1 m                 -Inf             2.51          -Inf   
#> 10 fam_ID_10 fam_ID_10… m                 -Inf             1.68          -Inf   
#> # ℹ 290 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1      -0.229       -0.374   FALSE                             31
#>  2 fam_ID_2       1.01         2.18    TRUE                              46
#>  3 fam_ID_3      -0.274       -0.00782 FALSE                             32
#>  4 fam_ID_4      -0.900       -1.57    FALSE                             29
#>  5 fam_ID_5       1.73         2.38    TRUE                              41
#>  6 fam_ID_6       0.197        1.13    FALSE                             17
#>  7 fam_ID_7      -0.778       -1.18    FALSE                             37
#>  8 fam_ID_8       1.20         1.01    FALSE                             33
#>  9 fam_ID_9       0.00800      1.24    FALSE                             12
#> 10 fam_ID_10      0.223        1.20    FALSE                             27
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1         1.43         1.44  TRUE                              69
#>  2 fam_ID_2         0.652        1.00  FALSE                             11
#>  3 fam_ID_3        -1.03        -2.04  FALSE                             32
#>  4 fam_ID_4        -0.973       -0.332 FALSE                             29
#>  5 fam_ID_5        -0.547        0.640 FALSE                             33
#>  6 fam_ID_6         1.09         1.07  FALSE                             17
#>  7 fam_ID_7        -0.268       -0.894 FALSE                             37
#>  8 fam_ID_8        -1.48        -1.49  FALSE                             33
#>  9 fam_ID_9        -0.304       -1.29  FALSE                             12
#> 10 fam_ID_10       -1.45         0.145 FALSE                             27
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.0613      -0.272  FALSE                             31
#>  2 fam_ID_2       -0.147        1.45   TRUE                              68
#>  3 fam_ID_3       -0.905       -2.00   FALSE                             32
#>  4 fam_ID_4        0.760       -0.168  FALSE                             29
#>  5 fam_ID_5        1.08        -0.228  FALSE                             33
#>  6 fam_ID_6        0.436        1.71   TRUE                              58
#>  7 fam_ID_7       -0.0908       0.123  FALSE                             37
#>  8 fam_ID_8       -1.07        -1.62   FALSE                             33
#>  9 fam_ID_9        0.291       -0.0902 FALSE                             12
#> 10 fam_ID_10       0.117        0.848  FALSE                             27
#> # ℹ 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                2.79             1.44
#>  2 fam_ID_2  fam_ID_2_1 o                 2.18             2.18          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                2.76          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                2.87          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o                 2.39             2.39          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                3.31          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                2.55          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                2.72          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                3.48          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                2.95          -Inf   
#> # ℹ 140 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#>