Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.882  -0.562   1.42    -0.728  0.554    2.11  -0.630  -0.754 
#>  2 fam_ID_2      -0.341  -0.559  -0.947   -0.469 -0.533   -0.782 -0.149   0.433 
#>  3 fam_ID_3       1.03    0.677   0.953    0.926  0.889   -0.420  0.964   1.07  
#>  4 fam_ID_4       0.842   0.763   0.242    0.780 -0.207   -1.76   0.206   0.147 
#>  5 fam_ID_5      -0.765   0.0127 -1.74     0.443 -0.545   -0.207  1.01   -0.549 
#>  6 fam_ID_6      -0.453  -1.12   -0.141   -1.96  -0.277   -0.631 -0.0405 -0.273 
#>  7 fam_ID_7       1.34    1.27    0.317   -0.955  0.663    1.69   0.474   0.895 
#>  8 fam_ID_8      -0.153   1.19   -0.246    0.890  0.0325  -0.393  0.294  -0.0907
#>  9 fam_ID_9      -0.0187  0.930   0.366    0.541  0.610    0.425 -0.566  -1.01  
#> 10 fam_ID_10     -0.308  -0.0970  0.0285  -1.38  -0.737    0.482 -0.318  -0.731 
#> # … with 990 more rows, 17 more variables: pgf_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   pgm_phenotype1_status <lgl>, pgf_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>,
#> #   s1_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phen…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>          <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      0.254   0.689   1.08      0.414 -1.43     1.61    1.86   0.360 
#>  2 fam_ID_2      1.17    0.427  -0.0475    1.12   0.719   -0.113  -1.08   0.163 
#>  3 fam_ID_3     -1.44   -0.349  -0.379    -1.08  -0.465   -0.141  -0.397 -1.17  
#>  4 fam_ID_4     -0.323   0.0243 -0.563    -0.127  0.902    0.682  -0.104  0.0350
#>  5 fam_ID_5     -0.0774  0.655   0.523    -0.465  1.75     0.182   1.13   0.683 
#>  6 fam_ID_6      0.0367  0.405   1.89     -0.710  1.03     0.334   1.44   0.621 
#>  7 fam_ID_7      0.877   0.612  -0.903     1.66   0.0181   0.745  -0.846  1.03  
#>  8 fam_ID_8      1.01    1.05   -0.00985   0.123  0.0746   1.23   -0.428  0.607 
#>  9 fam_ID_9     -0.257  -0.768   0.677    -1.76   0.967    1.19   -0.900 -1.47  
#> 10 fam_ID_10    -0.576  -0.173   0.474    -0.386 -1.99    -0.157  -0.663  1.00  
#> # … with 990 more rows, 17 more variables: pgf_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   pgm_phenotype2_status <lgl>, pgf_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>,
#> #   s1_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_ph…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1 fam_ID_1       0.355  0.0257 -0.0618 -1.54   -0.996  -0.524    1.05   1.20e+0
#>  2 fam_ID_2      -0.276 -1.62   -1.92   -0.706  -0.861   0.382    0.924 -5.60e-1
#>  3 fam_ID_3       1.11  -0.0517 -0.809  -0.215   1.15   -0.602    1.19  -4.63e-5
#>  4 fam_ID_4       1.01   0.790  -0.290   0.0429  0.348  -0.478   -0.672  3.63e-2
#>  5 fam_ID_5       1.55   2.06    0.693   1.71   -0.154   1.92     2.48   9.58e-1
#>  6 fam_ID_6       0.562 -1.20   -0.0924 -0.430  -0.747  -0.0711  -0.237 -1.99e-1
#>  7 fam_ID_7       1.56   1.44   -1.31    0.641  -0.712  -1.38    -0.802 -1.80e+0
#>  8 fam_ID_8      -0.206 -0.390   1.17   -0.350   1.82    0.353    1.61   3.40e-2
#>  9 fam_ID_9       0.268  0.903  -0.815   1.40   -0.0411 -0.418    0.454 -1.85e+0
#> 10 fam_ID_10     -0.851 -0.492  -0.581  -0.0235  0.222  -1.87     0.539 -1.49e+0
#> # … with 990 more rows, 17 more variables: pgf_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   pgm_phenotype3_status <lgl>, pgf_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>,
#> #   s1_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          -Inf    2.95    -Inf    2.95 -Inf       2.95
#>  2 fam_ID_2  fam_ID_2_1  o          -Inf    2.83    -Inf    2.83 -Inf       2.83
#>  3 fam_ID_3  fam_ID_3_1  o          -Inf    3.06    -Inf    3.06 -Inf       3.06
#>  4 fam_ID_4  fam_ID_4_1  o          -Inf    3.52    -Inf    3.52 -Inf       3.52
#>  5 fam_ID_5  fam_ID_5_1  o          -Inf    2.59    -Inf    2.59    2.05    2.05
#>  6 fam_ID_6  fam_ID_6_1  o          -Inf    3.06    -Inf    3.06 -Inf       3.06
#>  7 fam_ID_7  fam_ID_7_1  o          -Inf    3.10    -Inf    3.10    1.44    1.44
#>  8 fam_ID_8  fam_ID_8_1  o          -Inf    3.14    -Inf    3.14 -Inf       3.14
#>  9 fam_ID_9  fam_ID_9_1  o          -Inf    2.76    -Inf    2.76 -Inf       2.76
#> 10 fam_ID_10 fam_ID_10_1 o          -Inf    3.52    -Inf    3.52 -Inf       3.52
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phen…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.616  0.169     0.319   0.226   1.11    0.325  0.0497  -1.16 
#>  2 fam_ID_2       1.06   0.00697   0.604   0.380   1.04    0.840  0.154   -0.552
#>  3 fam_ID_3      -0.298  1.86     -0.696   0.572  -0.336  -1.15  -0.280   -0.174
#>  4 fam_ID_4       0.206 -0.463    -0.841   0.603   1.42    0.489  0.275   -1.16 
#>  5 fam_ID_5       0.521  1.71     -0.306  -0.647   0.903  -0.561 -1.19     0.622
#>  6 fam_ID_6      -0.380  1.46      0.204  -0.148   0.565  -0.479  0.484   -0.617
#>  7 fam_ID_7      -0.465  0.783    -0.366  -0.979   0.761   0.796  0.244    1.16 
#>  8 fam_ID_8      -0.277 -1.00     -1.63   -0.344  -0.951  -0.406  0.483   -2.25 
#>  9 fam_ID_9      -1.11  -1.78     -0.619   0.488  -0.584  -2.06  -1.62    -1.67 
#> 10 fam_ID_10     -1.34  -1.87     -1.53   -0.656   1.98   -2.96  -0.609   -2.64 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, s2_phenotype1_status <lgl>,
#> #   mhs1_phenotype1_status <lgl>, mhs2_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>,
#> #   mgf_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_phe…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      0.0258 -0.609   0.980    2.37  -0.785  -0.850    0.992   0.871 
#>  2 fam_ID_2      1.68    1.39    1.31     0.792 -0.0721  0.102    0.579   1.85  
#>  3 fam_ID_3     -0.0394  0.0733 -1.05    -0.419  0.593  -0.148    0.425   1.06  
#>  4 fam_ID_4      0.387   0.0767 -0.0468   1.36   0.469   0.157   -1.01   -0.270 
#>  5 fam_ID_5      1.07    0.534   0.294    0.240 -1.25   -1.19    -0.151   0.864 
#>  6 fam_ID_6     -0.841   0.931   0.0166  -1.20  -0.104  -1.32    -1.52   -0.275 
#>  7 fam_ID_7     -0.584   0.172  -0.161    0.168  1.24   -0.986   -1.68   -0.791 
#>  8 fam_ID_8     -0.564  -1.30    1.83     1.02   0.368   0.00461  1.65    0.421 
#>  9 fam_ID_9      0.136  -0.632  -0.963    0.930  0.213   0.716   -0.225  -0.145 
#> 10 fam_ID_10     1.02    1.55    1.86    -0.465 -0.0150  0.350    0.0293  0.0819
#> # … with 990 more rows, 17 more variables: mhs2_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, s2_phenotype2_status <lgl>,
#> #   mhs1_phenotype2_status <lgl>, mhs2_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>,
#> #   mgf_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phen…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      0.426   0.234    -1.71    1.90  -1.28     0.717 -0.480   -0.677
#>  2 fam_ID_2      0.892   0.00649  -0.884  -0.932 -0.768    0.184  0.0331  -2.41 
#>  3 fam_ID_3     -0.769  -0.428    -1.08    1.05  -0.755    0.226 -0.536   -1.32 
#>  4 fam_ID_4     -0.304   0.118    -0.473  -0.524  0.218    0.198 -0.942    1.24 
#>  5 fam_ID_5      0.849   2.45      0.161   0.370  1.19     1.90  -0.130   -0.869
#>  6 fam_ID_6      0.0343  1.12      0.609   1.28  -1.64    -0.794  1.21    -0.172
#>  7 fam_ID_7     -0.988  -1.09     -1.14    0.130  0.163   -0.478 -0.840   -2.33 
#>  8 fam_ID_8      0.637   0.0733    0.698  -1.22   0.370   -1.29   0.710    0.241
#>  9 fam_ID_9      0.517   0.709    -0.668   0.925 -0.658   -2.15  -1.16     0.745
#> 10 fam_ID_10    -0.201  -0.611     1.31   -0.709  0.0164   0.849  0.337   -1.20 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, s2_phenotype3_status <lgl>,
#> #   mhs1_phenotype3_status <lgl>, mhs2_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>,
#> #   mgf_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       3.24 -Inf       3.24 -Inf       3.24
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       3.48    1.39    1.39 -Inf       3.48
#>  3 fam_ID_3  fam_ID_3_1  o          1.85    1.85 -Inf       3.24 -Inf       3.24
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       3.28 -Inf       3.28 -Inf       3.28
#>  5 fam_ID_5  fam_ID_5_1  o          1.71    1.71 -Inf       2.76    2.43    2.43
#>  6 fam_ID_6  fam_ID_6_1  o          1.45    1.45 -Inf       3.45 -Inf       3.45
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       2.55 -Inf       2.55 -Inf       2.55
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       2.99 -Inf       2.99 -Inf       2.99
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       3.48 -Inf       3.48 -Inf       3.48
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       3.28    1.54    1.54 -Inf       3.28
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1        0.807 -0.182   -0.149 FALSE   FALSE   FALSE        39      39
#>  2 fam_ID_2        2.71   0.426    2.61  TRUE    FALSE   TRUE         33      56
#>  3 fam_ID_3        0.121  0.303   -0.696 FALSE   FALSE   FALSE        66      63
#>  4 fam_ID_4        0.281 -1.49     0.375 FALSE   FALSE   FALSE        65      58
#>  5 fam_ID_5        0.163 -1.52    -1.41  FALSE   FALSE   FALSE        63      57
#>  6 fam_ID_6        0.186 -0.550    1.10  FALSE   FALSE   FALSE        57      53
#>  7 fam_ID_7        0.622  0.604    1.93  FALSE   FALSE   TRUE         47      50
#>  8 fam_ID_8       -1.15   0.255    0.336 FALSE   FALSE   FALSE        52      53
#>  9 fam_ID_9        0.757  0.950    0.236 FALSE   FALSE   FALSE        48      56
#> 10 fam_ID_10      -0.257  0.0770  -0.602 FALSE   FALSE   FALSE        29      39
#> # … with 90 more rows, 1 more variable: s1_phenotype1_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype1, ²​f_phenotype1, ³​s1_phenotype1,
#> #   ⁴​m_phenotype1_status, ⁵​f_phenotype1_status, ⁶​s1_phenotype1_status,
#> #   ⁷​m_phenotype1_aoo, ⁸​f_phenotype1_aoo
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       0.0585 -0.285  -1.55   FALSE   FALSE   FALSE        39      39
#>  2 fam_ID_2      -0.557   2.09    2.44   FALSE   TRUE    TRUE         56      48
#>  3 fam_ID_3       0.930   0.347   0.0577 FALSE   FALSE   FALSE        66      63
#>  4 fam_ID_4       0.218  -1.97    1.43   FALSE   FALSE   TRUE         65      58
#>  5 fam_ID_5       1.65   -0.571  -0.674  TRUE    FALSE   FALSE        60      57
#>  6 fam_ID_6       1.07   -0.274   0.104  FALSE   FALSE   FALSE        57      53
#>  7 fam_ID_7      -1.13    1.05    0.752  FALSE   FALSE   FALSE        47      50
#>  8 fam_ID_8      -0.364   1.46    0.350  FALSE   TRUE    FALSE        52      67
#>  9 fam_ID_9       0.686   0.0262 -1.24   FALSE   FALSE   FALSE        48      56
#> 10 fam_ID_10     -0.767   1.61   -0.313  FALSE   TRUE    FALSE        29      61
#> # … with 90 more rows, 1 more variable: s1_phenotype2_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype2, ²​f_phenotype2, ³​s1_phenotype2,
#> #   ⁴​m_phenotype2_status, ⁵​f_phenotype2_status, ⁶​s1_phenotype2_status,
#> #   ⁷​m_phenotype2_aoo, ⁸​f_phenotype2_aoo
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1        1.25  -0.262   -1.92  FALSE   FALSE   FALSE        39      39
#>  2 fam_ID_2        0.164  0.701    1.26  FALSE   FALSE   FALSE        56      56
#>  3 fam_ID_3        1.03   0.641   -0.858 FALSE   FALSE   FALSE        66      63
#>  4 fam_ID_4       -2.59  -1.96    -0.431 FALSE   FALSE   FALSE        65      58
#>  5 fam_ID_5       -1.74  -1.26    -1.48  FALSE   FALSE   FALSE        63      57
#>  6 fam_ID_6        2.46   0.0928   1.21  TRUE    FALSE   FALSE        39      53
#>  7 fam_ID_7        1.62   0.819    1.76  TRUE    FALSE   TRUE         61      50
#>  8 fam_ID_8       -0.536 -0.188   -0.332 FALSE   FALSE   FALSE        52      53
#>  9 fam_ID_9       -0.523 -0.0195  -1.88  FALSE   FALSE   FALSE        48      56
#> 10 fam_ID_10       0.721 -0.644   -1.66  FALSE   FALSE   FALSE        29      39
#> # … with 90 more rows, 1 more variable: s1_phenotype3_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype3, ²​f_phenotype3, ³​s1_phenotype3,
#> #   ⁴​m_phenotype3_status, ⁵​f_phenotype3_status, ⁶​s1_phenotype3_status,
#> #   ⁷​m_phenotype3_aoo, ⁸​f_phenotype3_aoo
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m       -Inf       2.47 -Inf       2.47 -Inf       2.47
#>  2 fam_ID_2  fam_ID_2_1  m          2.72    2.72 -Inf       1.78 -Inf       1.78
#>  3 fam_ID_3  fam_ID_3_1  m       -Inf       1.49 -Inf       1.49 -Inf       1.49
#>  4 fam_ID_4  fam_ID_4_1  m       -Inf       1.51 -Inf       1.51 -Inf       1.51
#>  5 fam_ID_5  fam_ID_5_1  m       -Inf       1.56    1.64    1.64 -Inf       1.56
#>  6 fam_ID_6  fam_ID_6_1  m       -Inf       1.74 -Inf       1.74    2.47    2.47
#>  7 fam_ID_7  fam_ID_7_1  m       -Inf       2.13 -Inf       2.13    1.62    1.62
#>  8 fam_ID_8  fam_ID_8_1  m       -Inf       1.93 -Inf       1.93 -Inf       1.93
#>  9 fam_ID_9  fam_ID_9_1  m       -Inf       2.09 -Inf       2.09 -Inf       2.09
#> 10 fam_ID_10 fam_ID_10_1 m       -Inf       2.87 -Inf       2.87 -Inf       2.87
#> # … with 290 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.578       0.00563 FALSE                             12
#>  2 fam_ID_2       -0.0106     -0.701   FALSE                             13
#>  3 fam_ID_3        0.0390     -0.334   FALSE                             34
#>  4 fam_ID_4        0.275       0.447   FALSE                             12
#>  5 fam_ID_5       -0.742      -0.730   FALSE                             37
#>  6 fam_ID_6       -0.229      -1.16    FALSE                             22
#>  7 fam_ID_7       -0.0763     -0.285   FALSE                             20
#>  8 fam_ID_8        0.562      -0.337   FALSE                             31
#>  9 fam_ID_9        0.0640     -0.397   FALSE                             34
#> 10 fam_ID_10      -0.644      -1.60    FALSE                             23
#> # … with 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        -1.16        -1.36  FALSE                             12
#>  2 fam_ID_2         1.28         1.56  TRUE                              63
#>  3 fam_ID_3         0.260       -0.429 FALSE                             34
#>  4 fam_ID_4        -0.292       -0.232 FALSE                             12
#>  5 fam_ID_5        -0.153       -0.642 FALSE                             37
#>  6 fam_ID_6         0.292        0.565 FALSE                             22
#>  7 fam_ID_7        -0.677       -0.723 FALSE                             20
#>  8 fam_ID_8         0.524        1.09  FALSE                             31
#>  9 fam_ID_9        -0.696       -1.50  FALSE                             34
#> 10 fam_ID_10       -0.915        0.172 FALSE                             23
#> # … with 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.413         0.687 FALSE                             12
#>  2 fam_ID_2       -0.0269       -0.587 FALSE                             13
#>  3 fam_ID_3        0.852         1.15  FALSE                             34
#>  4 fam_ID_4        1.46          1.39  TRUE                              72
#>  5 fam_ID_5       -0.706        -0.837 FALSE                             37
#>  6 fam_ID_6       -0.514         0.860 FALSE                             22
#>  7 fam_ID_7        0.0491       -0.259 FALSE                             20
#>  8 fam_ID_8        0.161        -0.496 FALSE                             31
#>  9 fam_ID_9       -0.518        -0.306 FALSE                             34
#> 10 fam_ID_10      -1.07         -1.39  FALSE                             23
#> # … with 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          -Inf    3.48 -Inf       3.48 -Inf       3.48
#>  2 fam_ID_2  fam_ID_2_1  o          -Inf    3.45    1.56    1.56 -Inf       3.45
#>  3 fam_ID_3  fam_ID_3_1  o          -Inf    2.68 -Inf       2.68 -Inf       2.68
#>  4 fam_ID_4  fam_ID_4_1  o          -Inf    3.48 -Inf       3.48    1.39    1.39
#>  5 fam_ID_5  fam_ID_5_1  o          -Inf    2.55 -Inf       2.55 -Inf       2.55
#>  6 fam_ID_6  fam_ID_6_1  o          -Inf    3.14 -Inf       3.14 -Inf       3.14
#>  7 fam_ID_7  fam_ID_7_1  o          -Inf    3.21 -Inf       3.21 -Inf       3.21
#>  8 fam_ID_8  fam_ID_8_1  o          -Inf    2.79 -Inf       2.79 -Inf       2.79
#>  9 fam_ID_9  fam_ID_9_1  o          -Inf    2.68 -Inf       2.68 -Inf       2.68
#> 10 fam_ID_10 fam_ID_10_1 o          -Inf    3.10 -Inf       3.10 -Inf       3.10
#> # … with 140 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#>