Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype1 o_phenotype1 m_phenotype1 f_phenotype1 s1_phenotype1
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1       -1.55         -1.56       -1.52         -1.40         -1.33 
#>  2 fam_ID_2       -0.0822       -0.171       1.57          0.771        -0.556
#>  3 fam_ID_3        0.431        -0.958       0.284        -1.24         -1.40 
#>  4 fam_ID_4       -0.170         0.416      -0.133         0.441         0.195
#>  5 fam_ID_5       -0.804         0.470      -0.0246        0.303         0.898
#>  6 fam_ID_6       -2.09         -2.89        0.175        -1.26         -2.03 
#>  7 fam_ID_7        0.211         1.01       -0.414         0.490         0.267
#>  8 fam_ID_8        0.108        -0.677       1.29          1.10          1.27 
#>  9 fam_ID_9        0.339        -1.00        0.281        -1.44         -0.337
#> 10 fam_ID_10      -0.174        -1.05        0.0535        0.804        -0.199
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype1 <dbl>, mgf_phenotype1 <dbl>,
#> #   pgm_phenotype1 <dbl>, pgf_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, f_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, pgm_phenotype1_status <lgl>,
#> #   pgf_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype2 o_phenotype2 m_phenotype2 f_phenotype2 s1_phenotype2
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1         1.90        1.14          3.01         0.768        0.273 
#>  2 fam_ID_2        -0.113       0.0482        0.424       -1.62        -1.93  
#>  3 fam_ID_3        -1.23       -1.17         -1.59        -1.06        -1.68  
#>  4 fam_ID_4        -0.315      -1.25          0.691       -1.13        -1.12  
#>  5 fam_ID_5         0.355       0.733         0.921        1.64        -0.587 
#>  6 fam_ID_6         0.268      -0.0506       -0.529        1.21         0.338 
#>  7 fam_ID_7        -0.587      -0.762         0.398       -0.251        1.13  
#>  8 fam_ID_8         0.252       0.403        -0.681        0.925        0.119 
#>  9 fam_ID_9         1.16        1.63          1.56         0.768        0.0935
#> 10 fam_ID_10       -0.541      -0.637         0.864       -0.235       -0.555 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype2 <dbl>, mgf_phenotype2 <dbl>,
#> #   pgm_phenotype2 <dbl>, pgf_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, f_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, pgm_phenotype2_status <lgl>,
#> #   pgf_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype3 o_phenotype3 m_phenotype3 f_phenotype3 s1_phenotype3
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        -0.868       -0.777      0.00651       -0.601       -1.18  
#>  2 fam_ID_2        -0.368       -0.277      2.04          -0.752        0.226 
#>  3 fam_ID_3         0.139        1.05      -0.263         -1.56        -0.0433
#>  4 fam_ID_4        -0.456       -1.32      -1.21          -0.517        1.49  
#>  5 fam_ID_5         1.21         0.520      0.447          1.91         1.66  
#>  6 fam_ID_6        -0.693       -2.07      -0.257         -0.105       -0.610 
#>  7 fam_ID_7        -1.16         0.375     -0.683         -1.51         0.760 
#>  8 fam_ID_8        -1.01        -1.07       0.613         -1.41        -1.05  
#>  9 fam_ID_9         0.302        0.350     -0.780          0.403       -0.173 
#> 10 fam_ID_10        0.125        0.241     -0.179         -0.387       -0.582 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype3 <dbl>, mgf_phenotype3 <dbl>,
#> #   pgm_phenotype3 <dbl>, pgf_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, f_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, pgm_phenotype3_status <lgl>,
#> #   pgf_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o                 -Inf             2.99          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 -Inf             3.35          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o                 -Inf             2.79          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 -Inf             2.99          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o                 -Inf             3.55          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o                 -Inf             2.95          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 -Inf             3.06          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o                 -Inf             3.06          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o                 -Inf             2.72             1.64
#> 10 fam_ID_10 fam_ID_10… o                 -Inf             3.14          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype1 o_phenotype1 m_phenotype1 mgm_phenotype1 mgf_phenotype1
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1       0.252         0.912       0.664         -0.694           1.12 
#>  2 fam_ID_2      -0.616        -2.42       -0.105         -1.20            0.644
#>  3 fam_ID_3      -0.0384       -0.102       2.16          -0.624           1.66 
#>  4 fam_ID_4      -0.914         0.216       0.905          0.940          -1.22 
#>  5 fam_ID_5       0.0144       -0.273       0.153          0.586          -0.509
#>  6 fam_ID_6       0.979         0.834       1.25           1.23            2.74 
#>  7 fam_ID_7       0.328        -0.145       0.0239        -1.30           -1.11 
#>  8 fam_ID_8      -0.820        -1.39        0.401         -0.0952         -0.540
#>  9 fam_ID_9      -0.174        -0.620       0.662          0.373          -1.79 
#> 10 fam_ID_…      -0.414        -1.48       -0.401         -1.42           -1.10 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype1 <dbl>, s2_phenotype1 <dbl>,
#> #   mhs1_phenotype1 <dbl>, mhs2_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   s2_phenotype1_status <lgl>, mhs1_phenotype1_status <lgl>,
#> #   mhs2_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype2 o_phenotype2 m_phenotype2 mgm_phenotype2 mgf_phenotype2
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1        0.906        2.20         3.13           3.03          -0.507
#>  2 fam_ID_2       -0.661       -0.317        0.110          0.184         -1.95 
#>  3 fam_ID_3        0.485       -0.300        0.340          1.46           0.994
#>  4 fam_ID_4        0.576        1.79         0.373          0.660         -0.869
#>  5 fam_ID_5        0.563        1.46         1.02          -0.388          1.36 
#>  6 fam_ID_6       -1.04        -1.94        -0.680          0.713         -0.250
#>  7 fam_ID_7       -0.193        0.472       -0.879         -0.914         -1.23 
#>  8 fam_ID_8       -0.530       -0.686        0.497         -0.604          0.790
#>  9 fam_ID_9        0.283       -0.317        0.588          0.286         -0.258
#> 10 fam_ID_…       -0.693       -1.87        -0.497         -1.87          -0.251
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype2 <dbl>, s2_phenotype2 <dbl>,
#> #   mhs1_phenotype2 <dbl>, mhs2_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   s2_phenotype2_status <lgl>, mhs1_phenotype2_status <lgl>,
#> #   mhs2_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype3 o_phenotype3 m_phenotype3 mgm_phenotype3 mgf_phenotype3
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1     -1.74          -1.65        0.694          -0.150         -0.244
#>  2 fam_ID_2      1.16           0.790       1.32            0.142         -0.290
#>  3 fam_ID_3     -0.0159         0.676      -0.0144         -0.903          1.02 
#>  4 fam_ID_4     -0.522          0.537       0.0554          0.315          0.898
#>  5 fam_ID_5      0.506          0.964      -0.391           0.233         -0.925
#>  6 fam_ID_6      0.00829        0.270      -0.116           0.362         -0.118
#>  7 fam_ID_7     -0.574         -0.273      -0.0554         -0.168          0.950
#>  8 fam_ID_8      1.77           2.19        1.21            2.21           2.56 
#>  9 fam_ID_9     -0.839         -2.03       -0.194          -0.419          0.293
#> 10 fam_ID_…     -0.0366         0.628      -0.773          -0.478         -1.13 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype3 <dbl>, s2_phenotype3 <dbl>,
#> #   mhs1_phenotype3 <dbl>, mhs2_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   s2_phenotype3_status <lgl>, mhs1_phenotype3_status <lgl>,
#> #   mhs2_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o                 -Inf             2.55             2.18
#>  2 fam_ID_2  fam_ID_2_1 o                 -Inf             3.03          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o                 -Inf             2.43          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 -Inf             2.55             1.78
#>  5 fam_ID_5  fam_ID_5_1 o                 -Inf             3.42             1.47
#>  6 fam_ID_6  fam_ID_6_1 o                 -Inf             3.38          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 -Inf             2.72          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o                 -Inf             3.55          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o                 -Inf             3.31          -Inf   
#> 10 fam_ID_10 fam_ID_10… o                 -Inf             3.55          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype1 f_phenotype1 s1_phenotype1 m_phenotype1_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1       -0.515         0.471       -0.0547 FALSE              
#>  2 fam_ID_2       -0.0101        1.95        -1.11   FALSE              
#>  3 fam_ID_3        0.762         1.16        -0.212  FALSE              
#>  4 fam_ID_4        0.413        -1.29        -0.0593 FALSE              
#>  5 fam_ID_5       -0.232         0.909       -0.883  FALSE              
#>  6 fam_ID_6       -0.525        -0.985       -0.252  FALSE              
#>  7 fam_ID_7        0.596        -1.17        -0.114  FALSE              
#>  8 fam_ID_8        0.860         1.64        -0.377  FALSE              
#>  9 fam_ID_9        0.726         0.364       -1.01   FALSE              
#> 10 fam_ID_10       1.06         -0.662       -0.568  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype2 f_phenotype2 s1_phenotype2 m_phenotype2_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        1.00          1.31         -0.628 FALSE              
#>  2 fam_ID_2       -2.33          0.433        -1.52  FALSE              
#>  3 fam_ID_3        0.148         0.310        -0.718 FALSE              
#>  4 fam_ID_4        0.556        -1.07          0.400 FALSE              
#>  5 fam_ID_5       -0.0127       -0.185        -0.869 FALSE              
#>  6 fam_ID_6       -0.0569        0.128        -0.287 FALSE              
#>  7 fam_ID_7        0.0438       -0.390         1.56  FALSE              
#>  8 fam_ID_8       -0.0791        1.50         -0.433 FALSE              
#>  9 fam_ID_9       -0.801         0.144        -1.26  FALSE              
#> 10 fam_ID_10      -0.342        -0.270        -1.79  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype3 f_phenotype3 s1_phenotype3 m_phenotype3_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1       -1.47          1.39         -0.672 FALSE              
#>  2 fam_ID_2       -0.249         1.14         -1.57  FALSE              
#>  3 fam_ID_3       -0.328         0.678         0.201 FALSE              
#>  4 fam_ID_4       -0.413        -2.41         -0.232 FALSE              
#>  5 fam_ID_5       -0.443        -0.289        -1.86  FALSE              
#>  6 fam_ID_6        1.22         -0.619         0.373 FALSE              
#>  7 fam_ID_7        0.0886       -0.551         0.651 FALSE              
#>  8 fam_ID_8        1.46          0.970         0.109 TRUE               
#>  9 fam_ID_9        0.397         0.728         0.504 FALSE              
#> 10 fam_ID_10      -0.454         1.77          0.744 FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 m                 -Inf             1.93             -Inf
#>  2 fam_ID_2  fam_ID_2_1 m                 -Inf             1.97             -Inf
#>  3 fam_ID_3  fam_ID_3_1 m                 -Inf             1.81             -Inf
#>  4 fam_ID_4  fam_ID_4_1 m                 -Inf             2.30             -Inf
#>  5 fam_ID_5  fam_ID_5_1 m                 -Inf             1.78             -Inf
#>  6 fam_ID_6  fam_ID_6_1 m                 -Inf             2.34             -Inf
#>  7 fam_ID_7  fam_ID_7_1 m                 -Inf             1.64             -Inf
#>  8 fam_ID_8  fam_ID_8_1 m                 -Inf             1.89             -Inf
#>  9 fam_ID_9  fam_ID_9_1 m                 -Inf             1.49             -Inf
#> 10 fam_ID_10 fam_ID_10… m                 -Inf             1.68             -Inf
#> # ℹ 290 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.186        -0.674 FALSE                             26
#>  2 fam_ID_2       -0.878        -1.59  FALSE                             31
#>  3 fam_ID_3       -0.121         0.191 FALSE                             21
#>  4 fam_ID_4       -0.404        -0.269 FALSE                             17
#>  5 fam_ID_5        0.596         1.07  FALSE                             18
#>  6 fam_ID_6        0.395         0.471 FALSE                             27
#>  7 fam_ID_7        0.344         1.76  TRUE                              56
#>  8 fam_ID_8       -0.523        -0.763 FALSE                             11
#>  9 fam_ID_9        0.0317       -1.72  FALSE                             22
#> 10 fam_ID_10      -0.967        -1.44  FALSE                             31
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.227       -0.123  FALSE                             26
#>  2 fam_ID_2       -0.329        0.866  FALSE                             31
#>  3 fam_ID_3       -0.798       -1.70   FALSE                             21
#>  4 fam_ID_4        0.139        0.0180 FALSE                             17
#>  5 fam_ID_5        0.394       -0.617  FALSE                             18
#>  6 fam_ID_6       -0.262       -0.584  FALSE                             27
#>  7 fam_ID_7       -0.103       -0.701  FALSE                             13
#>  8 fam_ID_8       -0.0762      -0.719  FALSE                             11
#>  9 fam_ID_9       -0.286        0.121  FALSE                             22
#> 10 fam_ID_10       0.490        1.52   TRUE                              65
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.0808      -1.19   FALSE                             26
#>  2 fam_ID_2       -0.420       -0.683  FALSE                             31
#>  3 fam_ID_3       -0.565       -0.819  FALSE                             21
#>  4 fam_ID_4       -1.19        -1.14   FALSE                             17
#>  5 fam_ID_5       -0.0723      -0.215  FALSE                             18
#>  6 fam_ID_6       -0.345        0.193  FALSE                             27
#>  7 fam_ID_7       -0.267       -0.138  FALSE                             13
#>  8 fam_ID_8        0.882        0.681  FALSE                             11
#>  9 fam_ID_9        1.42        -0.0288 FALSE                             22
#> 10 fam_ID_10      -0.576       -0.245  FALSE                             31
#> # ℹ 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o              -Inf                2.99          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o              -Inf                2.79          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                3.17          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                3.31          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                3.28          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                2.95          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 1.78             1.78          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                3.52          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                3.14          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                2.79             1.51
#> # ℹ 140 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#>