Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_ph…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1 fam_ID_1      1.97     2.36   0.351    1.97    1.06    0.343 -0.214  -1.15   
#>  2 fam_ID_2      0.0845  -0.609  0.430    0.778   0.592   0.182  0.875  -0.00415
#>  3 fam_ID_3     -1.01    -1.12  -0.297   -0.123   0.273   0.194 -1.33    0.141  
#>  4 fam_ID_4     -0.409   -1.51  -0.150    1.12    0.982  -1.01  -1.37   -0.324  
#>  5 fam_ID_5     -0.311   -1.54  -0.128    0.832  -0.667   1.61  -1.18   -1.68   
#>  6 fam_ID_6     -1.43    -1.12  -1.88    -0.105  -0.110  -0.515  0.307  -0.0836 
#>  7 fam_ID_7     -0.294   -1.21  -0.0535   1.62    0.174   1.30  -0.0336  1.21   
#>  8 fam_ID_8     -0.904   -0.873 -0.303   -0.329  -2.41   -0.696 -2.23   -0.733  
#>  9 fam_ID_9      0.675    0.105  1.56     0.484  -0.504  -0.138 -0.437  -1.08   
#> 10 fam_ID_10     0.871    0.898 -0.379    0.863   1.49    0.273  1.70   -0.464  
#> # … with 990 more rows, 17 more variables: pgf_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   pgm_phenotype1_status <lgl>, pgf_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>,
#> #   s1_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1        0.555  0.658   -0.554  1.12    0.0416  -1.39   1.27     1.67 
#>  2 fam_ID_2        0.130 -0.600   -0.774  1.39    0.883    0.346 -0.744    1.33 
#>  3 fam_ID_3        0.768  0.706   -0.761  0.747   0.616   -0.743  0.599    0.270
#>  4 fam_ID_4       -1.26  -2.19    -1.47   0.0134 -2.43    -0.952 -2.02    -0.698
#>  5 fam_ID_5        0.155  1.14    -0.451  1.76    0.241   -0.558 -0.485    1.48 
#>  6 fam_ID_6        0.690  0.675    2.33   0.529   1.56     0.836  0.509    0.253
#>  7 fam_ID_7       -1.23  -0.514    0.111 -1.72    0.637   -0.127 -1.15    -0.611
#>  8 fam_ID_8        0.291 -0.0255  -0.173  1.66   -0.0942  -0.530 -0.0670   1.14 
#>  9 fam_ID_9       -0.389 -1.12    -0.600 -0.273  -2.31     0.925 -1.92    -0.920
#> 10 fam_ID_10      -0.546 -0.0284  -0.434 -1.43    0.188   -0.788 -0.0637  -0.961
#> # … with 990 more rows, 17 more variables: pgf_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   pgm_phenotype2_status <lgl>, pgf_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>,
#> #   s1_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_phe…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1     -1.44   -2.18   -0.370   -0.614  0.00255 -1.30   -0.997  -0.307 
#>  2 fam_ID_2      0.748  -0.0592  0.711    0.600  0.827   -1.78    0.219   0.257 
#>  3 fam_ID_3     -0.0452  0.953   0.343   -0.871 -1.77    -1.28   -0.849  -1.45  
#>  4 fam_ID_4      0.168   0.169  -0.619    0.592 -0.588   -0.546   0.0559  0.972 
#>  5 fam_ID_5      0.0521 -0.327   0.388   -0.618 -0.666    0.0563  0.117  -0.0603
#>  6 fam_ID_6     -0.121  -0.841  -1.90     0.453 -0.831    0.871  -0.226   2.04  
#>  7 fam_ID_7     -0.335  -0.890   0.249    0.323  0.199   -0.369   0.481   0.602 
#>  8 fam_ID_8     -0.516   0.371  -0.0394  -0.973 -0.182   -1.67    0.100  -1.23  
#>  9 fam_ID_9     -1.21   -1.79   -1.92     0.467  1.85    -0.205  -1.39   -0.450 
#> 10 fam_ID_10     0.493   2.29    1.46     0.322  1.02     1.43   -0.440   0.401 
#> # … with 990 more rows, 17 more variables: pgf_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   pgm_phenotype3_status <lgl>, pgf_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>,
#> #   s1_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          2.34    2.34    -Inf    2.59 -Inf       2.59
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.83    -Inf    2.83 -Inf       2.83
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.03    -Inf    3.03 -Inf       3.03
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       3.52    -Inf    3.52 -Inf       3.52
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       2.76    -Inf    2.76 -Inf       2.76
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       2.43    -Inf    2.43 -Inf       2.43
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.52    -Inf    3.52 -Inf       3.52
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.45    -Inf    3.45 -Inf       3.45
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.47    -Inf    2.47 -Inf       2.47
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.47    -Inf    2.47    2.30    2.30
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_phe…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.321  1.55   -0.637  -0.314  -0.337   0.00528  1.61   -0.418 
#>  2 fam_ID_2      -0.178 -0.708  -1.55    0.277  -0.844  -0.515   -0.151  -0.707 
#>  3 fam_ID_3      -0.305  0.0748 -0.127  -1.85   -1.70   -0.262    0.958  -0.0194
#>  4 fam_ID_4       0.198  1.07    1.02    1.86    0.399   0.768    0.774   1.03  
#>  5 fam_ID_5      -0.788 -0.890  -2.58   -1.43   -1.49   -1.91    -0.0847 -0.0873
#>  6 fam_ID_6      -0.460  0.0831 -0.606  -1.55   -0.683   1.33    -0.751   0.107 
#>  7 fam_ID_7       0.991 -0.263   0.762  -0.0974 -0.725  -0.0795   1.39    0.864 
#>  8 fam_ID_8       0.599  0.574   0.383   0.465  -0.682   0.434   -1.08   -0.862 
#>  9 fam_ID_9       0.924  0.975   0.285   0.146   0.180  -0.566    0.119   0.651 
#> 10 fam_ID_10     -0.742  0.121  -0.0516 -0.145   0.0423  0.138   -0.367  -0.294 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, s2_phenotype1_status <lgl>,
#> #   mhs1_phenotype1_status <lgl>, mhs2_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>,
#> #   mgf_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.700  -0.518   0.198    1.87   -0.815  0.0908   1.24  -0.0470
#>  2 fam_ID_2       0.0730  0.201   1.24    -0.326  -0.841 -1.43     2.36   0.527 
#>  3 fam_ID_3       0.155  -0.158   0.152   -1.05   -0.453 -1.64    -1.42  -1.12  
#>  4 fam_ID_4       0.190  -0.272  -1.07    -1.22   -0.214 -0.868   -0.485 -0.765 
#>  5 fam_ID_5       0.133  -0.0663 -0.268   -0.100  -1.10  -0.163    0.285 -0.0687
#>  6 fam_ID_6       0.162   1.26   -0.178   -0.215  -1.27  -1.59    -0.599  1.31  
#>  7 fam_ID_7       1.45    0.218   0.224    0.405   1.92  -0.588    0.720 -0.0284
#>  8 fam_ID_8       1.13    2.44    1.27    -0.784  -0.919  1.78    -0.902 -0.185 
#>  9 fam_ID_9       0.572   1.32   -0.348   -0.228   0.479 -0.158   -0.547 -0.417 
#> 10 fam_ID_10     -0.0533 -0.129   0.0207  -0.737  -0.398  1.33     0.967 -1.15  
#> # … with 990 more rows, 17 more variables: mhs2_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, s2_phenotype2_status <lgl>,
#> #   mhs1_phenotype2_status <lgl>, mhs2_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>,
#> #   mgf_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phen…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1     -0.417  -0.0154  0.515    0.875    1.23   -0.523  0.334   0.0229
#>  2 fam_ID_2     -0.405   0.405   0.433    0.0600  -2.10   -2.25  -0.487  -2.04  
#>  3 fam_ID_3      1.91    3.33    0.714   -0.275    1.58    0.344 -0.965   0.281 
#>  4 fam_ID_4     -1.06    0.0876 -1.76    -2.14    -1.88   -0.823  0.387   1.07  
#>  5 fam_ID_5      0.0916  0.995  -0.418    0.995   -0.236   0.879 -0.618  -1.14  
#>  6 fam_ID_6      0.791   1.55   -0.00180 -1.44    -0.610  -1.16   1.39   -0.704 
#>  7 fam_ID_7     -0.343   0.490  -1.42     1.77     0.161  -0.233  0.0712 -1.84  
#>  8 fam_ID_8     -1.50   -1.93    0.282   -0.659   -1.24   -0.410 -2.53    1.28  
#>  9 fam_ID_9      0.697   0.261  -0.718   -0.934   -1.82    0.228  0.529   0.949 
#> 10 fam_ID_10    -1.20   -1.49   -1.73    -0.222   -0.202  -1.15   1.65   -0.0301
#> # … with 990 more rows, 17 more variables: mhs2_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, s2_phenotype3_status <lgl>,
#> #   mhs1_phenotype3_status <lgl>, mhs2_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>,
#> #   mgf_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          1.54    1.54 -Inf       3.24 -Inf       3.24
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.43 -Inf       2.43 -Inf       2.43
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.17 -Inf       3.17    3.35    3.35
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       2.87 -Inf       2.87 -Inf       2.87
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       3.21 -Inf       3.21 -Inf       3.21
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       2.72 -Inf       2.72    1.56    1.56
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.48 -Inf       3.48 -Inf       3.48
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       2.43    2.43    2.43 -Inf       2.43
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       3.10    1.32    1.32 -Inf       3.10
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.59 -Inf       2.59 -Inf       2.59
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1      -0.827   -0.699  -1.34  FALSE   FALSE   FALSE        45      43
#>  2 fam_ID_2      -0.334    1.01   -0.119 FALSE   FALSE   FALSE        67      60
#>  3 fam_ID_3       1.29    -0.963   0.805 TRUE    FALSE   FALSE        90      35
#>  4 fam_ID_4      -0.0442   1.23   -1.08  FALSE   FALSE   FALSE        62      60
#>  5 fam_ID_5       0.293   -0.335  -0.631 FALSE   FALSE   FALSE        66      58
#>  6 fam_ID_6       1.73     0.453   0.861 TRUE    FALSE   FALSE        57      55
#>  7 fam_ID_7      -0.428   -0.897  -1.28  FALSE   FALSE   FALSE        35      40
#>  8 fam_ID_8      -0.731   -0.300  -2.04  FALSE   FALSE   FALSE        51      61
#>  9 fam_ID_9       0.0274  -2.26   -2.00  FALSE   FALSE   FALSE        41      33
#> 10 fam_ID_10      0.509   -2.40    0.739 FALSE   FALSE   FALSE        42      40
#> # … with 90 more rows, 1 more variable: s1_phenotype1_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype1, ²​f_phenotype1, ³​s1_phenotype1,
#> #   ⁴​m_phenotype1_status, ⁵​f_phenotype1_status, ⁶​s1_phenotype1_status,
#> #   ⁷​m_phenotype1_aoo, ⁸​f_phenotype1_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       1.57   -1.16     0.498 TRUE    FALSE   FALSE        63      43
#>  2 fam_ID_2       0.0587  0.299    0.520 FALSE   FALSE   FALSE        67      60
#>  3 fam_ID_3       1.37   -0.547    0.901 TRUE    FALSE   FALSE        74      35
#>  4 fam_ID_4       0.663   1.83     0.451 FALSE   TRUE    FALSE        62      55
#>  5 fam_ID_5      -0.379   0.124    0.875 FALSE   FALSE   FALSE        66      58
#>  6 fam_ID_6      -0.309  -0.830   -0.319 FALSE   FALSE   FALSE        44      55
#>  7 fam_ID_7      -0.286  -1.72    -0.527 FALSE   FALSE   FALSE        35      40
#>  8 fam_ID_8       0.535   0.0287  -1.56  FALSE   FALSE   FALSE        51      61
#>  9 fam_ID_9      -0.253  -1.90    -0.822 FALSE   FALSE   FALSE        41      33
#> 10 fam_ID_10      0.736  -1.55     0.579 FALSE   FALSE   FALSE        42      40
#> # … with 90 more rows, 1 more variable: s1_phenotype2_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype2, ²​f_phenotype2, ³​s1_phenotype2,
#> #   ⁴​m_phenotype2_status, ⁵​f_phenotype2_status, ⁶​s1_phenotype2_status,
#> #   ⁷​m_phenotype2_aoo, ⁸​f_phenotype2_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       -0.133  -0.553 -0.144  FALSE   FALSE   FALSE        45      43
#>  2 fam_ID_2       -0.719   0.644 -0.681  FALSE   FALSE   FALSE        67      60
#>  3 fam_ID_3        1.08   -2.22   0.0371 FALSE   FALSE   FALSE        42      35
#>  4 fam_ID_4       -0.167   1.61  -0.116  FALSE   TRUE    FALSE        62      61
#>  5 fam_ID_5        0.616  -0.483  0.156  FALSE   FALSE   FALSE        66      58
#>  6 fam_ID_6        0.690  -0.311  0.587  FALSE   FALSE   FALSE        44      55
#>  7 fam_ID_7        0.104  -0.523 -1.47   FALSE   FALSE   FALSE        35      40
#>  8 fam_ID_8       -0.406  -0.853 -3.64   FALSE   FALSE   FALSE        51      61
#>  9 fam_ID_9        0.627  -1.35  -0.784  FALSE   FALSE   FALSE        41      33
#> 10 fam_ID_10      -1.08   -0.131  1.02   FALSE   FALSE   FALSE        42      40
#> # … with 90 more rows, 1 more variable: s1_phenotype3_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype3, ²​f_phenotype3, ³​s1_phenotype3,
#> #   ⁴​m_phenotype3_status, ⁵​f_phenotype3_status, ⁶​s1_phenotype3_status,
#> #   ⁷​m_phenotype3_aoo, ⁸​f_phenotype3_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m       -Inf       2.22    1.56    1.56    -Inf    2.22
#>  2 fam_ID_2  fam_ID_2_1  m       -Inf       1.47 -Inf       1.47    -Inf    1.47
#>  3 fam_ID_3  fam_ID_3_1  m          1.29    1.29    1.37    1.37    -Inf    2.34
#>  4 fam_ID_4  fam_ID_4_1  m       -Inf       1.59 -Inf       1.59    -Inf    1.59
#>  5 fam_ID_5  fam_ID_5_1  m       -Inf       1.49 -Inf       1.49    -Inf    1.49
#>  6 fam_ID_6  fam_ID_6_1  m          1.74    1.74 -Inf       2.26    -Inf    2.26
#>  7 fam_ID_7  fam_ID_7_1  m       -Inf       2.63 -Inf       2.63    -Inf    2.63
#>  8 fam_ID_8  fam_ID_8_1  m       -Inf       1.97 -Inf       1.97    -Inf    1.97
#>  9 fam_ID_9  fam_ID_9_1  m       -Inf       2.39 -Inf       2.39    -Inf    2.39
#> 10 fam_ID_10 fam_ID_10_1 m       -Inf       2.34 -Inf       2.34    -Inf    2.34
#> # … with 290 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        1.31          2.56  TRUE                              37
#>  2 fam_ID_2        0.595         1.40  TRUE                              72
#>  3 fam_ID_3       -0.629        -0.297 FALSE                             26
#>  4 fam_ID_4        0.833         1.00  FALSE                             25
#>  5 fam_ID_5       -0.0599       -0.839 FALSE                             12
#>  6 fam_ID_6        1.22          1.25  FALSE                             19
#>  7 fam_ID_7       -0.541        -0.291 FALSE                             21
#>  8 fam_ID_8       -0.163        -0.126 FALSE                             14
#>  9 fam_ID_9        0.516         0.477 FALSE                             37
#> 10 fam_ID_10      -0.428        -0.169 FALSE                             19
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.238        0.239  FALSE                             26
#>  2 fam_ID_2       -0.669        0.0371 FALSE                             23
#>  3 fam_ID_3       -0.594       -1.43   FALSE                             26
#>  4 fam_ID_4       -1.12         0.0691 FALSE                             25
#>  5 fam_ID_5       -0.427       -1.04   FALSE                             12
#>  6 fam_ID_6        0.847        0.359  FALSE                             19
#>  7 fam_ID_7        1.10         0.312  FALSE                             21
#>  8 fam_ID_8       -0.0678       0.723  FALSE                             14
#>  9 fam_ID_9        1.03         1.88   TRUE                              53
#> 10 fam_ID_10       1.31         1.21   FALSE                             19
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        -0.504        0.997 FALSE                             26
#>  2 fam_ID_2        -0.866       -0.775 FALSE                             23
#>  3 fam_ID_3         0.418        1.05  FALSE                             26
#>  4 fam_ID_4         0.411        0.258 FALSE                             25
#>  5 fam_ID_5        -0.456       -0.274 FALSE                             12
#>  6 fam_ID_6        -1.20        -1.16  FALSE                             19
#>  7 fam_ID_7        -0.126        1.73  TRUE                              57
#>  8 fam_ID_8        -0.883       -1.04  FALSE                             14
#>  9 fam_ID_9        -0.250       -1.26  FALSE                             37
#> 10 fam_ID_10       -0.111       -0.266 FALSE                             19
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          2.55    2.55 -Inf       2.99 -Inf       2.99
#>  2 fam_ID_2  fam_ID_2_1  o          1.39    1.39 -Inf       3.10 -Inf       3.10
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       2.99 -Inf       2.99 -Inf       2.99
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       3.03 -Inf       3.03 -Inf       3.03
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       3.48 -Inf       3.48 -Inf       3.48
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       3.24 -Inf       3.24 -Inf       3.24
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.17 -Inf       3.17    1.74    1.74
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.42 -Inf       3.42 -Inf       3.42
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.55    1.89    1.89 -Inf       2.55
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       3.24 -Inf       3.24 -Inf       3.24
#> # … with 140 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#>