Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1        1.08    2.81   1.18   -0.260  -0.0719  1.43     0.488   0.113
#>  2 fam_ID_2        0.383   1.89   1.07   -0.0190 -0.0560  0.567    0.271  -0.472
#>  3 fam_ID_3        0.965   0.918  0.637  -1.99    0.695   1.81     1.91   -0.478
#>  4 fam_ID_4       -1.06   -1.12  -0.410  -0.748  -1.81   -1.29     1.02   -2.79 
#>  5 fam_ID_5        0.822   1.87  -1.04    1.51   -0.613   0.515   -0.634   0.301
#>  6 fam_ID_6        1.56    1.34   1.43    1.21    0.879   0.405    0.991  -1.45 
#>  7 fam_ID_7        0.346   0.411 -0.566  -0.761   0.217   0.284    0.472   0.321
#>  8 fam_ID_8       -0.206   0.640  0.0544 -0.328   1.14    0.0887  -1.21    0.418
#>  9 fam_ID_9       -1.03   -1.33   0.337   0.381  -0.300   1.30     0.348   0.282
#> 10 fam_ID_10      -0.773  -0.205 -0.866  -1.74    0.445  -0.660   -0.880   0.203
#> # … with 990 more rows, 17 more variables: pgf_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   pgm_phenotype1_status <lgl>, pgf_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>,
#> #   s1_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phen…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>          <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1     0.108    0.938   -0.794   0.476   0.825    1.46    0.875  1.51  
#>  2 fam_ID_2    -1.30    -0.507   -0.746  -1.34   -0.0931  -1.18   -1.31   0.523 
#>  3 fam_ID_3    -0.0997  -1.88    -1.04    0.585   0.619   -0.282   1.27   0.782 
#>  4 fam_ID_4     0.355    1.08    -1.47   -0.988  -0.988    0.992   0.813  0.0657
#>  5 fam_ID_5     0.690    0.380   -1.65    1.64    1.21     0.214   0.243  0.708 
#>  6 fam_ID_6     0.556    0.537   -0.374   3.60    0.770    1.04    0.876  1.15  
#>  7 fam_ID_7     0.00563  0.426   -0.514  -0.158  -1.68    -0.288   1.66  -0.936 
#>  8 fam_ID_8     1.05     0.340    0.379  -0.0809 -0.424    0.930   0.923  0.222 
#>  9 fam_ID_9     0.0354  -0.00492 -0.238   0.155  -1.65    -0.730  -0.481 -0.0152
#> 10 fam_ID_10    0.471    0.746   -0.0927  0.528   0.266   -0.661  -0.740 -1.53  
#> # … with 990 more rows, 17 more variables: pgf_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   pgm_phenotype2_status <lgl>, pgf_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>,
#> #   s1_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_phe…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      -1.50  -1.39    0.0931  -1.87  -0.00908   1.15    1.62  -0.774 
#>  2 fam_ID_2       0.250  1.29    0.517   -0.186 -0.603    -0.815   1.01   0.278 
#>  3 fam_ID_3      -1.30  -1.87   -1.91    -1.43  -0.394    -0.313   0.217  0.385 
#>  4 fam_ID_4       0.554  0.420  -0.312   -0.773  0.863     1.14   -1.04   0.554 
#>  5 fam_ID_5       0.107 -0.0469 -0.229    1.32   0.477    -0.226  -0.885  1.73  
#>  6 fam_ID_6      -0.773 -0.937  -0.515    1.03   0.171    -0.824   0.167 -0.0551
#>  7 fam_ID_7       0.137 -0.103   0.520    0.273 -0.349    -0.403   0.325 -0.0780
#>  8 fam_ID_8       0.161  0.0436  0.0308   0.762  1.01     -0.355  -1.00   0.128 
#>  9 fam_ID_9       0.398  1.32    0.493    1.06   1.64      0.624   0.102 -0.0367
#> 10 fam_ID_10     -0.789 -1.19   -0.316   -0.250  0.792     0.353   0.531 -1.21  
#> # … with 990 more rows, 17 more variables: pgf_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   pgm_phenotype3_status <lgl>, pgf_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>,
#> #   s1_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          2.79    2.79    -Inf    3.21 -Inf       3.21
#>  2 fam_ID_2  fam_ID_2_1  o          1.89    1.89    -Inf    2.55    1.29    1.29
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       2.68    -Inf    2.68 -Inf       2.68
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       2.91    -Inf    2.91 -Inf       2.91
#>  5 fam_ID_5  fam_ID_5_1  o          1.85    1.85    -Inf    2.72 -Inf       2.72
#>  6 fam_ID_6  fam_ID_6_1  o          1.34    1.34    -Inf    2.68 -Inf       2.68
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       2.63    -Inf    2.63 -Inf       2.63
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       3.17    -Inf    3.17 -Inf       3.17
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       3.14    -Inf    3.14    1.33    1.33
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.43    -Inf    2.43 -Inf       2.43
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_phe…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1     -0.0587  -0.548   0.271 -0.538   -0.295  0.351     1.15  -1.83  
#>  2 fam_ID_2     -0.173   -0.698   0.495  0.0302   0.593 -1.27      1.54   0.377 
#>  3 fam_ID_3      0.165    1.25    0.514  0.918    0.846  0.0651    1.07   0.415 
#>  4 fam_ID_4      1.02     0.542   0.803  3.68     0.839  0.451    -0.359 -0.668 
#>  5 fam_ID_5      1.03     1.79    0.374  2.13     0.712  1.39      0.585 -0.568 
#>  6 fam_ID_6     -0.0507   0.523   0.843 -0.0321  -0.412 -0.832     0.589  0.0232
#>  7 fam_ID_7     -0.700    0.475   0.137 -0.284    1.34  -0.0605    0.794  0.597 
#>  8 fam_ID_8      0.225    1.20    1.76  -1.53    -1.78   1.02     -1.86  -1.49  
#>  9 fam_ID_9      0.588    0.647  -0.838 -0.769    1.21   0.600     0.627 -1.05  
#> 10 fam_ID_10     1.69     1.23    0.989  0.0975  -0.154 -0.00235   1.54   0.165 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, s2_phenotype1_status <lgl>,
#> #   mhs1_phenotype1_status <lgl>, mhs2_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>,
#> #   mgf_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      0.00446 -0.0349  -1.89    1.61   -0.520  -1.48  -0.386   -0.505
#>  2 fam_ID_2      0.170    0.778    1.00   -1.33    0.535  -0.273  0.189    0.537
#>  3 fam_ID_3     -0.531   -0.334   -0.640   1.59    0.520   1.02  -0.297   -0.173
#>  4 fam_ID_4      0.324    0.189    0.652   0.691   0.396   0.321  0.144    0.232
#>  5 fam_ID_5     -0.185    0.364   -0.276   1.36   -1.02   -0.497 -0.309    1.64 
#>  6 fam_ID_6      0.688    1.63     0.737  -2.01    0.889   0.886  1.10    -2.49 
#>  7 fam_ID_7     -0.195   -1.39     1.33    0.235  -1.96   -1.01   0.0529  -1.75 
#>  8 fam_ID_8     -1.02    -0.871    0.657   0.321  -0.851  -1.72  -0.866    2.13 
#>  9 fam_ID_9     -0.972   -0.246   -0.597  -0.219   0.610   1.05  -0.221    1.09 
#> 10 fam_ID_10    -0.437   -1.86    -0.862  -1.46    1.14   -2.16  -0.183    0.291
#> # … with 990 more rows, 17 more variables: mhs2_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, s2_phenotype2_status <lgl>,
#> #   mhs1_phenotype2_status <lgl>, mhs2_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>,
#> #   mgf_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phen…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>          <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       1.75   2.22      0.236 -0.306   -0.605  1.05    0.216   -0.375
#>  2 fam_ID_2      -0.171 -0.469     1.35  -0.618    0.939 -0.963  -0.841    0.315
#>  3 fam_ID_3       0.261  0.386     0.509  0.572    1.46   1.57   -0.440    0.505
#>  4 fam_ID_4       0.781 -0.00912   1.83  -0.909   -0.716  0.806   1.32     0.809
#>  5 fam_ID_5       0.687  0.506    -0.425  0.0826  -0.532  0.685   1.63    -0.194
#>  6 fam_ID_6       0.585 -0.519     0.126  1.08     1.37  -0.0559 -0.178    0.437
#>  7 fam_ID_7       0.115 -0.675    -1.08   1.25    -2.27   1.20    0.0628   0.599
#>  8 fam_ID_8       0.962  1.79     -1.02  -0.0465   0.702  0.690  -0.172    1.07 
#>  9 fam_ID_9       0.634 -0.421    -0.169  1.09    -1.17  -0.771  -0.402   -0.672
#> 10 fam_ID_10      0.573  1.14     -1.19   1.55    -0.840  1.04    0.245   -0.985
#> # … with 990 more rows, 17 more variables: mhs2_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, s2_phenotype3_status <lgl>,
#> #   mhs1_phenotype3_status <lgl>, mhs2_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>,
#> #   mgf_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       2.95 -Inf       2.95    2.22    2.22
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       3.17 -Inf       3.17 -Inf       3.17
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.55 -Inf       3.55 -Inf       3.55
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       3.35 -Inf       3.35 -Inf       3.35
#>  5 fam_ID_5  fam_ID_5_1  o          1.78    1.78 -Inf       3.24 -Inf       3.24
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       3.35    1.62    1.62 -Inf       3.35
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.55 -Inf       3.55 -Inf       3.55
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       2.55 -Inf       2.55    1.78    1.78
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       3.55 -Inf       3.55 -Inf       3.55
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.68 -Inf       2.68 -Inf       2.68
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenot…¹ f_phe…² s1_phe…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>          <dbl>   <dbl>    <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1     -1.66    0.0688 -0.919   FALSE   FALSE   FALSE        50      43
#>  2 fam_ID_2     -0.646   0.374   1.22    FALSE   FALSE   FALSE        46      48
#>  3 fam_ID_3     -0.0242  0.597   0.597   FALSE   FALSE   FALSE        37      38
#>  4 fam_ID_4      0.987   0.213  -0.185   FALSE   FALSE   FALSE        61      63
#>  5 fam_ID_5      0.391   0.485  -0.00877 FALSE   FALSE   FALSE        57      69
#>  6 fam_ID_6      0.734   0.0686 -0.451   FALSE   FALSE   FALSE        31      37
#>  7 fam_ID_7      1.95    0.332   0.286   TRUE    FALSE   FALSE        52      70
#>  8 fam_ID_8      0.178   1.11    1.24    FALSE   FALSE   FALSE        42      35
#>  9 fam_ID_9      1.67   -0.163   0.151   TRUE    FALSE   FALSE        59      59
#> 10 fam_ID_10    -0.653   0.152   0.175   FALSE   FALSE   FALSE        36      34
#> # … with 90 more rows, 1 more variable: s1_phenotype1_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype1, ²​f_phenotype1, ³​s1_phenotype1,
#> #   ⁴​m_phenotype1_status, ⁵​f_phenotype1_status, ⁶​s1_phenotype1_status,
#> #   ⁷​m_phenotype1_aoo, ⁸​f_phenotype1_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1        0.112 -0.471  -0.262  FALSE   FALSE   FALSE        50      43
#>  2 fam_ID_2        1.59   1.04    1.33   TRUE    FALSE   TRUE         62      48
#>  3 fam_ID_3       -1.30   0.0100 -0.303  FALSE   FALSE   FALSE        37      38
#>  4 fam_ID_4        0.984  0.607   0.442  FALSE   FALSE   FALSE        61      63
#>  5 fam_ID_5       -1.47   1.13   -2.07   FALSE   FALSE   FALSE        57      69
#>  6 fam_ID_6       -0.986 -0.598  -0.731  FALSE   FALSE   FALSE        31      37
#>  7 fam_ID_7        1.38   1.56    0.639  TRUE    TRUE    FALSE        74      63
#>  8 fam_ID_8       -1.14   1.32    0.495  FALSE   TRUE    FALSE        42      82
#>  9 fam_ID_9        1.79   0.445  -0.106  TRUE    FALSE   FALSE        56      59
#> 10 fam_ID_10      -2.44  -0.647  -0.0532 FALSE   FALSE   FALSE        36      34
#> # … with 90 more rows, 1 more variable: s1_phenotype2_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype2, ²​f_phenotype2, ³​s1_phenotype2,
#> #   ⁴​m_phenotype2_status, ⁵​f_phenotype2_status, ⁶​s1_phenotype2_status,
#> #   ⁷​m_phenotype2_aoo, ⁸​f_phenotype2_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       -0.379 -0.586   -1.02  FALSE   FALSE   FALSE        50      43
#>  2 fam_ID_2        0.389  0.280    0.208 FALSE   FALSE   FALSE        46      48
#>  3 fam_ID_3       -0.188  0.341    0.649 FALSE   FALSE   FALSE        37      38
#>  4 fam_ID_4       -1.69  -0.731   -1.47  FALSE   FALSE   FALSE        61      63
#>  5 fam_ID_5       -0.109  0.920    0.517 FALSE   FALSE   FALSE        57      69
#>  6 fam_ID_6       -0.877 -0.666   -1.27  FALSE   FALSE   FALSE        31      37
#>  7 fam_ID_7       -0.144  0.770    0.714 FALSE   FALSE   FALSE        62      70
#>  8 fam_ID_8        1.08   2.40     2.50  FALSE   TRUE    TRUE         42      41
#>  9 fam_ID_9        0.524 -1.95     0.255 FALSE   FALSE   FALSE        58      59
#> 10 fam_ID_10      -1.23   0.0900   1.05  FALSE   FALSE   FALSE        36      34
#> # … with 90 more rows, 1 more variable: s1_phenotype3_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype3, ²​f_phenotype3, ³​s1_phenotype3,
#> #   ⁴​m_phenotype3_status, ⁵​f_phenotype3_status, ⁶​s1_phenotype3_status,
#> #   ⁷​m_phenotype3_aoo, ⁸​f_phenotype3_aoo
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m       -Inf       2.01 -Inf       2.01    -Inf    2.01
#>  2 fam_ID_2  fam_ID_2_1  m       -Inf       2.18    1.59    1.59    -Inf    2.18
#>  3 fam_ID_3  fam_ID_3_1  m       -Inf       2.55 -Inf       2.55    -Inf    2.55
#>  4 fam_ID_4  fam_ID_4_1  m       -Inf       1.62 -Inf       1.62    -Inf    1.62
#>  5 fam_ID_5  fam_ID_5_1  m       -Inf       1.74 -Inf       1.74    -Inf    1.74
#>  6 fam_ID_6  fam_ID_6_1  m       -Inf       2.79 -Inf       2.79    -Inf    2.79
#>  7 fam_ID_7  fam_ID_7_1  m          1.93    1.93    1.37    1.37    -Inf    1.59
#>  8 fam_ID_8  fam_ID_8_1  m       -Inf       2.34 -Inf       2.34    -Inf    2.34
#>  9 fam_ID_9  fam_ID_9_1  m          1.68    1.68    1.78    1.78    -Inf    1.71
#> 10 fam_ID_10 fam_ID_10_1 m       -Inf       2.59 -Inf       2.59    -Inf    2.59
#> # … with 290 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1      -0.434        -1.83   FALSE                             37
#>  2 fam_ID_2       0.00317       0.273  FALSE                             17
#>  3 fam_ID_3      -1.21         -1.84   FALSE                             25
#>  4 fam_ID_4       0.499         0.508  FALSE                             32
#>  5 fam_ID_5      -0.614        -0.446  FALSE                             33
#>  6 fam_ID_6      -0.672        -0.248  FALSE                             30
#>  7 fam_ID_7      -0.138        -0.279  FALSE                             40
#>  8 fam_ID_8       0.0558        1.23   FALSE                             35
#>  9 fam_ID_9       0.200         0.0932 FALSE                             34
#> 10 fam_ID_10     -0.579         0.578  FALSE                             27
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.603       -0.478  FALSE                             37
#>  2 fam_ID_2        0.454        0.0271 FALSE                             17
#>  3 fam_ID_3        0.0635      -0.412  FALSE                             25
#>  4 fam_ID_4       -0.906       -1.43   FALSE                             32
#>  5 fam_ID_5        0.420        0.359  FALSE                             33
#>  6 fam_ID_6        0.185        1.43   TRUE                              69
#>  7 fam_ID_7        0.272        0.326  FALSE                             40
#>  8 fam_ID_8       -0.789       -0.974  FALSE                             35
#>  9 fam_ID_9       -0.401       -0.0624 FALSE                             34
#> 10 fam_ID_10       0.953        1.34   TRUE                              78
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1         1.77        1.86   TRUE                              54
#>  2 fam_ID_2        -1.94       -0.0247 FALSE                             17
#>  3 fam_ID_3        -0.362      -0.168  FALSE                             25
#>  4 fam_ID_4         1.04        0.587  FALSE                             32
#>  5 fam_ID_5         1.34        0.252  FALSE                             33
#>  6 fam_ID_6        -0.924       0.616  FALSE                             30
#>  7 fam_ID_7        -0.398      -0.310  FALSE                             40
#>  8 fam_ID_8        -0.688      -1.10   FALSE                             35
#>  9 fam_ID_9         0.397      -1.10   FALSE                             34
#> 10 fam_ID_10        0.181       1.25   FALSE                             27
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          -Inf    2.55 -Inf       2.55    1.85    1.85
#>  2 fam_ID_2  fam_ID_2_1  o          -Inf    3.31 -Inf       3.31 -Inf       3.31
#>  3 fam_ID_3  fam_ID_3_1  o          -Inf    3.03 -Inf       3.03 -Inf       3.03
#>  4 fam_ID_4  fam_ID_4_1  o          -Inf    2.76 -Inf       2.76 -Inf       2.76
#>  5 fam_ID_5  fam_ID_5_1  o          -Inf    2.72 -Inf       2.72 -Inf       2.72
#>  6 fam_ID_6  fam_ID_6_1  o          -Inf    2.83    1.44    1.44 -Inf       2.83
#>  7 fam_ID_7  fam_ID_7_1  o          -Inf    2.43 -Inf       2.43 -Inf       2.43
#>  8 fam_ID_8  fam_ID_8_1  o          -Inf    2.63 -Inf       2.63 -Inf       2.63
#>  9 fam_ID_9  fam_ID_9_1  o          -Inf    2.68 -Inf       2.68 -Inf       2.68
#> 10 fam_ID_10 fam_ID_10_1 o          -Inf    2.95    1.34    1.34 -Inf       2.95
#> # … with 140 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> # ℹ Use `print(n = ...)` to see more rows
#>