Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_ph…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1 fam_ID_1      0.186  -0.0584   0.377 -0.0422 -1.34   -0.570    0.904  0.00667
#>  2 fam_ID_2     -0.481  -0.0355  -2.43   1.03   -0.487  -0.515   -1.14   0.284  
#>  3 fam_ID_3      0.706   0.227   -0.730  1.48    2.45   -0.832   -0.107  0.782  
#>  4 fam_ID_4      0.0142 -0.178    0.591 -0.280  -1.18    1.29     0.817  1.32   
#>  5 fam_ID_5      0.189   1.31     0.785  0.0553 -0.0619  0.0858   0.233  1.05   
#>  6 fam_ID_6      0.666   1.63     0.524  0.0499  1.02   -0.320    0.308  1.32   
#>  7 fam_ID_7      1.45    0.621    0.393  1.02   -0.858  -0.291   -1.61  -0.0498 
#>  8 fam_ID_8     -0.634  -0.677   -0.457 -1.05   -0.329  -0.363   -0.164  1.91   
#>  9 fam_ID_9      1.02   -0.351    0.979  0.970   0.264  -0.260   -0.498  0.212  
#> 10 fam_ID_10     0.995   1.83    -0.865  2.11    1.46    0.105    0.894  1.60   
#> # … with 990 more rows, 17 more variables: pgf_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   pgm_phenotype1_status <lgl>, pgf_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>,
#> #   s1_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1      -0.804  -1.41     0.969   0.265  0.415    1.05   -0.801 -0.0743
#>  2 fam_ID_2       0.0869 -0.367   -0.593  -0.211  0.0434   0.270   0.369  0.401 
#>  3 fam_ID_3       0.431  -0.0792   1.60   -0.294 -0.240   -0.291  -1.08   0.0327
#>  4 fam_ID_4      -0.0349 -0.331    1.02   -0.703  0.732    0.342   0.536 -0.226 
#>  5 fam_ID_5       0.853  -0.438    1.25   -0.372 -0.890    0.855   0.547 -0.0630
#>  6 fam_ID_6       1.12    2.59     0.800   1.61   0.284    1.37   -2.09   2.24  
#>  7 fam_ID_7       0.346   0.427   -1.20    1.38  -1.43    -1.37   -1.11   1.66  
#>  8 fam_ID_8       0.389   0.414    0.716  -0.345 -0.0327  -1.68   -1.14  -0.148 
#>  9 fam_ID_9       0.210  -0.585    1.34    0.501  0.596    1.09   -0.123 -2.48  
#> 10 fam_ID_10     -0.0553  0.280   -1.24   -1.52   0.102   -1.01   -0.204 -1.09  
#> # … with 990 more rows, 17 more variables: pgf_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   pgm_phenotype2_status <lgl>, pgf_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>,
#> #   s1_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ f_phe…⁴ s1_ph…⁵ mgm_p…⁶ mgf_p…⁷ pgm_p…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1        0.108  0.573    1.81   0.336   0.546   0.726   -0.299   0.331
#>  2 fam_ID_2        0.262  0.313   -0.406  0.178  -0.809   0.412    1.48   -0.638
#>  3 fam_ID_3       -0.328  2.07     0.885 -0.0323 -1.36    1.00    -1.27   -0.380
#>  4 fam_ID_4       -0.106  0.0736   1.70   0.593  -0.297  -0.495   -1.13   -1.01 
#>  5 fam_ID_5       -0.344  0.179    0.191  0.437  -0.277   0.204   -0.887   0.994
#>  6 fam_ID_6       -1.90  -1.67     0.266 -0.653  -1.47   -0.988   -1.46   -1.62 
#>  7 fam_ID_7       -0.159 -0.0825   0.999 -0.127  -1.61    1.67    -1.49    0.593
#>  8 fam_ID_8        0.674 -0.0339   1.27   0.577   2.35   -0.445    3.19    1.21 
#>  9 fam_ID_9        1.04   1.20     0.164 -0.228  -0.410  -0.0811  -0.227  -0.483
#> 10 fam_ID_10       0.596  1.30    -0.291  0.182  -0.0783 -1.43    -0.274   1.93 
#> # … with 990 more rows, 17 more variables: pgf_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   pgm_phenotype3_status <lgl>, pgf_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>,
#> #   s1_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       2.79 -Inf       2.79 -Inf       2.79
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       2.99 -Inf       2.99 -Inf       2.99
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.48 -Inf       3.48    2.09    2.09
#>  4 fam_ID_4  fam_ID_4_1  o       -Inf       3.21 -Inf       3.21 -Inf       3.21
#>  5 fam_ID_5  fam_ID_5_1  o          1.31    1.31 -Inf       3.28 -Inf       3.28
#>  6 fam_ID_6  fam_ID_6_1  o          1.62    1.62    2.59    2.59 -Inf       2.43
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       3.52 -Inf       3.52 -Inf       3.52
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       2.83 -Inf       2.83 -Inf       2.83
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.51 -Inf       2.51 -Inf       2.51
#> 10 fam_ID_10 fam_ID_10_1 o          1.81    1.81 -Inf       2.76    1.30    1.30
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.739   1.11    -1.18   0.691   0.419   -1.06   -1.14   0.544 
#>  2 fam_ID_2      -0.294  -0.615   -2.07  -0.0716 -0.538   -0.356  -0.237  0.0533
#>  3 fam_ID_3      -0.723  -0.140   -0.177 -0.158  -0.707   -0.790   0.272 -0.516 
#>  4 fam_ID_4      -0.271  -0.0508  -1.18   0.208  -0.740   -1.97   -0.394 -0.791 
#>  5 fam_ID_5      -0.935  -1.46     0.372  0.790  -0.952    0.595  -0.432 -0.171 
#>  6 fam_ID_6      -0.953  -0.456   -1.11  -1.11    0.0890  -1.34   -2.08   1.71  
#>  7 fam_ID_7       0.0401  0.441    0.458 -0.244   1.03    -0.183   0.129 -0.634 
#>  8 fam_ID_8       0.823   0.586    2.29  -1.17    1.94     1.03    0.958  2.45  
#>  9 fam_ID_9       0.460   1.06    -0.647  0.823   0.162   -0.186  -0.290 -1.79  
#> 10 fam_ID_10      0.928   0.473    1.61   0.541   1.12    -0.689   1.66  -2.23  
#> # … with 990 more rows, 17 more variables: mhs2_phenotype1 <dbl>,
#> #   o_phenotype1_status <lgl>, m_phenotype1_status <lgl>,
#> #   mgm_phenotype1_status <lgl>, mgf_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, s2_phenotype1_status <lgl>,
#> #   mhs1_phenotype1_status <lgl>, mhs2_phenotype1_status <lgl>,
#> #   o_phenotype1_aoo <dbl>, m_phenotype1_aoo <dbl>, mgm_phenotype1_aoo <dbl>,
#> #   mgf_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenot…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_phe…⁷ mhs1_…⁸
#>    <chr>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>
#>  1 fam_ID_1      0.0686   1.03  -2.18   -1.32   -0.470  -0.706  -0.789    -0.675
#>  2 fam_ID_2      0.518   -0.159  1.67    1.66    2.18    0.272  -0.00125   0.712
#>  3 fam_ID_3      0.122    0.767  1.35    0.175   1.50    0.275  -1.17      0.115
#>  4 fam_ID_4     -0.436   -0.323 -2.25   -0.0414 -0.291   1.04    0.335     1.08 
#>  5 fam_ID_5     -1.44    -2.36  -2.02   -0.450   0.135  -0.0657 -0.249    -0.355
#>  6 fam_ID_6     -0.218    0.235 -1.51    0.384  -0.0271  1.19   -1.04      0.713
#>  7 fam_ID_7      1.13     1.29   0.796   0.966  -1.31    0.324   0.321    -1.16 
#>  8 fam_ID_8     -0.559   -1.62  -0.951  -0.229  -0.815   0.261  -2.40     -0.657
#>  9 fam_ID_9     -1.71    -0.736  0.147  -1.12   -0.493  -0.522  -0.635    -0.288
#> 10 fam_ID_10    -0.0657   0.743 -0.0178 -0.0704  0.194   1.25   -0.0632    1.89 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype2 <dbl>,
#> #   o_phenotype2_status <lgl>, m_phenotype2_status <lgl>,
#> #   mgm_phenotype2_status <lgl>, mgf_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, s2_phenotype2_status <lgl>,
#> #   mhs1_phenotype2_status <lgl>, mhs2_phenotype2_status <lgl>,
#> #   o_phenotype2_aoo <dbl>, m_phenotype2_aoo <dbl>, mgm_phenotype2_aoo <dbl>,
#> #   mgf_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenoty…¹ o_phe…² m_phe…³ mgm_p…⁴ mgf_p…⁵ s1_ph…⁶ s2_ph…⁷ mhs1_…⁸
#>    <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1       0.175    0.646   0.236  -1.01    1.27   0.962    0.674  1.35  
#>  2 fam_ID_2       1.04     1.92   -0.868  -1.02    1.13  -0.417    0.455  0.694 
#>  3 fam_ID_3       0.0926   0.290   1.90    2.04    0.410  0.821   -0.916  0.633 
#>  4 fam_ID_4      -0.800   -1.91   -0.486   0.448  -0.410 -0.225   -1.48  -1.84  
#>  5 fam_ID_5       0.589    0.349   0.278   0.603   0.825 -0.957   -0.339 -0.0736
#>  6 fam_ID_6       1.27     1.69   -1.02   -0.400   0.120  1.34    -0.951  0.481 
#>  7 fam_ID_7       0.676    1.75    1.51   -0.551   1.07   0.601    0.726  1.01  
#>  8 fam_ID_8      -0.487   -0.343  -1.42    0.496  -0.238 -0.876   -0.851  1.84  
#>  9 fam_ID_9       0.351    0.386   0.371   1.24   -0.861  0.363   -0.433 -0.877 
#> 10 fam_ID_10      0.954    0.994   1.12    1.23   -0.595  0.0745  -1.29   0.987 
#> # … with 990 more rows, 17 more variables: mhs2_phenotype3 <dbl>,
#> #   o_phenotype3_status <lgl>, m_phenotype3_status <lgl>,
#> #   mgm_phenotype3_status <lgl>, mgf_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, s2_phenotype3_status <lgl>,
#> #   mhs1_phenotype3_status <lgl>, mhs2_phenotype3_status <lgl>,
#> #   o_phenotype3_aoo <dbl>, m_phenotype3_aoo <dbl>, mgm_phenotype3_aoo <dbl>,
#> #   mgf_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o          -Inf    3.38 -Inf       3.38 -Inf       3.38
#>  2 fam_ID_2  fam_ID_2_1  o          -Inf    3.35 -Inf       3.35    1.93    1.93
#>  3 fam_ID_3  fam_ID_3_1  o          -Inf    3.24 -Inf       3.24 -Inf       3.24
#>  4 fam_ID_4  fam_ID_4_1  o          -Inf    3.48 -Inf       3.48 -Inf       3.48
#>  5 fam_ID_5  fam_ID_5_1  o          -Inf    3.42 -Inf       3.42 -Inf       3.42
#>  6 fam_ID_6  fam_ID_6_1  o          -Inf    2.59 -Inf       2.59    1.68    1.68
#>  7 fam_ID_7  fam_ID_7_1  o          -Inf    3.35    1.29    1.29    1.74    1.74
#>  8 fam_ID_8  fam_ID_8_1  o          -Inf    3.42 -Inf       3.42 -Inf       3.42
#>  9 fam_ID_9  fam_ID_9_1  o          -Inf    3.21 -Inf       3.21 -Inf       3.21
#> 10 fam_ID_10 fam_ID_10_1 o          -Inf    2.68 -Inf       2.68 -Inf       2.68
#> # … with 7,990 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenot…¹ f_phen…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>          <dbl>    <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1      -0.981 -0.611   -0.0545 FALSE   FALSE   FALSE        32      37
#>  2 fam_ID_2      -0.232 -1.22    -1.40   FALSE   FALSE   FALSE        43      38
#>  3 fam_ID_3       0.143 -0.00188 -0.324  FALSE   FALSE   FALSE        43      49
#>  4 fam_ID_4      -4.35  -1.95    -2.80   FALSE   FALSE   FALSE        60      58
#>  5 fam_ID_5       1.30   0.0489   0.646  TRUE    FALSE   FALSE        89      61
#>  6 fam_ID_6       0.408  0.272    0.362  FALSE   FALSE   FALSE        35      33
#>  7 fam_ID_7       1.41  -1.38     0.492  TRUE    FALSE   FALSE        70      39
#>  8 fam_ID_8      -1.99  -1.07    -2.41   FALSE   FALSE   FALSE        56      63
#>  9 fam_ID_9       0.855  0.542    1.23   FALSE   FALSE   FALSE        55      59
#> 10 fam_ID_10     -0.140 -0.0518   0.842  FALSE   FALSE   FALSE        60      57
#> # … with 90 more rows, 1 more variable: s1_phenotype1_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype1, ²​f_phenotype1, ³​s1_phenotype1,
#> #   ⁴​m_phenotype1_status, ⁵​f_phenotype1_status, ⁶​s1_phenotype1_status,
#> #   ⁷​m_phenotype1_aoo, ⁸​f_phenotype1_aoo
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1      -0.226  -0.544   0.161  FALSE   FALSE   FALSE        32      37
#>  2 fam_ID_2      -1.03   -1.63   -2.52   FALSE   FALSE   FALSE        43      38
#>  3 fam_ID_3       0.503  -0.551  -0.0759 FALSE   FALSE   FALSE        43      49
#>  4 fam_ID_4      -0.376  -1.37    0.0496 FALSE   FALSE   FALSE        60      58
#>  5 fam_ID_5       1.19   -0.300   1.09   FALSE   FALSE   FALSE        63      61
#>  6 fam_ID_6       0.0384 -0.492  -0.163  FALSE   FALSE   FALSE        35      33
#>  7 fam_ID_7       1.20    0.0442  1.29   FALSE   FALSE   TRUE         47      39
#>  8 fam_ID_8       1.38   -0.0895 -0.913  TRUE    FALSE   FALSE        73      63
#>  9 fam_ID_9       0.101   0.638   0.935  FALSE   FALSE   FALSE        55      59
#> 10 fam_ID_10      1.06   -0.168  -0.388  FALSE   FALSE   FALSE        60      57
#> # … with 90 more rows, 1 more variable: s1_phenotype2_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype2, ²​f_phenotype2, ³​s1_phenotype2,
#> #   ⁴​m_phenotype2_status, ⁵​f_phenotype2_status, ⁶​s1_phenotype2_status,
#> #   ⁷​m_phenotype2_aoo, ⁸​f_phenotype2_aoo
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenoty…¹ f_phe…² s1_ph…³ m_phe…⁴ f_phe…⁵ s1_ph…⁶ m_phe…⁷ f_phe…⁸
#>    <chr>           <dbl>   <dbl>   <dbl> <lgl>   <lgl>   <lgl>     <dbl>   <dbl>
#>  1 fam_ID_1       0.775    1.09    0.941 FALSE   FALSE   FALSE        32      37
#>  2 fam_ID_2      -1.49    -1.29   -1.88  FALSE   FALSE   FALSE        43      38
#>  3 fam_ID_3       0.185   -1.65    0.397 FALSE   FALSE   FALSE        43      49
#>  4 fam_ID_4      -1.01    -1.64   -1.08  FALSE   FALSE   FALSE        60      58
#>  5 fam_ID_5       0.547   -0.896   0.933 FALSE   FALSE   FALSE        63      61
#>  6 fam_ID_6       0.0250  -0.439   0.109 FALSE   FALSE   FALSE        35      33
#>  7 fam_ID_7       1.08    -0.946  -0.245 FALSE   FALSE   FALSE        47      39
#>  8 fam_ID_8      -1.28    -0.268  -2.32  FALSE   FALSE   FALSE        56      63
#>  9 fam_ID_9       0.0853  -0.250   1.66  FALSE   FALSE   TRUE         55      59
#> 10 fam_ID_10      1.83    -1.11    0.340 TRUE    FALSE   FALSE        55      57
#> # … with 90 more rows, 1 more variable: s1_phenotype3_aoo <dbl>, and
#> #   abbreviated variable names ¹​m_phenotype3, ²​f_phenotype3, ³​s1_phenotype3,
#> #   ⁴​m_phenotype3_status, ⁵​f_phenotype3_status, ⁶​s1_phenotype3_status,
#> #   ⁷​m_phenotype3_aoo, ⁸​f_phenotype3_aoo
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m       -Inf       2.76 -Inf       2.76 -Inf       2.76
#>  2 fam_ID_2  fam_ID_2_1  m       -Inf       2.30 -Inf       2.30 -Inf       2.30
#>  3 fam_ID_3  fam_ID_3_1  m       -Inf       2.30 -Inf       2.30 -Inf       2.30
#>  4 fam_ID_4  fam_ID_4_1  m       -Inf       1.64 -Inf       1.64 -Inf       1.64
#>  5 fam_ID_5  fam_ID_5_1  m          1.30    1.30 -Inf       1.56 -Inf       1.56
#>  6 fam_ID_6  fam_ID_6_1  m       -Inf       2.63 -Inf       2.63 -Inf       2.63
#>  7 fam_ID_7  fam_ID_7_1  m          1.42    1.42 -Inf       2.13 -Inf       2.13
#>  8 fam_ID_8  fam_ID_8_1  m       -Inf       1.78    1.38    1.38 -Inf       1.78
#>  9 fam_ID_9  fam_ID_9_1  m       -Inf       1.81 -Inf       1.81 -Inf       1.81
#> 10 fam_ID_10 fam_ID_10_1 m       -Inf       1.64 -Inf       1.64    1.81    1.81
#> # … with 290 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        -0.694       0.0344 FALSE                             15
#>  2 fam_ID_2        -1.31       -2.01   FALSE                             12
#>  3 fam_ID_3        -0.932      -1.56   FALSE                             15
#>  4 fam_ID_4         0.949       2.61   TRUE                              36
#>  5 fam_ID_5        -1.28       -1.70   FALSE                             21
#>  6 fam_ID_6         0.542      -0.0869 FALSE                             17
#>  7 fam_ID_7        -0.599      -0.135  FALSE                             35
#>  8 fam_ID_8        -0.438      -0.629  FALSE                             26
#>  9 fam_ID_9        -1.43       -0.586  FALSE                             28
#> 10 fam_ID_10       -0.831      -0.327  FALSE                             33
#> # … with 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.761       -0.678  FALSE                             15
#>  2 fam_ID_2        0.380        0.0415 FALSE                             12
#>  3 fam_ID_3       -0.365       -0.454  FALSE                             15
#>  4 fam_ID_4        0.214        1.18   FALSE                             20
#>  5 fam_ID_5        0.0171      -0.0227 FALSE                             21
#>  6 fam_ID_6       -1.24        -1.13   FALSE                             17
#>  7 fam_ID_7       -0.755       -0.636  FALSE                             35
#>  8 fam_ID_8        0.0749      -0.115  FALSE                             26
#>  9 fam_ID_9        0.394        0.949  FALSE                             28
#> 10 fam_ID_10       0.690       -0.0130 FALSE                             33
#> # … with 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        -1.19       -0.172  FALSE                             15
#>  2 fam_ID_2         0.505       0.490  FALSE                             12
#>  3 fam_ID_3         0.231       0.0338 FALSE                             15
#>  4 fam_ID_4        -0.659      -1.26   FALSE                             20
#>  5 fam_ID_5        -0.734      -0.286  FALSE                             21
#>  6 fam_ID_6        -0.155       0.830  FALSE                             17
#>  7 fam_ID_7        -0.407       0.0151 FALSE                             35
#>  8 fam_ID_8        -1.32       -1.81   FALSE                             26
#>  9 fam_ID_9         0.363       0.444  FALSE                             28
#> 10 fam_ID_10        0.166       0.995  FALSE                             33
#> # … with 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID    role  lower_p…¹ upper…² lower…³ upper…⁴ lower…⁵ upper…⁶
#>    <chr>     <chr>       <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o       -Inf       3.38    -Inf    3.38    -Inf    3.38
#>  2 fam_ID_2  fam_ID_2_1  o       -Inf       3.48    -Inf    3.48    -Inf    3.48
#>  3 fam_ID_3  fam_ID_3_1  o       -Inf       3.38    -Inf    3.38    -Inf    3.38
#>  4 fam_ID_4  fam_ID_4_1  o          2.59    2.59    -Inf    3.21    -Inf    3.21
#>  5 fam_ID_5  fam_ID_5_1  o       -Inf       3.17    -Inf    3.17    -Inf    3.17
#>  6 fam_ID_6  fam_ID_6_1  o       -Inf       3.31    -Inf    3.31    -Inf    3.31
#>  7 fam_ID_7  fam_ID_7_1  o       -Inf       2.63    -Inf    2.63    -Inf    2.63
#>  8 fam_ID_8  fam_ID_8_1  o       -Inf       2.99    -Inf    2.99    -Inf    2.99
#>  9 fam_ID_9  fam_ID_9_1  o       -Inf       2.91    -Inf    2.91    -Inf    2.91
#> 10 fam_ID_10 fam_ID_10_1 o       -Inf       2.72    -Inf    2.72    -Inf    2.72
#> # … with 140 more rows, and abbreviated variable names ¹​lower_phenotype1,
#> #   ²​upper_phenotype1, ³​lower_phenotype2, ⁴​upper_phenotype2, ⁵​lower_phenotype3,
#> #   ⁶​upper_phenotype3
#>