Simulate under the liability threshold model (multiple phenotypes).

simulate_under_LTM_multi simulates families and thresholds under the liability threshold model for a given family structure and multiple phenotypes. Please note that it is not possible to simulate different family structures.

Usage

simulate_under_LTM_multi(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  genetic_corrmat = diag(3),
  full_corrmat = diag(3),
  h2_vec = rep(0.5, 3),
  phen_names = NULL,
  n_sim = 1000,
  pop_prev = rep(0.1, 3)
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

genetic_corrmat

A numeric matrix holding the genetic correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

full_corrmat

A numeric matrix holding the full correlations between the desired phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries must be between -1 and 1. In addition, the matrix must be symmetric. Defaults to diag(3).

h2_vec

A numeric vector holding the liability-scale heritabilities for a number of phenotype. All entries must be non-negative. Note that under the liability threshold model, the heritabilities must also be at most 1. Defaults to rep(0.5,3).

phen_names

A character vector holding the phenotype names. These names will be used to create the row and column names for the covariance matrix. If it is not specified, the names will default to phenotype1, phenotype2, etc. Defaults to NULL.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A numeric vector holding the population prevalences, i.e. the overall prevalences in the population. All entries in pop_prev must be positive and smaller than 1. Defaults to rep(.1,3).

Value

If either fam_vec or n_fam is used as the argument and if it is of the required format, if genetic_corrmat and full_corrmat are two numeric and symmetric matrices satisfying that all diagonal entries are one and that all off-diagonal entries are between -1 and 1, if the liability-scale heritabilities in h2_vec are numbers satisfying \(0 \leq h^2_i\) for all \(i \in \{1,...,n_pheno\}\), n_sim is a strictly positive number, and pop_prev is a positive numeric vector such that all entries are at most one, then the output will be a list containing lists for each phenotype. The first outer list, which is named after the first phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the first phenotype. As the first outer list, the second outer list, which is named after the second phenotype in phen_names, holds the tibble sim_obs, which holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families for the second phenotype. There is a list containing sim_obs for each phenotype in phen_names. The last list entry, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families and all phenotypes. Note that this tibble has the format required in estimate_liability. Finally, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_multi()
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype1 o_phenotype1 m_phenotype1 f_phenotype1 s1_phenotype1
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        0.959        1.34         1.38          0.820         2.04 
#>  2 fam_ID_2        1.46         2.33        -0.548         1.43          2.03 
#>  3 fam_ID_3       -0.573        0.0225      -0.481         1.46         -0.288
#>  4 fam_ID_4       -0.104       -0.343        0.848        -0.355         0.226
#>  5 fam_ID_5        0.603       -0.165        0.287        -0.640        -1.40 
#>  6 fam_ID_6        0.538        0.910        0.830        -1.45          0.693
#>  7 fam_ID_7       -0.0690      -0.425       -0.0800        0.844         0.250
#>  8 fam_ID_8       -0.175        0.131       -0.885        -0.675        -0.207
#>  9 fam_ID_9       -0.718       -0.706       -0.377         0.495        -0.166
#> 10 fam_ID_10       0.379        0.373        0.978        -0.412         0.384
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype1 <dbl>, mgf_phenotype1 <dbl>,
#> #   pgm_phenotype1 <dbl>, pgf_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, f_phenotype1_status <lgl>,
#> #   s1_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, pgm_phenotype1_status <lgl>,
#> #   pgf_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype2 o_phenotype2 m_phenotype2 f_phenotype2 s1_phenotype2
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1        -1.43      -0.746        -1.10        -0.444         -1.49 
#>  2 fam_ID_2        -0.428     -1.34         -0.0952      -0.791          1.40 
#>  3 fam_ID_3        -0.301     -0.645         1.71        -1.28           0.258
#>  4 fam_ID_4        -1.11      -2.16         -0.869       -2.03          -1.93 
#>  5 fam_ID_5         0.424      1.10          0.922        0.871         -0.895
#>  6 fam_ID_6         0.114     -1.16          1.10        -1.62          -0.263
#>  7 fam_ID_7         0.993      1.45          1.31         0.0951        -0.787
#>  8 fam_ID_8         0.434      1.04         -0.585        2.83           1.05 
#>  9 fam_ID_9        -0.174     -0.00258      -1.58        -1.02          -0.461
#> 10 fam_ID_10        0.424      0.970         1.06        -0.963          1.69 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype2 <dbl>, mgf_phenotype2 <dbl>,
#> #   pgm_phenotype2 <dbl>, pgf_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, f_phenotype2_status <lgl>,
#> #   s1_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, pgm_phenotype2_status <lgl>,
#> #   pgf_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID    g_phenotype3 o_phenotype3 m_phenotype3 f_phenotype3 s1_phenotype3
#>    <chr>            <dbl>        <dbl>        <dbl>        <dbl>         <dbl>
#>  1 fam_ID_1       -0.0725     -1.13          -0.784       -0.972        0.0342
#>  2 fam_ID_2        0.313       0.00468        0.990       -1.18         0.570 
#>  3 fam_ID_3       -0.510      -0.928          0.570       -0.349       -0.0342
#>  4 fam_ID_4        0.0908     -0.0394         1.63         0.305        2.16  
#>  5 fam_ID_5       -0.941      -1.83           0.512        0.956       -0.377 
#>  6 fam_ID_6       -0.980      -0.356         -1.78        -0.548       -1.06  
#>  7 fam_ID_7       -0.723      -0.798         -1.35        -0.793       -0.999 
#>  8 fam_ID_8       -0.475      -1.20           0.422       -1.64         0.597 
#>  9 fam_ID_9        0.655      -0.592          1.30        -1.26        -0.836 
#> 10 fam_ID_10      -1.02       -0.0343         0.926       -0.677       -0.893 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: mgm_phenotype3 <dbl>, mgf_phenotype3 <dbl>,
#> #   pgm_phenotype3 <dbl>, pgf_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, f_phenotype3_status <lgl>,
#> #   s1_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, pgm_phenotype3_status <lgl>,
#> #   pgf_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o                 1.34             1.34          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 2.34             2.34          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o              -Inf                2.72          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o              -Inf                3.21          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o              -Inf                2.99          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o              -Inf                2.59          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o              -Inf                3.21             1.45
#>  8 fam_ID_8  fam_ID_8_1 o              -Inf                2.95          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o              -Inf                2.43          -Inf   
#> 10 fam_ID_10 fam_ID_10… o              -Inf                2.79          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

genetic_corrmat <- matrix(0.4, 3, 3)
diag(genetic_corrmat) <- 1
full_corrmat <- matrix(0.6, 3, 3)
diag(full_corrmat) <- 1

simulate_under_LTM_multi(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2,2), 
c("m","mgm","mgf","s","mhs")))
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype1 o_phenotype1 m_phenotype1 mgm_phenotype1 mgf_phenotype1
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1       0.364         1.27         0.731         0.919          -0.485
#>  2 fam_ID_2      -1.04         -1.87        -1.70         -0.353          -0.990
#>  3 fam_ID_3      -0.634        -1.41        -0.667        -0.892          -1.71 
#>  4 fam_ID_4      -0.216         0.203       -0.102        -0.195           0.679
#>  5 fam_ID_5       0.0215       -0.109       -0.187        -0.465           0.611
#>  6 fam_ID_6       0.477        -0.552       -0.365         1.49            0.318
#>  7 fam_ID_7       0.402        -0.265       -0.267        -0.703           0.211
#>  8 fam_ID_8      -0.489        -0.622       -0.882        -0.0498         -0.670
#>  9 fam_ID_9       1.00          1.17         0.394        -0.555           1.25 
#> 10 fam_ID_…      -0.210         0.405        0.512        -1.04           -0.327
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype1 <dbl>, s2_phenotype1 <dbl>,
#> #   mhs1_phenotype1 <dbl>, mhs2_phenotype1 <dbl>, o_phenotype1_status <lgl>,
#> #   m_phenotype1_status <lgl>, mgm_phenotype1_status <lgl>,
#> #   mgf_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   s2_phenotype1_status <lgl>, mhs1_phenotype1_status <lgl>,
#> #   mhs2_phenotype1_status <lgl>, o_phenotype1_aoo <dbl>, …
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype2 o_phenotype2 m_phenotype2 mgm_phenotype2 mgf_phenotype2
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1      -0.0964      -0.290        -0.220         -0.820        -0.259 
#>  2 fam_ID_2      -0.0480      -0.0149        0.625          0.260         1.07  
#>  3 fam_ID_3      -0.406        0.941         0.587         -0.670        -0.0582
#>  4 fam_ID_4      -0.363       -0.338         1.69           1.01          0.137 
#>  5 fam_ID_5       0.908        1.35          0.792          0.461        -0.771 
#>  6 fam_ID_6       0.315       -0.0514       -0.552         -0.328         0.288 
#>  7 fam_ID_7      -0.950       -0.129         0.515          1.33          1.85  
#>  8 fam_ID_8      -0.0159      -0.567         1.02          -0.513        -1.85  
#>  9 fam_ID_9      -0.581        0.112         0.142         -0.528         1.68  
#> 10 fam_ID_…       0.770        0.633         1.53           1.66          0.574 
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype2 <dbl>, s2_phenotype2 <dbl>,
#> #   mhs1_phenotype2 <dbl>, mhs2_phenotype2 <dbl>, o_phenotype2_status <lgl>,
#> #   m_phenotype2_status <lgl>, mgm_phenotype2_status <lgl>,
#> #   mgf_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   s2_phenotype2_status <lgl>, mhs1_phenotype2_status <lgl>,
#> #   mhs2_phenotype2_status <lgl>, o_phenotype2_aoo <dbl>, …
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID   g_phenotype3 o_phenotype3 m_phenotype3 mgm_phenotype3 mgf_phenotype3
#>    <chr>           <dbl>        <dbl>        <dbl>          <dbl>          <dbl>
#>  1 fam_ID_1      -0.0144      -0.0573       0.130          0.0207        -0.862 
#>  2 fam_ID_2       0.105       -0.232        1.72           0.366          1.80  
#>  3 fam_ID_3       0.126        0.322       -0.349          0.860         -0.0310
#>  4 fam_ID_4       0.667       -0.0342       1.57          -1.03           0.448 
#>  5 fam_ID_5       1.08         1.33         0.936         -0.214          0.634 
#>  6 fam_ID_6       0.492        0.934        1.08           0.204          0.823 
#>  7 fam_ID_7       0.560        1.76        -0.474         -0.230         -0.250 
#>  8 fam_ID_8       0.501        0.737        0.0492         0.539          0.808 
#>  9 fam_ID_9       1.43         1.98         2.74           2.63          -0.841 
#> 10 fam_ID_…       0.228        0.624        0.524          0.944         -0.0716
#> # ℹ 990 more rows
#> # ℹ 20 more variables: s1_phenotype3 <dbl>, s2_phenotype3 <dbl>,
#> #   mhs1_phenotype3 <dbl>, mhs2_phenotype3 <dbl>, o_phenotype3_status <lgl>,
#> #   m_phenotype3_status <lgl>, mgm_phenotype3_status <lgl>,
#> #   mgf_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   s2_phenotype3_status <lgl>, mhs1_phenotype3_status <lgl>,
#> #   mhs2_phenotype3_status <lgl>, o_phenotype3_aoo <dbl>, …
#> 
#> 
#> $thresholds
#> # A tibble: 8,000 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o                 -Inf             2.87          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 -Inf             2.87          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o                 -Inf             3.31          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 -Inf             2.43          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o                 -Inf             2.59             1.35
#>  6 fam_ID_6  fam_ID_6_1 o                 -Inf             3.03          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 -Inf             3.21          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o                 -Inf             2.79          -Inf   
#>  9 fam_ID_9  fam_ID_9_1 o                 -Inf             3.48          -Inf   
#> 10 fam_ID_10 fam_ID_10… o                 -Inf             2.51          -Inf   
#> # ℹ 7,990 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c("m","f","s1"), add_ind = FALSE, 
genetic_corrmat = genetic_corrmat, full_corrmat = full_corrmat, n_sim = 100)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype1 f_phenotype1 s1_phenotype1 m_phenotype1_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        1.29        -0.656        -0.519  TRUE               
#>  2 fam_ID_2       -1.31         0.225        -0.430  FALSE              
#>  3 fam_ID_3        0.187        0.435         0.220  FALSE              
#>  4 fam_ID_4        0.0120       0.637        -0.231  FALSE              
#>  5 fam_ID_5        0.199        1.46         -0.779  FALSE              
#>  6 fam_ID_6        1.04        -1.28         -0.761  FALSE              
#>  7 fam_ID_7       -0.233        0.581         0.365  FALSE              
#>  8 fam_ID_8       -2.32         0.0143       -0.768  FALSE              
#>  9 fam_ID_9       -0.659       -0.0169       -0.0988 FALSE              
#> 10 fam_ID_10      -0.765       -0.0794       -1.73   FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype1_status <lgl>, s1_phenotype1_status <lgl>,
#> #   m_phenotype1_aoo <dbl>, f_phenotype1_aoo <dbl>, s1_phenotype1_aoo <dbl>
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype2 f_phenotype2 s1_phenotype2 m_phenotype2_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        0.491        -1.03         0.418  FALSE              
#>  2 fam_ID_2       -1.00          0.141       -0.605  FALSE              
#>  3 fam_ID_3        0.991         1.35         0.787  FALSE              
#>  4 fam_ID_4        0.306         0.373       -0.652  FALSE              
#>  5 fam_ID_5        0.0785        1.52        -0.0798 FALSE              
#>  6 fam_ID_6       -0.665        -0.928       -0.751  FALSE              
#>  7 fam_ID_7       -0.587         0.732       -0.109  FALSE              
#>  8 fam_ID_8        1.08         -0.769       -0.565  FALSE              
#>  9 fam_ID_9       -1.52          0.369        0.487  FALSE              
#> 10 fam_ID_10      -0.788         0.773       -0.322  FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype2_status <lgl>, s1_phenotype2_status <lgl>,
#> #   m_phenotype2_aoo <dbl>, f_phenotype2_aoo <dbl>, s1_phenotype2_aoo <dbl>
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 100 × 10
#>    fam_ID    m_phenotype3 f_phenotype3 s1_phenotype3 m_phenotype3_status
#>    <chr>            <dbl>        <dbl>         <dbl> <lgl>              
#>  1 fam_ID_1        0.0131       -0.574        0.137  FALSE              
#>  2 fam_ID_2       -0.387         0.634        0.0247 FALSE              
#>  3 fam_ID_3        2.01         -0.632        1.24   TRUE               
#>  4 fam_ID_4       -1.67          1.37        -1.10   FALSE              
#>  5 fam_ID_5        2.31          0.246        1.48   TRUE               
#>  6 fam_ID_6        0.816        -1.32        -1.25   FALSE              
#>  7 fam_ID_7       -0.326         0.230       -0.565  FALSE              
#>  8 fam_ID_8       -0.900        -0.895       -1.03   FALSE              
#>  9 fam_ID_9       -0.689         0.190       -1.16   FALSE              
#> 10 fam_ID_10      -2.11         -0.136       -1.80   FALSE              
#> # ℹ 90 more rows
#> # ℹ 5 more variables: f_phenotype3_status <lgl>, s1_phenotype3_status <lgl>,
#> #   m_phenotype3_aoo <dbl>, f_phenotype3_aoo <dbl>, s1_phenotype3_aoo <dbl>
#> 
#> 
#> $thresholds
#> # A tibble: 300 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 m                 1.29             1.29             -Inf
#>  2 fam_ID_2  fam_ID_2_1 m              -Inf                1.64             -Inf
#>  3 fam_ID_3  fam_ID_3_1 m              -Inf                1.45             -Inf
#>  4 fam_ID_4  fam_ID_4_1 m              -Inf                2.13             -Inf
#>  5 fam_ID_5  fam_ID_5_1 m              -Inf                2.18             -Inf
#>  6 fam_ID_6  fam_ID_6_1 m              -Inf                1.44             -Inf
#>  7 fam_ID_7  fam_ID_7_1 m              -Inf                1.54             -Inf
#>  8 fam_ID_8  fam_ID_8_1 m              -Inf                2.26             -Inf
#>  9 fam_ID_9  fam_ID_9_1 m              -Inf                1.81             -Inf
#> 10 fam_ID_10 fam_ID_10… m              -Inf                2.22             -Inf
#> # ℹ 290 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#> 

simulate_under_LTM_multi(fam_vec = c(), n_fam = NULL, add_ind = TRUE, n_sim = 150)
#> $phenotype1
#> $phenotype1$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype1 o_phenotype1 o_phenotype1_status o_phenotype1_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.437       -0.0939 FALSE                             15
#>  2 fam_ID_2       -0.898       -0.937  FALSE                             27
#>  3 fam_ID_3       -0.144       -0.293  FALSE                             36
#>  4 fam_ID_4       -0.492       -2.29   FALSE                             34
#>  5 fam_ID_5        1.21        -0.194  FALSE                             35
#>  6 fam_ID_6       -0.331       -0.384  FALSE                             22
#>  7 fam_ID_7        1.62         1.09   FALSE                             24
#>  8 fam_ID_8        0.377       -0.485  FALSE                             12
#>  9 fam_ID_9       -0.0210      -0.167  FALSE                             15
#> 10 fam_ID_10       0.0714      -0.639  FALSE                             30
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype2
#> $phenotype2$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype2 o_phenotype2 o_phenotype2_status o_phenotype2_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1        0.103       -0.0641 FALSE                             15
#>  2 fam_ID_2        0.523        0.384  FALSE                             27
#>  3 fam_ID_3       -0.482       -0.886  FALSE                             36
#>  4 fam_ID_4        0.565        1.10   FALSE                             34
#>  5 fam_ID_5        0.271       -0.645  FALSE                             35
#>  6 fam_ID_6       -0.946       -0.905  FALSE                             22
#>  7 fam_ID_7       -0.0989      -0.0752 FALSE                             24
#>  8 fam_ID_8        0.802        1.78   TRUE                              56
#>  9 fam_ID_9       -0.201       -0.538  FALSE                             15
#> 10 fam_ID_10      -0.452        0.795  FALSE                             30
#> # ℹ 140 more rows
#> 
#> 
#> $phenotype3
#> $phenotype3$sim_obs
#> # A tibble: 150 × 5
#>    fam_ID    g_phenotype3 o_phenotype3 o_phenotype3_status o_phenotype3_aoo
#>    <chr>            <dbl>        <dbl> <lgl>                          <dbl>
#>  1 fam_ID_1       -0.356      -0.0968  FALSE                             15
#>  2 fam_ID_2       -0.278      -1.13    FALSE                             27
#>  3 fam_ID_3        1.12        1.96    TRUE                              51
#>  4 fam_ID_4        1.11        0.524   FALSE                             34
#>  5 fam_ID_5        0.379       0.772   FALSE                             35
#>  6 fam_ID_6        0.0283     -0.00122 FALSE                             22
#>  7 fam_ID_7        0.248       0.807   FALSE                             24
#>  8 fam_ID_8        0.117      -1.22    FALSE                             12
#>  9 fam_ID_9       -1.12       -0.290   FALSE                             15
#> 10 fam_ID_10      -0.599      -0.880   FALSE                             30
#> # ℹ 140 more rows
#> 
#> 
#> $thresholds
#> # A tibble: 150 × 9
#>    fam_ID    indiv_ID   role  lower_phenotype1 upper_phenotype1 lower_phenotype2
#>    <chr>     <chr>      <chr>            <dbl>            <dbl>            <dbl>
#>  1 fam_ID_1  fam_ID_1_1 o                 -Inf             3.38          -Inf   
#>  2 fam_ID_2  fam_ID_2_1 o                 -Inf             2.95          -Inf   
#>  3 fam_ID_3  fam_ID_3_1 o                 -Inf             2.59          -Inf   
#>  4 fam_ID_4  fam_ID_4_1 o                 -Inf             2.68          -Inf   
#>  5 fam_ID_5  fam_ID_5_1 o                 -Inf             2.63          -Inf   
#>  6 fam_ID_6  fam_ID_6_1 o                 -Inf             3.14          -Inf   
#>  7 fam_ID_7  fam_ID_7_1 o                 -Inf             3.06          -Inf   
#>  8 fam_ID_8  fam_ID_8_1 o                 -Inf             3.48             1.78
#>  9 fam_ID_9  fam_ID_9_1 o                 -Inf             3.38          -Inf   
#> 10 fam_ID_10 fam_ID_10… o                 -Inf             2.83          -Inf   
#> # ℹ 140 more rows
#> # ℹ 3 more variables: upper_phenotype2 <dbl>, lower_phenotype3 <dbl>,
#> #   upper_phenotype3 <dbl>
#>