Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g       o        m      f     s1    mgm     mgf    pgm     pgf
#>    <chr>      <dbl>   <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 fam_ID_1 -0.384   0.587  -0.667    0.268 -1.15  -2.01   0.418  -0.815 -0.222 
#>  2 fam_ID_2 -0.0553 -0.475  -0.139   -1.04  -1.16   0.766 -0.634  -0.138 -0.498 
#>  3 fam_ID_3  0.932   0.997   0.132   -0.285  1.46   0.826  0.0614 -1.29   0.744 
#>  4 fam_ID_4  0.737   1.07    0.299   -0.875  0.431 -0.403  1.61   -1.42   0.0840
#>  5 fam_ID_5  0.333   0.115  -0.702    1.29   0.560 -0.437  0.285   2.39  -1.37  
#>  6 fam_ID_6 -0.384  -1.35    0.00208 -0.791 -2.60  -0.592 -0.415  -0.609 -1.59  
#>  7 fam_ID_7  0.0221  0.744  -0.595   -0.604 -1.38  -0.224 -0.963  -0.637  1.86  
#>  8 fam_ID_8  0.160   0.506   1.04     0.675  1.24  -0.880  0.464  -1.16  -0.431 
#>  9 fam_ID_9 -0.164  -0.0716  0.653   -0.571  0.576 -0.790 -0.111  -0.466 -0.454 
#> 10 fam_ID_… -0.222   0.680   1.09     1.62  -0.896 -2.56   1.43    0.292 -0.449 
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.83
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  2.87
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.06
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.10
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.14
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.72
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.59
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.35
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.31
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.10
#> # … with 7,990 more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID         g      o       m     mgm     mgf   mhs1   mhs2 o_sta…¹ m_sta…²
#>    <chr>      <dbl>  <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl> <lgl>   <lgl>  
#>  1 fam_ID_1 -0.854  -0.874 -0.0952  1.22    1.16    1.81  -0.702 FALSE   FALSE  
#>  2 fam_ID_2 -1.07   -0.735 -0.477  -0.536  -1.38   -0.137 -1.88  FALSE   FALSE  
#>  3 fam_ID_3  0.832   1.38  -0.128  -0.661   0.607   1.91  -0.873 TRUE    FALSE  
#>  4 fam_ID_4  0.0508 -0.504 -0.547  -0.0961 -1.08   -0.343 -1.62  FALSE   FALSE  
#>  5 fam_ID_5 -0.767  -0.647 -2.17    0.582  -0.988   0.468 -1.15  FALSE   FALSE  
#>  6 fam_ID_6  0.299   1.01   0.634  -0.567  -1.59    0.243  0.525 FALSE   FALSE  
#>  7 fam_ID_7 -0.561  -0.424 -0.0973 -0.449   0.972   0.952  1.31  FALSE   FALSE  
#>  8 fam_ID_8 -0.209  -0.677 -0.538   2.92   -1.69    1.80   1.21  FALSE   FALSE  
#>  9 fam_ID_9 -0.278  -0.673 -0.284  -1.64   -0.0172 -0.346 -1.29  FALSE   FALSE  
#> 10 fam_ID_…  0.310   0.531  1.03    0.891   1.14   -1.03  -1.64  FALSE   FALSE  
#> # … with 990 more rows, 10 more variables: mgm_status <lgl>, mgf_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>, and
#> #   abbreviated variable names ¹​o_status, ²​m_status
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.24
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.99
#>  3 fam_ID_3  fam_ID_3_1  o        1.38  1.38
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.14
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.31
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.42
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.87
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.38
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.55
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.28
#> # … with 5,990 more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID          m      f      s1 m_status f_status s1_st…¹ m_aoo f_aoo s1_aoo
#>    <chr>       <dbl>  <dbl>   <dbl> <lgl>    <lgl>    <lgl>   <dbl> <dbl>  <dbl>
#>  1 fam_ID_1  -0.0465 -0.763 -0.0540 FALSE    FALSE    FALSE      58    68     38
#>  2 fam_ID_2   0.120   0.841  1.74   FALSE    FALSE    TRUE       50    54     72
#>  3 fam_ID_3   1.72    1.10  -2.32   TRUE     FALSE    FALSE      75    46     16
#>  4 fam_ID_4   0.951  -1.39  -2.43   FALSE    FALSE    FALSE      57    60     36
#>  5 fam_ID_5   0.154   0.491 -1.88   FALSE    FALSE    FALSE      38    39     14
#>  6 fam_ID_6   1.04   -0.554  0.307  FALSE    FALSE    FALSE      57    60     31
#>  7 fam_ID_7   1.32   -0.728 -0.642  FALSE    FALSE    FALSE      47    42     19
#>  8 fam_ID_8   0.805   0.366  0.298  FALSE    FALSE    FALSE      30    31     12
#>  9 fam_ID_9   0.648  -0.111 -1.28   FALSE    FALSE    FALSE      48    50     24
#> 10 fam_ID_10 -0.676   2.77   1.09   FALSE    TRUE     FALSE      64    37     39
#> # … with 490 more rows, and abbreviated variable name ¹​s1_status
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     2.02
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     2.29
#>  3 fam_ID_3  fam_ID_3_1  m        1.71  1.71
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     2.05
#>  5 fam_ID_5  fam_ID_5_1  m     -Inf     2.75
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     2.05
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     2.40
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     3.05
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     2.36
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     1.86
#> # … with 1,490 more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID          g        o o_status o_aoo
#>    <chr>       <dbl>    <dbl> <lgl>    <dbl>
#>  1 fam_ID_1   0.245   0.00759 FALSE       19
#>  2 fam_ID_2  -1.14   -0.847   FALSE       25
#>  3 fam_ID_3   0.178   1.46    FALSE       34
#>  4 fam_ID_4  -0.0981 -0.734   FALSE       22
#>  5 fam_ID_5   1.32    1.19    FALSE       40
#>  6 fam_ID_6   0.744  -0.345   FALSE       19
#>  7 fam_ID_7  -1.20   -1.54    FALSE       29
#>  8 fam_ID_8  -0.879  -1.75    FALSE       39
#>  9 fam_ID_9   0.248   0.158   FALSE       29
#> 10 fam_ID_10  0.191  -0.775   FALSE       11
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.44
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.23
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.90
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.33
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.67
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.44
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.09
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  2.71
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.09
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.70
#> # … with 190 more rows
#>