Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g       o       m       f     s1    mgm    mgf    pgm    pgf
#>    <chr>      <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 fam_ID_1   0.318  1.13   -0.807   0.230   0.598  0.333 -1.50  -0.858  0.369
#>  2 fam_ID_2  -0.796 -1.25   -0.938   0.0496  0.288 -3.13   0.844 -0.491  0.937
#>  3 fam_ID_3  -0.482 -0.760   0.929  -1.84   -1.32   0.133  1.02  -0.610 -0.897
#>  4 fam_ID_4  -0.154 -0.385  -0.548  -0.519   0.960  1.13  -0.934 -0.449 -0.177
#>  5 fam_ID_5  -0.282  0.268   0.412   0.277  -0.186  1.16   1.66   0.666  0.730
#>  6 fam_ID_6   1.15   2.20   -0.392   0.228   1.99   0.366 -0.155  1.72   0.167
#>  7 fam_ID_7  -0.923 -0.0222  0.299  -1.51   -2.23   0.865 -1.31  -1.37  -1.59 
#>  8 fam_ID_8  -1.05  -0.700  -0.815  -1.30   -2.22  -0.279 -1.15  -1.09  -0.286
#>  9 fam_ID_9   0.270  0.185   0.0700  1.20    1.11  -0.860 -0.249 -0.692  2.06 
#> 10 fam_ID_10 -0.858 -0.231  -0.387   0.669   0.394 -0.167 -1.10  -1.21  -0.860
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.72
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.76
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.28
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     2.43
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.68
#>  6 fam_ID_6  fam_ID_6_1  o        2.18  2.18
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.72
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.91
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.24
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.17
#> # … with 7,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID           g      o        m     mgm    mgf     mhs1   mhs2 o_status
#>    <chr>        <dbl>  <dbl>    <dbl>   <dbl>  <dbl>    <dbl>  <dbl> <lgl>   
#>  1 fam_ID_1  -0.0678  -0.792 -0.646   -0.181  -1.68   0.193   -0.455 FALSE   
#>  2 fam_ID_2   0.499    1.06   0.257   -1.03    0.446  0.865   -1.07  FALSE   
#>  3 fam_ID_3  -0.00381 -0.901 -0.00387 -2.05   -1.44   0.102    0.175 FALSE   
#>  4 fam_ID_4  -0.347   -0.548  1.63     0.308   0.548  0.731    2.23  FALSE   
#>  5 fam_ID_5  -0.941   -1.63  -2.34    -1.37    0.144 -1.93    -2.78  FALSE   
#>  6 fam_ID_6  -0.0947  -0.956 -0.464    0.0433 -0.358  1.38    -0.747 FALSE   
#>  7 fam_ID_7  -0.460    1.58   1.03    -0.123   2.11  -0.802    0.921 TRUE    
#>  8 fam_ID_8   0.461    0.632  2.07     1.24   -0.959  0.192    1.00  FALSE   
#>  9 fam_ID_9   1.04     1.20  -0.742   -0.158   1.37  -0.00178  0.547 FALSE   
#> 10 fam_ID_10 -0.229   -2.30   0.633   -0.105  -0.505  1.15     1.06  FALSE   
#> # … with 990 more rows, and 11 more variables: m_status <lgl>,
#> #   mgm_status <lgl>, mgf_status <lgl>, mhs1_status <lgl>, mhs2_status <lgl>,
#> #   o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, mhs1_aoo <dbl>,
#> #   mhs2_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.63
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.38
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.55
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.31
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.59
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     2.72
#>  7 fam_ID_7  fam_ID_7_1  o        1.59  1.59
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.55
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.52
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.06
#> # … with 5,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID         m       f      s1 m_status f_status s1_st…¹ m_aoo f_aoo s1_aoo
#>    <chr>      <dbl>   <dbl>   <dbl> <lgl>    <lgl>    <lgl>   <dbl> <dbl>  <dbl>
#>  1 fam_ID_1  -0.245  1.11   -0.814  FALSE    FALSE    FALSE      52    54     30
#>  2 fam_ID_2  -1.20  -0.636   0.773  FALSE    FALSE    FALSE      61    67     38
#>  3 fam_ID_3   0.450 -0.513   1.93   FALSE    FALSE    TRUE       50    49     61
#>  4 fam_ID_4  -0.958 -0.821  -0.0809 FALSE    FALSE    FALSE      63    61     38
#>  5 fam_ID_5   0.216 -0.0993 -0.650  FALSE    FALSE    FALSE      60    60     40
#>  6 fam_ID_6  -0.430 -0.307  -0.238  FALSE    FALSE    FALSE      66    69     40
#>  7 fam_ID_7   0.610  0.750  -2.34   FALSE    FALSE    FALSE      30    30     10
#>  8 fam_ID_8  -0.234 -1.09   -1.34   FALSE    FALSE    FALSE      52    52     27
#>  9 fam_ID_9   1.83   0.611  -0.718  TRUE     FALSE    FALSE      66    40     10
#> 10 fam_ID_10 -1.61  -0.269  -0.0651 FALSE    FALSE    FALSE      57    57     33
#> # … with 490 more rows, and abbreviated variable name ¹​s1_status
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     2.21
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     1.93
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     2.29
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     1.89
#>  5 fam_ID_5  fam_ID_5_1  m     -Inf     1.96
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     1.83
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     3.05
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     2.21
#>  9 fam_ID_9  fam_ID_9_1  m        1.83  1.83
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     2.05
#> # … with 1,490 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID         g       o o_status o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -0.269 -0.649  FALSE       17
#>  2 fam_ID_2   0.211 -0.0511 FALSE       17
#>  3 fam_ID_3  -0.580 -0.0201 FALSE       39
#>  4 fam_ID_4  -0.642 -1.62   FALSE       24
#>  5 fam_ID_5   0.347 -1.30   FALSE       15
#>  6 fam_ID_6   0.603  0.701  FALSE       38
#>  7 fam_ID_7   0.466 -0.674  FALSE       18
#>  8 fam_ID_8  -0.453 -0.688  FALSE       31
#>  9 fam_ID_9   0.376  1.11   FALSE       38
#> 10 fam_ID_10 -0.231 -1.26   FALSE       28
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.50
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.50
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.71
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.26
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.57
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.75
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.47
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.01
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  2.75
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.12
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>