Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID           g       o       m      f     s1    mgm     mgf    pgm    pgf
#>    <chr>        <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1 fam_ID_1  -0.418   -0.608  -0.655   0.747 -1.09  -0.632 -0.228   0.203  1.27 
#>  2 fam_ID_2   1.33     1.14   -1.14    1.63   1.82  -0.295 -1.58    1.91  -0.158
#>  3 fam_ID_3  -0.478   -0.776   0.0326  0.966  1.03  -0.425 -0.873   1.24  -1.40 
#>  4 fam_ID_4  -0.833   -0.784  -1.90   -1.61   1.39  -0.628 -1.60    1.19   0.631
#>  5 fam_ID_5  -0.170   -0.687  -0.0454  0.359 -0.980 -0.371  0.492   1.23   0.621
#>  6 fam_ID_6   0.00742  0.394  -0.642   0.590  1.03  -0.162 -0.945   2.75  -0.241
#>  7 fam_ID_7   0.217    0.0646 -0.319   0.705  1.77  -1.76  -0.0450  1.80   0.134
#>  8 fam_ID_8   0.852    0.0846  0.288  -0.653 -1.24  -0.554  1.52   -1.46   0.813
#>  9 fam_ID_9   0.153    0.474   0.0326 -0.285  1.12   0.489  0.367  -1.68   0.422
#> 10 fam_ID_10  0.258   -0.463   1.40    0.499  0.818  1.02  -0.812   0.941 -0.662
#> # ℹ 990 more rows
#> # ℹ 16 more variables: o_status <lgl>, m_status <lgl>, f_status <lgl>,
#> #   s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, pgm_status <lgl>,
#> #   pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.28
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  2.99
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.55
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.10
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.59
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.14
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.48
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.52
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  2.79
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  2.55
#> # ℹ 7,990 more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID       g      o      m     mgm     mgf    mhs1   mhs2 o_status m_status
#>    <chr>    <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>  <dbl> <lgl>    <lgl>   
#>  1 fam_I…  0.0593 -0.823  0.598  0.819   1.36    0.337  -0.987 FALSE    FALSE   
#>  2 fam_I…  1.05    1.14  -0.261  0.515  -0.606   0.521  -1.40  FALSE    FALSE   
#>  3 fam_I…  0.372  -0.892 -0.563 -0.508  -0.983   0.378   0.237 FALSE    FALSE   
#>  4 fam_I… -0.471  -0.694 -0.784 -0.993  -0.572   0.0112 -1.87  FALSE    FALSE   
#>  5 fam_I… -0.656  -0.114  1.12  -0.0224  0.996   1.91   -0.412 FALSE    FALSE   
#>  6 fam_I… -0.397   0.408 -1.76   0.470  -1.28    0.115  -1.19  FALSE    FALSE   
#>  7 fam_I…  0.293   1.83  -0.288 -0.289   1.45   -1.30   -0.701 TRUE     FALSE   
#>  8 fam_I…  0.301   0.818  0.524  1.06    0.0433  1.36    0.642 FALSE    FALSE   
#>  9 fam_I…  0.393   0.468 -1.08  -0.234  -0.984   1.34   -1.08  FALSE    FALSE   
#> 10 fam_I… -0.573  -0.697 -0.507  0.187  -0.686  -0.572   0.364 FALSE    FALSE   
#> # ℹ 990 more rows
#> # ℹ 10 more variables: mgm_status <lgl>, mgf_status <lgl>, mhs1_status <lgl>,
#> #   mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>,
#> #   mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.21
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.28
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     2.55
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.21
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.76
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.42
#>  7 fam_ID_7  fam_ID_7_1  o        1.81  1.81
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.43
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.03
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.99
#> # ℹ 5,990 more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID         m      f     s1 m_status f_status s1_status m_aoo f_aoo s1_aoo
#>    <chr>      <dbl>  <dbl>  <dbl> <lgl>    <lgl>    <lgl>     <dbl> <dbl>  <dbl>
#>  1 fam_ID_1  -0.894 -1.53  -1.69  FALSE    FALSE    FALSE        63    62     38
#>  2 fam_ID_2   1.82   1.29   1.50  TRUE     FALSE    FALSE        66    64     34
#>  3 fam_ID_3   2.29  -0.896 -0.716 TRUE     FALSE    FALSE        50    47     26
#>  4 fam_ID_4   1.07   1.63   0.925 FALSE    FALSE    FALSE        46    46     24
#>  5 fam_ID_5   0.818  0.221 -0.760 FALSE    FALSE    FALSE        56    55     28
#>  6 fam_ID_6  -0.527  0.340 -0.914 FALSE    FALSE    FALSE        58    52     32
#>  7 fam_ID_7  -0.899 -1.17  -0.585 FALSE    FALSE    FALSE        54    49     26
#>  8 fam_ID_8  -1.54  -1.43  -1.54  FALSE    FALSE    FALSE        56    56     31
#>  9 fam_ID_9   0.480  0.189  1.82  FALSE    FALSE    TRUE         69    60     66
#> 10 fam_ID_10 -1.89  -1.03  -0.982 FALSE    FALSE    FALSE        56    48     30
#> # ℹ 490 more rows
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     1.89
#>  2 fam_ID_2  fam_ID_2_1  m        1.83  1.83
#>  3 fam_ID_3  fam_ID_3_1  m        2.29  2.29
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     2.44
#>  5 fam_ID_5  fam_ID_5_1  m     -Inf     2.08
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     2.02
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     2.14
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     2.08
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     1.78
#> 10 fam_ID_10 fam_ID_10_1 m     -Inf     2.08
#> # ℹ 1,490 more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID         g      o o_status o_aoo
#>    <chr>      <dbl>  <dbl> <lgl>    <dbl>
#>  1 fam_ID_1   0.205 -1.21  FALSE       27
#>  2 fam_ID_2  -0.636 -1.58  FALSE       14
#>  3 fam_ID_3   1.59   2.54  TRUE        43
#>  4 fam_ID_4   0.426  0.164 FALSE       15
#>  5 fam_ID_5  -0.153 -1.07  FALSE       21
#>  6 fam_ID_6   1.34   2.49  TRUE        45
#>  7 fam_ID_7  -0.519 -2.29  FALSE       13
#>  8 fam_ID_8  -0.533  0.539 FALSE       16
#>  9 fam_ID_9   0.129  0.700 FALSE       18
#> 10 fam_ID_10  0.311 -0.437 FALSE       18
#> # ℹ 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.16
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.60
#>  3 fam_ID_3  fam_ID_3_1  o        2.55  2.55
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.57
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.37
#>  6 fam_ID_6  fam_ID_6_1  o        2.48  2.48
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     3.63
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.54
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.47
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.47
#> # ℹ 190 more rows
#>