Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g      o       m      f     s1    mgm    mgf     pgm     pgf
#>    <chr>      <dbl>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>
#>  1 fam_ID_1  -0.345  0.273  0.706   0.747  0.523  1.17  -1.14   1.51   -1.14  
#>  2 fam_ID_2  -0.227 -0.253 -0.642  -0.599  0.126  1.49   0.158  0.129   0.876 
#>  3 fam_ID_3   0.828  1.87   0.458   0.756  1.10   0.617  1.27   1.04   -0.517 
#>  4 fam_ID_4   0.450  1.41   0.780   0.869  1.56   0.384  0.747  1.54    0.233 
#>  5 fam_ID_5   0.231 -0.734  0.737   0.566 -0.368  1.30  -0.420 -1.49    0.695 
#>  6 fam_ID_6  -0.250  0.449 -0.980  -1.37  -0.676 -0.167 -1.38   0.465   0.866 
#>  7 fam_ID_7   0.174  0.254 -0.115   1.83   1.01  -1.16   1.05   0.456   1.80  
#>  8 fam_ID_8  -0.586 -2.05   0.0767  1.15   1.85   0.546 -0.492  0.0604  0.988 
#>  9 fam_ID_9  -0.153  1.90   0.465   1.29   0.436  0.656  0.179  1.27    1.24  
#> 10 fam_ID_10 -0.117 -0.217  0.327   0.785  0.875  0.859 -0.419  0.362   0.0753
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.91
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.95
#>  3 fam_ID_3  fam_ID_3_1  o        1.89  1.89
#>  4 fam_ID_4  fam_ID_4_1  o        1.42  1.42
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.43
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.10
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.72
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.76
#>  9 fam_ID_9  fam_ID_9_1  o        1.89  1.89
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.99
#> # … with 7,990 more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID        g      o       m    mgm     mgf    mhs1    mhs2 o_sta…¹ m_sta…²
#>    <chr>     <dbl>  <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl> <lgl>   <lgl>  
#>  1 fam_I…  0.850    1.58   0.427  -0.354  0.179   0.300   0.575  TRUE    FALSE  
#>  2 fam_I… -0.535   -1.40  -0.0883 -0.321 -0.952  -0.418   1.34   FALSE   FALSE  
#>  3 fam_I…  0.114   -0.370 -0.0191 -1.18   1.57    0.0408  0.263  FALSE   FALSE  
#>  4 fam_I… -0.561   -1.86   0.768   0.206  0.354   0.251   0.0987 FALSE   FALSE  
#>  5 fam_I…  1.67     1.55   0.285  -0.220 -0.431   0.477   0.661  TRUE    FALSE  
#>  6 fam_I… -0.500   -0.371 -0.966   0.300 -0.0905 -0.524  -0.788  FALSE   FALSE  
#>  7 fam_I…  0.00811 -0.477  0.478  -0.224 -0.861   0.351  -0.246  FALSE   FALSE  
#>  8 fam_I…  0.394    0.339  0.587   0.430 -0.301  -0.792  -0.243  FALSE   FALSE  
#>  9 fam_I…  0.0859   0.818  0.606   0.873  1.23    1.51    0.392  FALSE   FALSE  
#> 10 fam_I… -0.596    0.387  0.291  -0.363 -0.410   1.17   -1.36   FALSE   FALSE  
#> # … with 990 more rows, 10 more variables: mgm_status <lgl>, mgf_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>, and
#> #   abbreviated variable names ¹​o_status, ²​m_status
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o        1.59  1.59
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.06
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.10
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     2.43
#>  5 fam_ID_5  fam_ID_5_1  o        1.54  1.54
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     2.47
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     3.42
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.68
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.31
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.06
#> # … with 5,990 more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID          m      f      s1 m_status f_status s1_st…¹ m_aoo f_aoo s1_aoo
#>    <chr>       <dbl>  <dbl>   <dbl> <lgl>    <lgl>    <lgl>   <dbl> <dbl>  <dbl>
#>  1 fam_ID_1   1.16    2.29   0.393  FALSE    TRUE     FALSE      60    50     31
#>  2 fam_ID_2  -1.57   -1.92  -0.942  FALSE    FALSE    FALSE      37    32     14
#>  3 fam_ID_3  -0.171   0.248  0.509  FALSE    FALSE    FALSE      53    46     28
#>  4 fam_ID_4  -0.311   1.15   1.14   FALSE    FALSE    FALSE      60    64     38
#>  5 fam_ID_5   0.533   1.48   0.707  FALSE    FALSE    FALSE      55    62     35
#>  6 fam_ID_6  -1.02   -1.27  -0.766  FALSE    FALSE    FALSE      57    52     27
#>  7 fam_ID_7  -0.644   1.42  -0.0675 FALSE    FALSE    FALSE      56    51     30
#>  8 fam_ID_8  -0.0612 -2.09  -1.49   FALSE    FALSE    FALSE      35    35     11
#>  9 fam_ID_9  -1.01   -0.108  0.587  FALSE    FALSE    FALSE      64    58     38
#> 10 fam_ID_10  0.714  -0.857 -0.361  FALSE    FALSE    FALSE      54    52     29
#> # … with 490 more rows, and abbreviated variable name ¹​s1_status
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m      -Inf  1.96
#>  2 fam_ID_2  fam_ID_2_1  m      -Inf  2.79
#>  3 fam_ID_3  fam_ID_3_1  m      -Inf  2.18
#>  4 fam_ID_4  fam_ID_4_1  m      -Inf  1.96
#>  5 fam_ID_5  fam_ID_5_1  m      -Inf  2.11
#>  6 fam_ID_6  fam_ID_6_1  m      -Inf  2.05
#>  7 fam_ID_7  fam_ID_7_1  m      -Inf  2.08
#>  8 fam_ID_8  fam_ID_8_1  m      -Inf  2.86
#>  9 fam_ID_9  fam_ID_9_1  m      -Inf  1.86
#> 10 fam_ID_10 fam_ID_10_1 m      -Inf  2.14
#> # … with 1,490 more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID          g       o o_status o_aoo
#>    <chr>       <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1   1.11    1.34   FALSE       35
#>  2 fam_ID_2   0.256   0.615  FALSE       37
#>  3 fam_ID_3   0.414  -0.263  FALSE       23
#>  4 fam_ID_4   1.28    1.84   TRUE        65
#>  5 fam_ID_5   0.416   0.0757 FALSE       29
#>  6 fam_ID_6   0.0566  0.330  FALSE       32
#>  7 fam_ID_7  -0.192  -0.868  FALSE       10
#>  8 fam_ID_8   1.21    1.43   FALSE       10
#>  9 fam_ID_9   0.957   0.870  FALSE       25
#> 10 fam_ID_10  1.16    1.22   FALSE       40
#> # … with 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.86
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.79
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.30
#>  4 fam_ID_4  fam_ID_4_1  o        1.84  1.84
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.09
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     2.97
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     3.73
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.73
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.23
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     2.67
#> # … with 190 more rows
#>