Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID        g        o      m      f       s1    mgm     mgf     pgm    pgf
#>    <chr>     <dbl>    <dbl>  <dbl>  <dbl>    <dbl>  <dbl>   <dbl>   <dbl>  <dbl>
#>  1 fam_ID…  0.0751 -0.677    2.10   2.07   2.69    -0.391 -0.693   0.264   0.261
#>  2 fam_ID… -0.855  -1.72    -2.38  -0.615 -0.974   -1.03  -1.80   -0.324  -0.448
#>  3 fam_ID… -0.0348 -0.762   -0.927 -1.13  -0.826    0.849 -0.576  -0.906  -0.840
#>  4 fam_ID…  0.308   0.600   -0.856  0.526 -1.25    -0.642 -0.584  -1.45   -0.733
#>  5 fam_ID…  0.525  -0.303    2.19   0.162  1.81     1.70   1.33   -0.481  -1.01 
#>  6 fam_ID…  0.643   0.385   -0.588 -0.774  0.857    1.36   0.701  -0.114   1.07 
#>  7 fam_ID…  0.475   0.541    2.22   0.988 -0.233    2.16   2.45   -0.544   0.486
#>  8 fam_ID…  1.21   -0.254    0.121  1.98   2.29    -0.546  0.509   0.0184 -0.150
#>  9 fam_ID…  1.06    3.00     1.29  -0.864 -0.00213  0.995  0.649  -1.65    1.63 
#> 10 fam_ID…  0.135  -0.00962 -0.315  0.759 -0.665   -0.750 -0.0177  2.00    0.301
#> # … with 990 more rows, and 16 more variables: o_status <lgl>, m_status <lgl>,
#> #   f_status <lgl>, s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   pgm_status <lgl>, pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>,
#> #   s1_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.55
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.17
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     2.91
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.03
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.95
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.14
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     3.55
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.24
#>  9 fam_ID_9  fam_ID_9_1  o        2.99  2.99
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.14
#> # … with 7,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID        g       o       m     mgm     mgf   mhs1   mhs2 o_sta…¹ m_sta…²
#>    <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl> <lgl>   <lgl>  
#>  1 fam_ID… -0.0398 -0.592  -0.659  -1.27   -0.991   0.694  1.68  FALSE   FALSE  
#>  2 fam_ID… -0.622  -0.417  -1.10    0.0652 -0.419  -0.478  1.23  FALSE   FALSE  
#>  3 fam_ID… -0.479  -0.819  -0.696   0.0284  0.340  -1.28   1.44  FALSE   FALSE  
#>  4 fam_ID…  0.436  -1.48   -1.05    0.666   0.987  -0.685 -1.08  FALSE   FALSE  
#>  5 fam_ID…  0.642   0.0821 -0.215  -0.753   2.24    0.447  0.729 FALSE   FALSE  
#>  6 fam_ID… -0.178  -1.06   -0.647  -1.39   -0.673  -0.474 -0.737 FALSE   FALSE  
#>  7 fam_ID… -0.650  -0.380   0.318   1.79   -0.0749  0.148 -1.21  FALSE   FALSE  
#>  8 fam_ID… -0.228  -0.103  -0.554   0.817  -1.25   -0.537  0.615 FALSE   FALSE  
#>  9 fam_ID… -0.617  -0.848   0.321  -0.832   1.64    1.31   0.560 FALSE   FALSE  
#> 10 fam_ID… -0.0548 -0.498   0.0398 -0.339  -1.15   -0.873 -1.79  FALSE   FALSE  
#> # … with 990 more rows, 10 more variables: mgm_status <lgl>, mgf_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>, and
#> #   abbreviated variable names ¹​o_status, ²​m_status
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.55
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.28
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.10
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.47
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.91
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.06
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.45
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.24
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  2.83
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.35
#> # … with 5,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID          m       f     s1 m_status f_status s1_st…¹ m_aoo f_aoo s1_aoo
#>    <chr>       <dbl>   <dbl>  <dbl> <lgl>    <lgl>    <lgl>   <dbl> <dbl>  <dbl>
#>  1 fam_ID_1   1.51    0.0268 -0.315 FALSE    FALSE    FALSE      67    62     37
#>  2 fam_ID_2  -0.976  -0.134   0.882 FALSE    FALSE    FALSE      46    43     18
#>  3 fam_ID_3  -1.22    1.47    1.29  FALSE    FALSE    FALSE      43    46     20
#>  4 fam_ID_4  -1.22   -0.206  -0.866 FALSE    FALSE    FALSE      62    62     35
#>  5 fam_ID_5  -1.40    1.41    1.11  FALSE    FALSE    FALSE      59    62     38
#>  6 fam_ID_6   0.151   1.31   -0.335 FALSE    FALSE    FALSE      44    41     22
#>  7 fam_ID_7   0.602  -0.419   0.122 FALSE    FALSE    FALSE      46    46     24
#>  8 fam_ID_8   0.0332  0.616   1.22  FALSE    FALSE    FALSE      49    53     31
#>  9 fam_ID_9   0.533   0.0139 -0.447 FALSE    FALSE    FALSE      37    39     13
#> 10 fam_ID_10  0.757   0.886  -0.218 FALSE    FALSE    FALSE      39    40     13
#> # … with 490 more rows, and abbreviated variable name ¹​s1_status
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m      -Inf  1.81
#>  2 fam_ID_2  fam_ID_2_1  m      -Inf  2.44
#>  3 fam_ID_3  fam_ID_3_1  m      -Inf  2.55
#>  4 fam_ID_4  fam_ID_4_1  m      -Inf  1.91
#>  5 fam_ID_5  fam_ID_5_1  m      -Inf  1.99
#>  6 fam_ID_6  fam_ID_6_1  m      -Inf  2.51
#>  7 fam_ID_7  fam_ID_7_1  m      -Inf  2.44
#>  8 fam_ID_8  fam_ID_8_1  m      -Inf  2.32
#>  9 fam_ID_9  fam_ID_9_1  m      -Inf  2.79
#> 10 fam_ID_10 fam_ID_10_1 m      -Inf  2.71
#> # … with 1,490 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID           g        o o_status o_aoo
#>    <chr>        <dbl>    <dbl> <lgl>    <dbl>
#>  1 fam_ID_1   0.225    0.279   FALSE       26
#>  2 fam_ID_2  -0.962   -1.07    FALSE       25
#>  3 fam_ID_3  -0.815    0.382   FALSE       33
#>  4 fam_ID_4   0.849   -0.0781  FALSE       39
#>  5 fam_ID_5  -0.222    0.00559 FALSE       40
#>  6 fam_ID_6   0.0507   0.424   FALSE       38
#>  7 fam_ID_7   0.127   -0.195   FALSE       35
#>  8 fam_ID_8   0.862    0.824   FALSE       25
#>  9 fam_ID_9  -1.10    -2.38    FALSE       19
#> 10 fam_ID_10  0.00540  0.148   FALSE       25
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.19
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.23
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.94
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.71
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.67
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.75
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.86
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.23
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.44
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.23
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>