Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID          g       o      m      f     s1    mgm    mgf    pgm      pgf
#>    <chr>       <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>    <dbl>
#>  1 fam_ID_1  -0.968  -0.577  -1.14  -1.12   0.495  0.466  0.991  0.377 -0.0957 
#>  2 fam_ID_2   1.24    0.783   0.430  0.684  0.325  0.803  0.191 -0.520  0.213  
#>  3 fam_ID_3  -0.283  -0.113  -0.118 -2.77  -0.158  0.510  1.47  -0.231 -0.388  
#>  4 fam_ID_4   0.191   0.747   0.715 -1.12  -1.15  -0.146  1.63  -0.139 -2.15   
#>  5 fam_ID_5   0.794  -0.0910  0.448 -0.183  2.03   0.412  0.417 -0.683  0.729  
#>  6 fam_ID_6  -0.730  -2.75    0.324 -1.16   0.949  0.757  0.267 -1.38  -1.51   
#>  7 fam_ID_7  -0.146   0.144  -0.976 -0.831 -1.80  -0.932 -0.530 -0.531  0.00280
#>  8 fam_ID_8  -0.0312 -1.06   -0.203 -0.849  0.413  0.571  0.622  1.67  -1.14   
#>  9 fam_ID_9   0.771   0.520  -0.367  0.458  0.736 -0.274 -0.445 -0.352  0.690  
#> 10 fam_ID_10  0.346   0.840   0.694  1.26  -0.263 -0.675  1.28   0.886  0.256  
#> # ℹ 990 more rows
#> # ℹ 16 more variables: o_status <lgl>, m_status <lgl>, f_status <lgl>,
#> #   s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, pgm_status <lgl>,
#> #   pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  2.87
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  2.55
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.38
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.76
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.48
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  2.99
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.83
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  2.76
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.38
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  2.79
#> # ℹ 7,990 more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID         g        o       m     mgm     mgf    mhs1    mhs2 o_status
#>    <chr>      <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <lgl>   
#>  1 fam_ID_1   0.130  1.19     0.992  -0.604   0.649   0.663  -0.780  FALSE   
#>  2 fam_ID_2   0.292 -0.800    0.0344  2.23   -0.989  -0.141  -0.863  FALSE   
#>  3 fam_ID_3  -0.224 -0.277   -1.14    0.0236 -0.359   0.246   0.0718 FALSE   
#>  4 fam_ID_4   0.250 -0.0881   1.27    0.379  -0.500   1.48   -0.747  FALSE   
#>  5 fam_ID_5  -0.259 -0.00750  0.685   1.30   -0.316  -0.342  -1.56   FALSE   
#>  6 fam_ID_6   0.127 -0.0447  -0.342  -0.0496 -0.447   0.0326  0.125  FALSE   
#>  7 fam_ID_7   0.510  2.01    -0.210   0.778  -1.63    1.35   -2.40   TRUE    
#>  8 fam_ID_8   0.463  0.751   -0.283  -0.0470 -0.378  -0.323  -0.874  FALSE   
#>  9 fam_ID_9   0.932  0.439    0.283   0.292  -0.0742  0.718   0.579  FALSE   
#> 10 fam_ID_10 -0.372 -1.83    -0.454  -0.215   0.378  -0.389   0.0452 FALSE   
#> # ℹ 990 more rows
#> # ℹ 11 more variables: m_status <lgl>, mgm_status <lgl>, mgf_status <lgl>,
#> #   mhs1_status <lgl>, mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.17
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     3.55
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.17
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.48
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.14
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     2.59
#>  7 fam_ID_7  fam_ID_7_1  o        2.01  2.01
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     2.76
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.38
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.17
#> # ℹ 5,990 more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID        m       f     s1 m_status f_status s1_status m_aoo f_aoo s1_aoo
#>    <chr>     <dbl>   <dbl>  <dbl> <lgl>    <lgl>    <lgl>     <dbl> <dbl>  <dbl>
#>  1 fam_ID_1 -0.486 -0.796   0.414 FALSE    FALSE    FALSE        42    49     23
#>  2 fam_ID_2 -0.541  0.380   0.290 FALSE    FALSE    FALSE        50    54     25
#>  3 fam_ID_3 -0.282 -0.353  -0.373 FALSE    FALSE    FALSE        37    40     16
#>  4 fam_ID_4  0.311 -0.0661  1.50  FALSE    FALSE    FALSE        56    54     26
#>  5 fam_ID_5 -1.30   1.06   -0.340 FALSE    FALSE    FALSE        44    49     23
#>  6 fam_ID_6  0.194  0.165   0.734 FALSE    FALSE    FALSE        41    52     22
#>  7 fam_ID_7  0.245  0.168   0.813 FALSE    FALSE    FALSE        48    52     24
#>  8 fam_ID_8 -0.432 -0.232   0.876 FALSE    FALSE    FALSE        32    40     13
#>  9 fam_ID_9  0.434  0.458   0.975 FALSE    FALSE    FALSE        50    48     27
#> 10 fam_ID_…  2.46  -0.690   1.54  TRUE     FALSE    FALSE        45    44     20
#> # ℹ 490 more rows
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m     -Inf     2.59
#>  2 fam_ID_2  fam_ID_2_1  m     -Inf     2.29
#>  3 fam_ID_3  fam_ID_3_1  m     -Inf     2.79
#>  4 fam_ID_4  fam_ID_4_1  m     -Inf     2.08
#>  5 fam_ID_5  fam_ID_5_1  m     -Inf     2.51
#>  6 fam_ID_6  fam_ID_6_1  m     -Inf     2.63
#>  7 fam_ID_7  fam_ID_7_1  m     -Inf     2.36
#>  8 fam_ID_8  fam_ID_8_1  m     -Inf     2.97
#>  9 fam_ID_9  fam_ID_9_1  m     -Inf     2.29
#> 10 fam_ID_10 fam_ID_10_1 m        2.48  2.48
#> # ℹ 1,490 more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID         g       o o_status o_aoo
#>    <chr>      <dbl>   <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -0.457  0.706  FALSE       15
#>  2 fam_ID_2  -0.354 -0.241  FALSE       37
#>  3 fam_ID_3   0.106 -0.882  FALSE       23
#>  4 fam_ID_4  -0.256 -0.0976 FALSE       19
#>  5 fam_ID_5   0.322  1.46   FALSE       14
#>  6 fam_ID_6   0.686  1.21   FALSE       30
#>  7 fam_ID_7   1.25   1.21   FALSE       16
#>  8 fam_ID_8  -0.450 -0.915  FALSE       20
#>  9 fam_ID_9   0.421  0.515  FALSE       30
#> 10 fam_ID_10 -1.18  -0.487  FALSE       19
#> # ℹ 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.57
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  2.79
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  3.30
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  3.44
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  3.60
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.05
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  3.54
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.40
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.05
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.44
#> # ℹ 190 more rows
#>