Simulate under the liability threshold model (single phenotype).

simulate_under_LTM_single simulates families and thresholds under the liability threshold model for a given family structure and a single phenotype. Please note that it is not possible to simulate different family structures.

simulate_under_LTM_single(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  h2 = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)

Arguments

fam_vec

A vector of strings holding the different family members. All family members must be represented by strings from the following list:

m (Mother)

f (Father)

c[0-9]*.[0-9]* (Children)

mgm (Maternal grandmother)

mgf (Maternal grandfather)

pgm (Paternal grandmother)

pgf (Paternal grandfather)

s[0-9]* (Full siblings)

mhs[0-9]* (Half-siblings - maternal side)

phs[0-9]* (Half-siblings - paternal side)

mau[0-9]* (Aunts/Uncles - maternal side)

pau[0-9]* (Aunts/Uncles - paternal side). Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").

n_fam

A named vector holding the desired number of family members. See setNames. All names must be picked from the list mentioned above. Defaults to NULL.

add_ind

A logical scalar indicating whether the genetic component of the full liability as well as the full liability for the underlying target individual should be included in the covariance matrix. Defaults to TRUE.

h2

A number representing the liability-scale heritability for a single phenotype. Must be non-negative. Note that under the liability threshold model, the heritability must also be at most 1. Defaults to 0.5.

n_sim

A positive number representing the number of simulations. Defaults to 1000.

pop_prev

A positive number representing the population prevalence, i.e. the overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.

Value

If either fam_vec or n_fam is used as the argument, if it is of the required format, if the liability-scale heritability h2

is a number satisfying \(0 \leq h^2\), n_sim is a strictly positive number, and pop_prev is a positive number that is at most one, then the output will be a list holding two tibbles. The first tibble, sim_obs, holds the simulated liabilities, the disease status and the current age/age-of-onset for all family members in each of the n_sim families. The second tibble, thresholds, holds the family identifier, the personal identifier, the role (specified in fam_vec or n_fam) as well as the lower and upper thresholds for all individuals in all families. Note that this tibble has the format required in estimate_liability. In addition, note that if neither fam_vec nor n_fam are specified, the function returns the disease status, the current age/age-of-onset, the lower and upper thresholds, as well as the personal identifier for a single individual, namely the individual under consideration (called o). If both fam_vec and n_fam are defined, the user is asked to ' decide on which of the two vectors to use.

See also

Examples

simulate_under_LTM_single()
#> $sim_obs
#> # A tibble: 1,000 × 26
#>    fam_ID         g      o      m       f      s1    mgm     mgf    pgm     pgf
#>    <chr>      <dbl>  <dbl>  <dbl>   <dbl>   <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 fam_ID_1  -0.631 -1.06   0.916  0.516   0.416   0.229  1.31    0.698 -0.393 
#>  2 fam_ID_2   0.473  0.333  0.289  0.131  -0.124  -0.160  0.0828  0.696 -0.385 
#>  3 fam_ID_3  -0.177 -0.538  0.786  0.568  -0.662   0.250 -0.0635 -0.459  0.877 
#>  4 fam_ID_4   0.122  0.333  1.30   1.49   -0.263   0.519 -0.815   0.503 -0.507 
#>  5 fam_ID_5   0.824  0.518  0.619  1.19    1.49   -0.610  0.570   1.68  -0.634 
#>  6 fam_ID_6   0.382 -0.373 -0.528  1.63   -1.68    0.135  1.70    1.81  -1.01  
#>  7 fam_ID_7  -0.788 -1.17  -0.216 -1.14    0.412   2.95   1.12   -0.660 -1.25  
#>  8 fam_ID_8  -0.382  0.326 -0.163  0.572  -0.656   0.625  0.250   0.803 -0.658 
#>  9 fam_ID_9  -0.238  0.243  1.57   0.0768  0.0668 -0.414  0.719  -1.85  -0.0738
#> 10 fam_ID_10  0.279  1.65  -0.389  1.53   -0.572  -2.23   1.04    1.42  -0.394 
#> # ℹ 990 more rows
#> # ℹ 16 more variables: o_status <lgl>, m_status <lgl>, f_status <lgl>,
#> #   s1_status <lgl>, mgm_status <lgl>, mgf_status <lgl>, pgm_status <lgl>,
#> #   pgf_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, f_aoo <dbl>, s1_aoo <dbl>,
#> #   mgm_aoo <dbl>, mgf_aoo <dbl>, pgm_aoo <dbl>, pgf_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 8,000 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     2.95
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.59
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.21
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     2.72
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     3.52
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.42
#>  7 fam_ID_7  fam_ID_7_1  o     -Inf     2.63
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.55
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     2.87
#> 10 fam_ID_10 fam_ID_10_1 o        1.64  1.64
#> # ℹ 7,990 more rows
#> 
simulate_under_LTM_single(fam_vec = NULL, n_fam = stats::setNames(c(1,1,1,2), 
c("m","mgm","mgf","mhs")))
#> $sim_obs
#> # A tibble: 1,000 × 20
#>    fam_ID      g       o      m    mgm     mgf    mhs1    mhs2 o_status m_status
#>    <chr>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl> <lgl>    <lgl>   
#>  1 fam_I… -0.305  0.508  -1.83  -1.06  -1.36    0.0710 -1.46   FALSE    FALSE   
#>  2 fam_I… -1.27  -1.02    0.727  0.919 -1.62    0.489   0.0354 FALSE    FALSE   
#>  3 fam_I… -0.795 -0.0452 -1.09  -1.60  -0.634   0.449   0.295  FALSE    FALSE   
#>  4 fam_I…  0.710  0.0896 -1.24  -1.04  -0.0404 -1.47   -1.20   FALSE    FALSE   
#>  5 fam_I… -0.780 -1.44   -0.958 -0.922  0.934  -1.51   -0.283  FALSE    FALSE   
#>  6 fam_I…  0.118 -0.403   1.92   1.97  -1.09    0.756  -0.252  FALSE    TRUE    
#>  7 fam_I…  0.507  0.810   0.438 -1.26   1.61    1.22    0.975  FALSE    FALSE   
#>  8 fam_I… -1.33  -1.97   -2.49  -0.711 -1.13   -0.162  -1.85   FALSE    FALSE   
#>  9 fam_I… -0.312 -0.132   0.467  0.547  0.0115  1.27    0.814  FALSE    FALSE   
#> 10 fam_I… -0.538 -0.631  -1.49  -0.246 -0.102  -0.548   0.626  FALSE    FALSE   
#> # ℹ 990 more rows
#> # ℹ 10 more variables: mgm_status <lgl>, mgf_status <lgl>, mhs1_status <lgl>,
#> #   mhs2_status <lgl>, o_aoo <dbl>, m_aoo <dbl>, mgm_aoo <dbl>, mgf_aoo <dbl>,
#> #   mhs1_aoo <dbl>, mhs2_aoo <dbl>
#> 
#> $thresholds
#> # A tibble: 6,000 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o      -Inf  3.55
#>  2 fam_ID_2  fam_ID_2_1  o      -Inf  3.14
#>  3 fam_ID_3  fam_ID_3_1  o      -Inf  2.59
#>  4 fam_ID_4  fam_ID_4_1  o      -Inf  2.79
#>  5 fam_ID_5  fam_ID_5_1  o      -Inf  2.91
#>  6 fam_ID_6  fam_ID_6_1  o      -Inf  3.24
#>  7 fam_ID_7  fam_ID_7_1  o      -Inf  2.87
#>  8 fam_ID_8  fam_ID_8_1  o      -Inf  3.52
#>  9 fam_ID_9  fam_ID_9_1  o      -Inf  3.24
#> 10 fam_ID_10 fam_ID_10_1 o      -Inf  3.42
#> # ℹ 5,990 more rows
#> 
simulate_under_LTM_single(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = FALSE, 
h2 = 0.5, n_sim = 500, pop_prev = .05)
#> $sim_obs
#> # A tibble: 500 × 10
#>    fam_ID        m       f     s1 m_status f_status s1_status m_aoo f_aoo s1_aoo
#>    <chr>     <dbl>   <dbl>  <dbl> <lgl>    <lgl>    <lgl>     <dbl> <dbl>  <dbl>
#>  1 fam_ID_1  0.754  0.749  -0.613 FALSE    FALSE    FALSE        44    51     25
#>  2 fam_ID_2 -1.63  -0.112   0.554 FALSE    FALSE    FALSE        69    62     40
#>  3 fam_ID_3  1.39   0.727   1.51  FALSE    FALSE    FALSE        61    58     33
#>  4 fam_ID_4 -0.585 -0.0125 -1.19  FALSE    FALSE    FALSE        53    49     26
#>  5 fam_ID_5 -1.09  -0.399   0.734 FALSE    FALSE    FALSE        58    49     29
#>  6 fam_ID_6 -1.51   0.801  -0.919 FALSE    FALSE    FALSE        55    59     32
#>  7 fam_ID_7  0.401 -1.69   -1.49  FALSE    FALSE    FALSE        53    55     33
#>  8 fam_ID_8 -0.435 -1.31   -0.189 FALSE    FALSE    FALSE        47    47     24
#>  9 fam_ID_9  0.703  0.437   1.65  FALSE    FALSE    TRUE         56    61     92
#> 10 fam_ID_… -0.429  0.418  -0.259 FALSE    FALSE    FALSE        56    60     37
#> # ℹ 490 more rows
#> 
#> $thresholds
#> # A tibble: 1,500 × 5
#>    fam_ID    indiv_ID    role  lower upper
#>    <chr>     <chr>       <chr> <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  m      -Inf  2.51
#>  2 fam_ID_2  fam_ID_2_1  m      -Inf  1.78
#>  3 fam_ID_3  fam_ID_3_1  m      -Inf  1.93
#>  4 fam_ID_4  fam_ID_4_1  m      -Inf  2.18
#>  5 fam_ID_5  fam_ID_5_1  m      -Inf  2.02
#>  6 fam_ID_6  fam_ID_6_1  m      -Inf  2.11
#>  7 fam_ID_7  fam_ID_7_1  m      -Inf  2.18
#>  8 fam_ID_8  fam_ID_8_1  m      -Inf  2.40
#>  9 fam_ID_9  fam_ID_9_1  m      -Inf  2.08
#> 10 fam_ID_10 fam_ID_10_1 m      -Inf  2.08
#> # ℹ 1,490 more rows
#> 
simulate_under_LTM_single(fam_vec = c(), n_fam = NULL, add_ind = TRUE, h2 = 0.5, 
n_sim = 200, pop_prev = 0.05)
#> Warning message: 
#>  Neither fam_vec nor n_fam is specified...
#> $sim_obs
#> # A tibble: 200 × 5
#>    fam_ID          g        o o_status o_aoo
#>    <chr>       <dbl>    <dbl> <lgl>    <dbl>
#>  1 fam_ID_1  -0.772  -0.198   FALSE       13
#>  2 fam_ID_2  -0.421  -0.552   FALSE       38
#>  3 fam_ID_3  -0.317  -0.00619 FALSE       23
#>  4 fam_ID_4  -0.147   0.304   FALSE       29
#>  5 fam_ID_5   0.428   0.564   FALSE       38
#>  6 fam_ID_6   0.428   0.366   FALSE       15
#>  7 fam_ID_7   1.32    3.15    TRUE        27
#>  8 fam_ID_8   0.310  -1.70    FALSE       10
#>  9 fam_ID_9   0.0167 -0.912   FALSE       25
#> 10 fam_ID_10 -1.19   -1.28    FALSE       30
#> # ℹ 190 more rows
#> 
#> $thresholds
#> # A tibble: 200 × 5
#>    fam_ID    indiv_ID    role    lower upper
#>    <chr>     <chr>       <chr>   <dbl> <dbl>
#>  1 fam_ID_1  fam_ID_1_1  o     -Inf     3.63
#>  2 fam_ID_2  fam_ID_2_1  o     -Inf     2.75
#>  3 fam_ID_3  fam_ID_3_1  o     -Inf     3.30
#>  4 fam_ID_4  fam_ID_4_1  o     -Inf     3.09
#>  5 fam_ID_5  fam_ID_5_1  o     -Inf     2.75
#>  6 fam_ID_6  fam_ID_6_1  o     -Inf     3.57
#>  7 fam_ID_7  fam_ID_7_1  o        3.16  3.16
#>  8 fam_ID_8  fam_ID_8_1  o     -Inf     3.73
#>  9 fam_ID_9  fam_ID_9_1  o     -Inf     3.23
#> 10 fam_ID_10 fam_ID_10_1 o     -Inf     3.05
#> # ℹ 190 more rows
#>